### 2021-04-04 34865f4
move README-code checking to a pytest test
