+++ PySDM_tests/devops_tests/test_readme_code.py
@@ -1,0 +1,3 @@
+import exdown
+
+test_readme = exdown.pytests("../../README.md", syntax_filter="Python")
