+++ PySDM_tests/devops_tests/test_readme_code.py
@@ -1,3 +1,7 @@
 import exdown
+import pathlib
 
-test_readme = exdown.pytests("../../README.md", syntax_filter="Python")
+test_readme = exdown.pytests(
+    pathlib.Path(__file__).parent.parent.parent.joinpath("README.md").absolute(),
+    syntax_filter="Python"
+)
