+++ PySDM_tests/devops_tests/test_readme_code.py
@@ -1,7 +1,10 @@
 import exdown
 import pathlib
+import pytest
+import platform
 
-test_readme = exdown.pytests(
+# note: skipping Windows to avoid non-UTF encoding trouble
+test_readme = pytest.mark.skipif('platform.system() == "Windows"')(exdown.pytests(
     pathlib.Path(__file__).parent.parent.parent.joinpath("README.md").absolute(),
     syntax_filter="Python"
-)
+))
