### 2021-04-05 34e79f2
skipping README test on Windows through sys.getfilesystemencoding() != 'utf-8'
