+++ PySDM_tests/devops_tests/test_readme_code.py
@@ -1,10 +1,9 @@
 import exdown
 import pathlib
-import pytest
-import sys
 
 
-test_readme = pytest.mark.skipif("sys.getfilesystemencoding() != 'utf-8'")(exdown.pytests(
+test_readme = exdown.pytests(
     pathlib.Path(__file__).parent.parent.parent.joinpath("README.md").absolute(),
-    syntax_filter="Python"
-))
+    syntax_filter="Python",
+    encoding='utf8'
+)
