### 2021-04-07 e620145
fix spectral sampling + use Logarithmic sampling without custom range in test_spectrum (Kreidenweis et al example)
