### 2021-04-07 19b6029
adding a TODO, removing a bogus assert
