+++ PySDM_tests/smoke_tests/Kreidenweis_et_al_2003/test_Fig_1.py
@@ -68,5 +68,4 @@ class TestFig1:
             pyplot.xlabel('pH')
             pyplot.show()
 
-        assert example_output['pH'][0] == 0
         assert 5 < example_output['pH'][-1] < 5.2
