### 2021-04-11 e42d484
solving fastmath problem with np.ldexp, exponential steps & larger limit for interval search in condensation
