### 2021-04-11 c9bc7a5
trying with math.ldexp? (and 128 limit)
