### 2021-04-16 399ca26
draft of test_fig_3 for Partridge_et_al example
