### 2021-04-16 3e0d0cf
removing numba_helpers, introducing some cross-backend magick
