+++ PySDM/backends/numba/impl/_chemistry_methods.py
@@ -2,10 +2,9 @@ import numba
 import numpy as np
 
 from PySDM.backends.numba import conf
-from PySDM.backends.numba.numba_helpers import pH2H, H2pH
 from PySDM.backends.numba.toms748 import toms748_solve
 from PySDM.physics.constants import Md, R_str, Rd, M, K_H2O
-from PySDM.physics.formulae import radius
+from PySDM.physics.formulae import radius, pH2H, H2pH
 from PySDM.dynamics.aqueous_chemistry.support import HENRY_CONST, SPECIFIC_GRAVITY, \
     MASS_ACCOMMODATION_COEFFICIENTS, DIFFUSION_CONST, GASEOUS_COMPOUNDS, DISSOCIATION_FACTORS, \
     KINETIC_CONST, EQUILIBRIUM_CONST
