+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -4,8 +4,8 @@ Created at 11.2019
 
 from PySDM.physics import constants as const
 from PySDM.backends.numba import conf
-from PySDM.backends.numba.numba_helpers import \
-    radius, temperature_pressure_RH, dr_dt_MM, dr_dt_FF, dT_i_dt_FF, dthd_dt, within_tolerance
+from PySDM.physics.formulae import temperature_pressure_RH, dr_dt_MM, dr_dt_FF, dT_i_dt_FF, radius, dthd_dt, \
+    within_tolerance
 from PySDM.backends.numba.toms748 import toms748_solve
 from PySDM.backends.numba.coordinates import mapper as coordinates
 import numba
