+++ PySDM/physics/formulae.py
@@ -2,29 +2,95 @@
 Crated at 2019
 """
 
-from PySDM.backends.numba import numba_helpers as physics
+from PySDM.backends.formulae import formula
+from PySDM.backends.numba import conf
 from PySDM.physics import constants as const
-
 import numpy as np
-
-dr_dt_MM = physics.dr_dt_MM
-R = physics.R
-r_cr = physics.r_cr
-pvs = physics.pvs
-lv = physics.lv
-c_p = physics.c_p
-dthd_dt = physics.dthd_dt
-temperature_pressure_RH = physics.temperature_pressure_RH
-radius = physics.radius
-volume = physics.volume
-RH_eq = physics.RH_eq
-pH2H = physics.pH2H
+from PySDM.physics.constants import R_str
+from numpy import exp
 
 
+@formula(inline='never')
+def dr_dt_MM(r, T, p, RH, kp, rd):
+    nom = (RH - RH_eq(r, T, kp, rd))
+    den = (
+            Fd(T, D(r, T)) +
+            Fk(T, K(r, T, p), lv(T))
+    )
+    return 1 / r * nom / den
+
+
+@formula
+def R(q):
+    return _mix(q, const.Rd, const.Rv)
+
+
+@formula
+def r_cr(kp, rd, T):
+    return np.sqrt(3 * kp * rd ** 3 / A(T))
+
+
+@formula
+def lv(T):
+    return const.l_tri + (const.c_pv - const.c_pw) * (T - const.T_tri)
+
+
+@formula
+def c_p(q):
+    return _mix(q, const.c_pd, const.c_pv)
+
+
+@formula
+def dthd_dt(rhod, thd, T, dqv_dt):
+    return - lv(T) * dqv_dt / const.c_pd / T * thd * rhod
+
+
+@formula(fastmath=False)
+def temperature_pressure_RH(rhod, thd, qv):
+    # equivalent to eqs A11 & A12 in libcloudph++ 1.0 paper
+    exponent = const.Rd / const.c_pd
+    pd = np.power((rhod * const.Rd * thd) / const.p1000 ** exponent, 1 / (1 - exponent))
+    T = thd * (pd / const.p1000) ** exponent
+
+    R = const.Rv / (1 / qv + 1) + const.Rd / (1 + qv)
+    p = rhod * (1 + qv) * R * T
+
+    RH = (p - pd) / pvs(T)
+
+    return T, p, RH
+
+
+@formula
+def radius(volume):
+    return (volume * 3 / 4 / np.pi) ** (1 / 3)
+
+
+@formula
+def volume(radius):
+    return 4 / 3 * np.pi * radius ** 3
+
+
+@formula
+def RH_eq(r, T, kp, rd):
+    return 1 + A(T) / r - B(kp, rd) / r ** 3
+
+
+@formula
+def pH2H(pH):
+    return 10**(-pH) * 1e3
+
+
+@formula
+def pvs(T):
+    return const.ARM_C1 * exp((const.ARM_C2 * (T - const.T0)) / (T - const.T0 + const.ARM_C3))
+
+
+@formula
 def th_dry(th_std, qv):
     return th_std * np.power(1 + qv / const.eps, const.Rd / const.c_pd)
 
 
+@formula
 def th_std(p, T):
     return T * (const.p1000 / p)**(const.Rd / const.c_pd)
 
@@ -86,13 +152,109 @@ def explicit_euler(y, dt, dy_dt):
     y += dt * dy_dt
 
 
+@formula
 def mole_fraction_2_mixing_ratio(mole_fraction, specific_gravity):
[…]
