### 2021-04-17 cc76bba
plot-False in Partridge example smoke tests
