+++ PySDM/initialisation/r_wet_init.py
@@ -22,7 +22,7 @@ def r_wet_init(r_dry: np.ndarray, environment, cell_id: np.ndarray, kappa, rtol=
 def r_wet_init_impl(r_dry: np.ndarray, T, p, RH, cell_id: np.ndarray, kappa, rtol, RH_range=(0, 1)):
     r_wet = np.empty_like(r_dry)
 
-    for i in prange(len(r_dry)):
+    for i in prange(len(r_wet)):
         r_d = r_dry[i]
         cid = cell_id[i]
         # root-finding initial guess
