### 2021-04-17 5ff10c2
precision resolver in c_inline; reverting parallel setting in r_wet_init
