+++ PySDM/backends/formulae.py
@@ -2,6 +2,7 @@ from PySDM.physics import constants as const
 from PySDM.backends.numba import conf
 import inspect
 import re
+from PySDM.backends.thrustRTC.impl.precision_resolver import PrecisionResolver
 
 from PySDM.physics import _flag
 if _flag.DIMENSIONAL_ANALYSIS:
@@ -19,12 +20,13 @@ else:
 def c_inline(fun, **args):
     prae = r"([+\-*/( ]|^)"
     post = r"([ )/*\-+]|$)"
+    real_t = PrecisionResolver.get_C_type()
     source = inspect.getsourcelines(fun)[0]
     assert len(source) == 3
     source = source[-1].strip()
     source = re.sub("^return ", "", source)
     for arg in inspect.signature(fun).parameters:
-        source = re.sub(f"{prae}({arg}){post}", f"\\1({args[arg]})\\3", source)
-    source = re.sub(f"{prae}const\\.([^\\d\\W]\\w*]*){post}", "\\1{const.\\2}\\3", source)
+        source = re.sub(f"{prae}({arg}){post}", f"\\1{real_t}({args[arg]})\\3", source)
+    source = re.sub(f"{prae}const\\.([^\\d\\W]\\w*]*){post}", "\\1" + real_t + "({const.\\2})\\3", source)
     source = eval(f'f"""{source}"""')
-    return f'({source})'
+    return f'{real_t}({source})'
