### 2021-04-17 aa48f03
moving c_inline to backends/thrustRTC/impl
