+++ PySDM/backends/thrustRTC/thrustRTC.py
@@ -24,8 +24,8 @@ class ThrustRTC(
 
     default_croupier = 'global'
 
-    def __init__(self):
-        raise Exception("Backend is stateless.")
+    def __init__(self, formulae):
+        self.formulae = formulae
 
     @staticmethod
     def sanity_check():
