+++ PySDM/builder.py
@@ -12,12 +12,15 @@ from PySDM.attributes.impl.mapper import get_class as attr_class
 from PySDM.attributes.physics.multiplicities import Multiplicities
 from PySDM.attributes.physics.volume import Volume
 from PySDM.attributes.numerics.cell_id import CellID
+from PySDM.physics.formulae import Formulae
+import inspect
 
 
 class Builder:
 
-    def __init__(self, n_sd, backend):
-        self.core = Core(n_sd, backend)
+    def __init__(self, n_sd, backend, formulae=Formulae()):
+        assert inspect.isclass(backend)
+        self.core = Core(n_sd, backend(formulae))
         self.req_attr = {'n': Multiplicities(self), 'volume': Volume(self), 'cell id': CellID(self)}
         self.aerosol_radius_threshold = 0
         self.condensation_params = None
