+++ PySDM/physics/formulae.py
@@ -2,7 +2,7 @@
 Crated at 2019
 """
 
-from PySDM.physics.impl.formulae import formula
+from PySDM.physics.impl.formula import formula
 from PySDM.physics import constants as const
 import numpy as np
 from PySDM.physics.constants import R_str
