### 2021-04-18 f09e99c
parallel r_wet_init
