### 2021-04-18 a52931a
reverting parallel change in r_wet_init; float format in c_inline
