### 2021-04-18 374826e
lv sanitization; cleanups in physics_methods
