### 2021-04-19 40023a9
fix in formulae (unused code...); TOMS "error reporting"
