### 2021-04-19 0374eed
propagating condensation failures using bool flags instead of exceptions/asserts (which are not supported as of now within numba parallel code)
