+++ PySDM/dynamics/condensation.py
@@ -31,6 +31,7 @@ class Condensation:
         self.rtol_thd = rtol_thd
 
         self.RH_max = None
+        self.success = None
 
         self.__substeps = substeps
         self.adaptive = adaptive
@@ -54,6 +55,8 @@ class Condensation:
 
         self.RH_max = self.core.Storage.empty(self.core.mesh.n_cell, dtype=float)
         self.RH_max[:] = np.nan
+        self.success = self.core.Storage.empty(self.core.mesh.n_cell, dtype=bool)
+        self.success[:] = False
         self.cell_order = np.arange(self.core.mesh.n_cell)
 
     def __call__(self):
@@ -71,8 +74,11 @@ class Condensation:
                 rtol_thd=self.rtol_thd,
                 counters=self.counters,
                 RH_max=self.RH_max,
+                success=self.success,
                 cell_order=self.cell_order
             )
+            if not (self.success.data[:] == True).all():
+                raise RuntimeError("Condensation failed")
             # note: this makes order of dynamics matter (e.g., condensation after chemistry or before)
             self.core.update_TpRH()
 
