### 2021-04-19 8042352
fixing TpRH for GPU
