### 2021-04-19 6a5a37e
workaround for type problem in FakeThrust
