+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -4,7 +4,7 @@ Created at 11.2019
 
 from PySDM.physics import constants as const
 from PySDM.backends.numba import conf
-from PySDM.physics.formulae import temperature_pressure_RH, dr_dt_MM, dr_dt_FF, dT_i_dt_FF, radius, dthd_dt, \
+from PySDM.physics.formulae import temperature_pressure_pv, dr_dt_MM, dr_dt_FF, dT_i_dt_FF, radius, dthd_dt, \
     within_tolerance
 import PySDM.physics.formulae as phys
 from PySDM.backends.numba.toms748 import toms748_solve
@@ -84,10 +84,12 @@ class CondensationMethods:
                 thd += dt * dthd_dt_pred / 2  # TODO #48 example showing that it makes sense
                 qv += dt * dqv_dt_pred / 2
 
-                T, p, RH = temperature_pressure_RH(rhod_mean, thd, qv)
+                T, p, pv = temperature_pressure_pv(rhod_mean, thd, qv)
                 lv = phys.lv(T)
+                pvs = phys.pvs(T)
+                RH = pv / pvs
                 ml_new, success_within_substep, n_activating, n_deactivating, n_ripening = \
-                    calculate_ml_new(dt, fake, T, p, RH, v, particle_T, v_cr, n, vdry, cell_idx, kappa, qv, lv, rtol_x)
+                    calculate_ml_new(dt, fake, T, p, RH, v, particle_T, v_cr, n, vdry, cell_idx, kappa, qv, lv, pvs, rtol_x)
                 dml_dt = (ml_new - ml_old) / dt
                 dqv_dt_corr = - dml_dt / m_d
                 dthd_dt_corr = dthd_dt(rhod=rhod_mean, thd=thd, T=T, dqv_dt=dqv_dt_corr, lv=lv)
@@ -118,21 +120,21 @@ class CondensationMethods:
     @staticmethod
     def make_calculate_ml_new(dx_dt, volume_of_x, x, enable_drop_temperatures):
         @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
-        def _minfun_FF(x_new, x_old, dt, T, p, RH, qv, lv, kappa, rd, T_i):
+        def _minfun_FF(x_new, x_old, dt, T, p, RH, qv, lv, pvs, kappa, rd, T_i):
             r_new = radius(volume_of_x(x_new))
             dr_dt = dr_dt_FF(r_new, T, p, qv, lv, kappa, rd, T_i)
             return x_old - x_new + dt * dx_dt(x_new, dr_dt)
 
         @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
-        def _minfun_MM(x_new, x_old, dt, T, p, RH, qv, lv, kappa, rd, _):
+        def _minfun_MM(x_new, x_old, dt, T, p, RH, qv, lv, pvs, kappa, rd, _):
             r_new = radius(volume_of_x(x_new))
-            dr_dt = dr_dt_MM(r_new, T, p, RH, lv, kappa, rd)
+            dr_dt = dr_dt_MM(r_new, T, p, RH, lv, pvs, kappa, rd)
             return x_old - x_new + dt * dx_dt(x_new, dr_dt)
 
         minfun = _minfun_FF if enable_drop_temperatures else _minfun_MM
 
         @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
-        def calculate_ml_new(dt, fake, T, p, RH, v, particle_T, v_cr, n, vdry, cell_idx, kappa, qv, lv, rtol_x):
[…]
