+++ PySDM/initialisation/r_wet_init.py
@@ -22,6 +22,7 @@ def r_wet_init(r_dry: np.ndarray, environment, cell_id: np.ndarray, kappa, rtol=
 def r_wet_init_impl(r_dry: np.ndarray, T, p, RH, cell_id: np.ndarray, kappa, rtol, RH_range=(0, 1)):
     r_wet = np.empty_like(r_dry)
     lv = formulae.lv(T)
+    pvs = formulae.pvs(T)
     for i in prange(len(r_dry)):
         r_d = r_dry[i]
         cid = cell_id[i]
@@ -34,6 +35,7 @@ def r_wet_init_impl(r_dry: np.ndarray, T, p, RH, cell_id: np.ndarray, kappa, rto
             p[cid],
             np.maximum(RH_range[0], np.minimum(RH_range[1], RH[cid])),
             lv[cid],
+            pvs[cid],
             kappa,
             r_d
         )
