+++ PySDM/backends/numba/impl/_physics_methods.py
@@ -5,7 +5,7 @@ Created at 11.2019
 import numba
 from numba import prange
 from PySDM.backends.numba import conf
-from PySDM.physics.formulae import temperature_pressure_RH
+from PySDM.physics.formulae import temperature_pressure_pv, pvs
 
 
 class PhysicsMethods:
@@ -27,7 +27,8 @@ class PhysicsMethods:
     @numba.njit(**conf.JIT_FLAGS)
     def temperature_pressure_RH_body(rhod, thd, qv, T, p, RH):
         for i in prange(T.shape[0]):
-            T[i], p[i], RH[i] = temperature_pressure_RH(rhod[i], thd[i], qv[i])
+            T[i], p[i], pv = temperature_pressure_pv(rhod[i], thd[i], qv[i])
+            RH[i] = pv / pvs(T[i])
 
     @staticmethod
     def temperature_pressure_RH(rhod, thd, qv, T, p, RH):
