+++ PySDM/backends/numba/bdf.py
@@ -2,7 +2,7 @@
 Created at 09.01.2020
 """
 
-from PySDM.physics.constants import rho_w
+import PySDM.physics.constants as const
 from PySDM.physics import formulae as phys
 from PySDM.backends.numba.numba import Numba
 from PySDM.backends.numba.conf import JIT_FLAGS
@@ -18,7 +18,7 @@ rtol = 1e-4
 
 
 def patch_core(core):
-    core.condensation_solver = make_solve(core.backend.formulae.condensation_coord)
+    core.condensation_solver = make_solve(core.backend.formulae)
     core.condensation = types.MethodType(bdf_condensation, core)
 
 
@@ -62,10 +62,12 @@ def bdf_condensation(core, kappa, rtol_x, rtol_thd, counters, RH_max, success, c
     )
 
 
-def make_solve(coord):
-    x = coord.x
-    volume = coord.volume
-    dx_dt = coord.dx_dt
+def make_solve(formulae):
+    x = formulae.condensation_coordinate.x
+    volume = formulae.condensation_coordinate.volume
+    dx_dt = formulae.condensation_coordinate.dx_dt
+    pvs_C = formulae.saturation_vapour_pressure.pvs_Celsius
+    lv = formulae.latent_heat.lv
 
     def solve(
             v, particle_temperatures, v_cr, n, vdry,
@@ -80,6 +82,8 @@ def make_solve(coord):
         qt = qv + _ODESystem.ql(n[cell_idx], y0[idx_x:], m_d_mean)
 
         odesys = _ODESystem(
+            pvs_C,
+            lv,
             kappa,
             vdry[cell_idx],
             n[cell_idx],
@@ -109,13 +113,15 @@ def make_solve(coord):
         m_new = 0
         for i in range(n_sd_in_cell):
             v_new = volume(y1[idx_x + i])
-            m_new += n[cell_idx[i]] * v_new * rho_w
+            m_new += n[cell_idx[i]] * v_new * const.rho_w
             v[cell_idx[i]] = v_new
 
         return integ.success, qt - m_new / m_d_mean, y1[idx_thd], 1, 1, 1, 1, np.nan
 
     class _ODESystem:
-        def __init__(self, kappa, dry_volume: np.ndarray, n: np.ndarray, dthd_dt, dqv_dt, m_d_mean, rhod_mean, qt):
+        def __init__(self, pvs_C, lv, kappa, dry_volume: np.ndarray, n: np.ndarray, dthd_dt, dqv_dt, m_d_mean, rhod_mean, qt):
+            self.pvs_C = pvs_C
+            self.lv = lv
             self.kappa = kappa
             self.rd = phys.radius(volume=dry_volume)
             self.n = n
@@ -131,25 +137,24 @@ def make_solve(coord):
 
             qv = self.qt + self.dqv_dt * t - self.ql(self.n, x, self.m_d_mean)
             T, p, pv = phys.temperature_pressure_pv(self.rhod_mean, thd, qv)
-            pvs = phys.pvs(T)
+            pvs = self.pvs_C(T - const.T0)
             RH = pv / pvs
             dy_dt = np.empty_like(y)
             self.impl(dy_dt, x, T, p, self.n, RH, self.kappa, self.rd, thd, self.dthd_dt, self.dqv_dt, self.m_d_mean,
-                      self.rhod_mean, pvs)
+                      self.rhod_mean, pvs, self.lv(T))
             return dy_dt
 
         @staticmethod
[…]
