+++ PySDM/attributes/physics/terminal_velocity.py
@@ -3,7 +3,7 @@ Created at 11.05.2020
 """
 
 from PySDM.attributes.impl.derived_attribute import DerivedAttribute
-from .gunn_and_kinzer import Interpolation
+from PySDM.physics.terminal_velocity.gunn_and_kinzer import Interpolation
 
 
 class TerminalVelocity(DerivedAttribute):
