+++ requirements.txt
@@ -5,4 +5,4 @@ pytest
 exdown>=0.8.8
 
 # note: if cloning both PySDM and PySDM examples, consider "pip install -e"
-git+git://github.com/atmos-cloud-sim-uj/PySDM-examples
+git+git://github.com/slayoo/PySDM-examples@382ba61#egg=PySDM-examples
