+++ PySDM/physics/condensation_coordinate/volume.py
@@ -4,8 +4,7 @@ from numpy import pi
 class Volume:
     @staticmethod
     def dx_dt(x, dr_dt):
-        r = radius(x)
-        return 4 * pi * r**2 * dr_dt
+        return 4 * pi * (x * 3/4 / pi)**(2/3) * dr_dt
 
     @staticmethod
     def volume(x):
