+++ PySDM/backends/numba/test_helpers/bdf.py
@@ -18,11 +18,11 @@ rtol = 1e-4
 
 
 def patch_core(core):
-    core.condensation_solver = make_solve(core.backend.formulae)
-    core.condensation = types.MethodType(bdf_condensation, core)
+    core.condensation_solver = _make_solve(core.backend.formulae)
+    core.condensation = types.MethodType(_bdf_condensation, core)
 
 
-def bdf_condensation(core, kappa, rtol_x, rtol_thd, counters, RH_max, success, cell_order):
+def _bdf_condensation(core, kappa, rtol_x, rtol_thd, counters, RH_max, success, cell_order):
     n_threads = 1
     if core.particles.has_attribute("temperature"):
         raise NotImplementedError()
@@ -62,12 +62,39 @@ def bdf_condensation(core, kappa, rtol_x, rtol_thd, counters, RH_max, success, c
     )
 
 
-def make_solve(formulae):
+def _make_solve(formulae):
     x = formulae.condensation_coordinate.x
     volume = formulae.condensation_coordinate.volume
     dx_dt = formulae.condensation_coordinate.dx_dt
     pvs_C = formulae.saturation_vapour_pressure.pvs_Celsius
     lv = formulae.latent_heat.lv
+
+    @numba.njit(**{**JIT_FLAGS, **{'parallel': False, 'inline': 'always'}})
+    def _ql(n, x, m_d_mean):
+        return np.sum(n * volume(x)) * const.rho_w / m_d_mean
+
+    @numba.njit(**JIT_FLAGS)
+    def _impl(dy_dt, x, T, p, n, RH, kappa, rd, thd, dot_thd, dot_qv, m_d_mean, rhod_mean, pvs, lv):
+        for i in numba.prange(len(x)):
+            dy_dt[idx_x + i] = dx_dt(x[i], phys.dr_dt_MM(phys.radius(volume(x[i])), T, p, RH, lv, pvs, kappa, rd[i]))
+        dqv_dt = dot_qv - np.sum(n * volume(x) * dy_dt[idx_x:]) * const.rho_w / m_d_mean
+        dy_dt[idx_thd] = dot_thd + phys.dthd_dt(rhod_mean, thd, T, dqv_dt, lv)
+
+    @numba.njit(**{**JIT_FLAGS, **{'parallel': False}})
+    def _odesys(t, y, kappa, dry_volume, n, dthd_dt, dqv_dt, m_d_mean, rhod_mean, qt):
+        rd = phys.radius(volume=dry_volume)
+
+        thd = y[idx_thd]
+        x = y[idx_x:]
+
+        qv = qt + dqv_dt * t - _ql(n, x, m_d_mean)
+        T, p, pv = phys.temperature_pressure_pv(rhod_mean, thd, qv)
+        pvs = pvs_C(T - const.T0)
+        RH = pv / pvs
+
+        dy_dt = np.empty_like(y)
+        _impl(dy_dt, x, T, p, n, RH, kappa, rd, thd, dthd_dt, dqv_dt, m_d_mean, rhod_mean, pvs, lv(T))
+        return dy_dt
 
     def solve(
             v, particle_temperatures, v_cr, n, vdry,
@@ -79,29 +106,18 @@ def make_solve(formulae):
         y0 = np.empty(n_sd_in_cell + idx_x)
         y0[idx_thd] = thd
         y0[idx_x:] = x(v[cell_idx])
-        qt = qv + _ODESystem.ql(n[cell_idx], y0[idx_x:], m_d_mean)
-
-        odesys = _ODESystem(
-            pvs_C,
-            lv,
-            kappa,
-            vdry[cell_idx],
-            n[cell_idx],
-            dthd_dt,
-            dqv_dt,
-            m_d_mean,
-            rhod_mean,
-            qt
-        )
[…]
