### 2021-04-21 29f0e9a
introducing test_helpers directories; physics methods ctor in Numba; cleanups
