+++ PySDM/backends/numba/numba.py
@@ -27,6 +27,7 @@ class Numba(
 
     def __init__(self, formulae):
         self.formulae = formulae
+        PhysicsMethods.__init__(self)
 
     @staticmethod
     def sanity_check():
