+++ PySDM/backends/thrustRTC/conf.py
@@ -2,7 +2,7 @@
 Created at 20.05.2020
 """
 
-from PySDM.backends.thrustRTC.fakeThrustRTC._flag import fakeThrustRTC
+from PySDM.backends.thrustRTC.test_helpers._flag import fakeThrustRTC
 
 if not fakeThrustRTC:
     # noinspection PyUnresolvedReferences
