+++ PySDM_tests/smoke_tests/Yang_et_al_2018/test_just_do_it.py
@@ -5,7 +5,7 @@ Created at 2020
 from PySDM_examples.Yang_et_al_2018.example import Simulation
 from PySDM_examples.Yang_et_al_2018.settings import Settings
 from PySDM.physics.constants import si
-from PySDM.backends.numba import bdf
+from PySDM.backends.numba.test_helpers import bdf
 import pytest
 import numpy as np
 
