+++ PySDM/backends/thrustRTC/random.py
@@ -4,7 +4,7 @@ Created at 25.08.2020
 
 import numpy as np
 from .conf import trtc, rndrtc
-from PySDM.backends.thrustRTC.nice_thrust import nice_thrust
+from PySDM.backends.thrustRTC.impl.nice_thrust import nice_thrust
 from PySDM.backends.thrustRTC.conf import NICE_THRUST_FLAGS
 
 
