+++ PySDM/backends/numba/impl/_physics_methods.py
@@ -24,17 +24,18 @@ class PhysicsMethods:
         dC = c_r - c_l
         return (omega * dC + c_l) / (1 - dC)
 
-    @staticmethod
-    @numba.njit(**conf.JIT_FLAGS)
-    def temperature_pressure_RH_body(pvs_C, rhod, thd, qv, T, p, RH):
-        for i in prange(T.shape[0]):
-            T[i], p[i], pv = temperature_pressure_pv(rhod[i], thd[i], qv[i])
-            RH[i] = pv / pvs_C(T[i] - const.T0)
+    def __init__(self):
+        pvs_C = self.formulae.saturation_vapour_pressure.pvs_Celsius
+
+        @numba.njit(**conf.JIT_FLAGS)
+        def temperature_pressure_RH_body(rhod, thd, qv, T, p, RH):
+            for i in prange(T.shape[0]):
+                T[i], p[i], pv = temperature_pressure_pv(rhod[i], thd[i], qv[i])
+                RH[i] = pv / pvs_C(T[i] - const.T0)
+        self.temperature_pressure_RH_body = temperature_pressure_RH_body
 
     def temperature_pressure_RH(self, rhod, thd, qv, T, p, RH):
-        return PhysicsMethods.temperature_pressure_RH_body(
-            self.formulae.saturation_vapour_pressure.pvs_Celsius,
-            rhod.data, thd.data, qv.data, T.data, p.data, RH.data)
+        self.temperature_pressure_RH_body(rhod.data, thd.data, qv.data, T.data, p.data, RH.data)
 
     @staticmethod
     @numba.njit(**conf.JIT_FLAGS)
