+++ PySDM/backends/thrustRTC/impl/nice_thrust.py
@@ -2,7 +2,7 @@
 Created at 18.05.2020
 """
 
-from .conf import trtc
+from PySDM.backends.thrustRTC.conf import trtc
 
 
 def nice_thrust(*, wait=False, debug_print=False):
