### 2021-04-21 50f2577
less doplets in tests; fix __init__ in backends
