### 2021-04-22 f8fc6a9
bump examples requirement
