### 2021-04-22 11d31a0
A&S tests cleanup + testing Volume coordinate
