+++ PySDM_tests/smoke_tests/Arabas_and_Shima_2017/test_vs_BDF.py
@@ -26,7 +26,7 @@ def split(arg1, arg2):
 @pytest.mark.parametrize("settings_idx", range(setups_num))
 @pytest.mark.parametrize("rtol", rtols)
 @pytest.mark.parametrize("leg", ['ascent', 'descent'])
-def test_vs_BDF(settings_idx, data, rtol):
+def test_vs_BDF(settings_idx, data, rtol, leg):
     # Arrange
     supersaturation = {}
     for scheme in schemes:
