### 2021-04-22 c39767d
let's check timing with inline=always for formulae after earlier sanitizations
