### 2021-04-22 27253f9
removing drop temperature code
