+++ PySDM/core.py
@@ -35,6 +35,7 @@ class Core:
         self.IndexedStorage = make_IndexedStorage(backend)
 
         self.timers = {}
+        self.null = self.Storage.empty(0, dtype=float)
 
     @property
     def env(self):
@@ -85,16 +86,15 @@ class Core:
         # TODO #443: mark_updated
 
     def condensation(self, kappa, rtol_x, rtol_thd, counters, RH_max, success, cell_order):
-        particle_temperatures = \
-            self.particles["temperature"] if self.particles.has_attribute("temperature") else \
-            self.Storage.empty(0, dtype=float)
+        particle_heat = \
+            self.particles["heat"] if self.particles.has_attribute("heat") else \
+            self.null
 
         self.backend.condensation(
                 solver=self.condensation_solver,
                 n_cell=self.mesh.n_cell,
                 cell_start_arg=self.particles.cell_start,
                 v=self.particles["volume"],
-                particle_temperatures=particle_temperatures,
                 n=self.particles['n'],
                 vdry=self.particles["dry volume"],
                 idx=self.particles._Particles__idx,
