+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -73,7 +73,7 @@ class CondensationMethods:
     @staticmethod
     def make_step_impl(phys_pvs_C, phys_lv, calculate_ml_old, calculate_ml_new):
         @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
-        def step_impl(v, particle_T, v_cr, n, vdry, cell_idx, kappa, thd, qv, dthd_dt_pred, dqv_dt_pred,
+        def step_impl(v, v_cr, n, vdry, cell_idx, kappa, thd, qv, dthd_dt_pred, dqv_dt_pred,
                       m_d, rhod_mean, rtol_x, dt, n_substeps, fake):
             dt /= n_substeps
             ml_old = calculate_ml_old(v, n, cell_idx)
@@ -89,7 +89,7 @@ class CondensationMethods:
                 pvs = phys_pvs_C(T - const.T0)
                 RH = pv / pvs
                 ml_new, success_within_substep, n_activating, n_deactivating, n_ripening = \
-                    calculate_ml_new(dt, fake, T, p, RH, v, particle_T, v_cr, n, vdry, cell_idx, kappa, qv, lv, pvs, rtol_x)
+                    calculate_ml_new(dt, fake, T, p, RH, v, v_cr, n, vdry, cell_idx, kappa, qv, lv, pvs, rtol_x)
                 dml_dt = (ml_new - ml_old) / dt
                 dqv_dt_corr = - dml_dt / m_d
                 dthd_dt_corr = dthd_dt(rhod=rhod_mean, thd=thd, T=T, dqv_dt=dqv_dt_corr, lv=lv)
@@ -118,23 +118,15 @@ class CondensationMethods:
         return calculate_ml_old
 
     @staticmethod
-    def make_calculate_ml_new(dx_dt, volume_of_x, x, enable_drop_temperatures):
+    def make_calculate_ml_new(dx_dt, volume_of_x, x):
         @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
-        def _minfun_FF(x_new, x_old, dt, T, p, RH, qv, lv, pvs, kappa, rd, T_i):
-            r_new = radius(volume_of_x(x_new))
-            dr_dt = dr_dt_FF(r_new, T, p, qv, lv, kappa, rd, T_i)
-            return x_old - x_new + dt * dx_dt(x_new, dr_dt)
-
-        @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
-        def _minfun_MM(x_new, x_old, dt, T, p, RH, qv, lv, pvs, kappa, rd, _):
+        def minfun(x_new, x_old, dt, T, p, RH, qv, lv, pvs, kappa, rd):
             r_new = radius(volume_of_x(x_new))
             dr_dt = dr_dt_MM(r_new, T, p, RH, lv, pvs, kappa, rd)
             return x_old - x_new + dt * dx_dt(x_new, dr_dt)
 
-        minfun = _minfun_FF if enable_drop_temperatures else _minfun_MM
-
         @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
-        def calculate_ml_new(dt, fake, T, p, RH, v, particle_T, v_cr, n, vdry, cell_idx, kappa, qv, lv, pvs, rtol_x):
+        def calculate_ml_new(dt, fake, T, p, RH, v, v_cr, n, vdry, cell_idx, kappa, qv, lv, pvs, rtol_x):
             result = 0
             n_activating = 0
             n_deactivating = 0
@@ -144,13 +136,8 @@ class CondensationMethods:
                 x_old = x(v[drop])
[…]
