+++ PySDM/physics/drop_growth/__init__.py
@@ -1,2 +1,1 @@
-from .fick_fourier import FickFourier
 from .maxwell_mason import MaxwellMason
