+++ PySDM/physics/drop_growth/fick_fourier.py
@@ -1,2 +1,0 @@
-class FickFourier:
-    pass
