+++ PySDM/physics/drop_growth/maxwell_mason.py
@@ -1,2 +1,10 @@
+import PySDM.physics.constants as const
+
+
 class MaxwellMason:
-    pass
+    @staticmethod
+    def dr_dt(r, RH_eq, T, RH, lv, pvs, D, K):
+        return (RH - RH_eq) / (
+                const.rho_w * const.Rv * T / D / pvs +
+                const.rho_w * lv / K / T * (lv / const.Rv / T - 1)
+        ) / r
