### 2021-04-22 192ff90
r_dr_dt
