+++ PySDM/physics/condensation_coordinate/volume_logarithm.py
@@ -3,8 +3,8 @@ from numpy import pi, log, exp
 
 class VolumeLogarithm:
     @staticmethod
-    def dx_dt(x, dr_dt):
-        return exp(-x/3) * dr_dt * 3 * (3/4/pi)**(-1/3)
+    def dx_dt(x, r_dr_dt):
+        return exp(-2*x/3) * r_dr_dt * 3 * (3/4/pi)**(-2/3)
 
     @staticmethod
     def volume(x):
