+++ PySDM/physics/condensation_coordinate/volume.py
@@ -3,8 +3,8 @@ from numpy import pi
 
 class Volume:
     @staticmethod
-    def dx_dt(x, dr_dt):
-        return 4 * pi * (x * 3/4 / pi)**(2/3) * dr_dt
+    def dx_dt(x, r_dr_dt):
+        return 4 * pi * (x * 3/4 / pi)**(1/3) * r_dr_dt
 
     @staticmethod
     def volume(x):
