+++ PySDM/physics/drop_growth/maxwell_mason.py
@@ -3,8 +3,8 @@ import PySDM.physics.constants as const
 
 class MaxwellMason:
     @staticmethod
-    def dr_dt(r, RH_eq, T, RH, lv, pvs, D, K):
+    def r_dr_dt(RH_eq, T, RH, lv, pvs, D, K):
         return (RH - RH_eq) / (
                 const.rho_w * const.Rv * T / D / pvs +
                 const.rho_w * lv / K / T * (lv / const.Rv / T - 1)
-        ) / r
+        )
