### 2021-04-23 #3144 @slayoo
> Where are the review instructions?

https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
https://joss.readthedocs.io/en/latest/review_criteria.html
https://joss.readthedocs.io/en/latest/review_checklist.html

Thanks,
Sylwester
