### 2021-04-23 238b709
test_spin_up with fastmath on and off
