+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -16,7 +16,7 @@ from functools import lru_cache
 
 class CondensationMethods:
     @staticmethod
-    def make_adapt_substeps(dt, step_fake, dt_range, fuse=100, multiplier=2):
+    def make_adapt_substeps(jit_flags, dt, step_fake, dt_range, fuse=100, multiplier=2):
         if not isinstance(multiplier, int):
             raise ValueError()
         if dt_range[1] > dt:
@@ -28,7 +28,7 @@ class CondensationMethods:
             n_substeps_max = math.floor(dt / dt_range[0])
         n_substeps_min = math.ceil(dt / dt_range[1])
 
-        @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
+        @numba.njit(**jit_flags)
         def adapt_substeps(args, n_substeps, thd, rtol_thd):
             n_substeps = np.maximum(n_substeps_min, n_substeps // multiplier)
             thd_new_long, success = step_fake(args, dt, n_substeps)
@@ -53,8 +53,8 @@ class CondensationMethods:
         return adapt_substeps
 
     @staticmethod
-    def make_step_fake(step_impl):
-        @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
+    def make_step_fake(jit_flags, step_impl):
+        @numba.njit(**jit_flags)
         def step_fake(args, dt, n_substeps):
             dt /= n_substeps
             _, thd_new, _, _, _, _, success = step_impl(*args, dt, 1, True)
@@ -63,16 +63,16 @@ class CondensationMethods:
         return step_fake
 
     @staticmethod
-    def make_step(step_impl):
-        @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
+    def make_step(jit_flags, step_impl):
+        @numba.njit(**jit_flags)
         def step(args, dt, n_substeps):
             return step_impl(*args, dt, n_substeps, False)
 
         return step
 
     @staticmethod
-    def make_step_impl(phys_pvs_C, phys_lv, calculate_ml_old, calculate_ml_new):
-        @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
+    def make_step_impl(jit_flags, phys_pvs_C, phys_lv, calculate_ml_old, calculate_ml_new):
+        @numba.njit(**jit_flags)
         def step_impl(v, v_cr, n, vdry, cell_idx, kappa, thd, qv, dthd_dt_pred, dqv_dt_pred,
                       m_d, rhod_mean, rtol_x, dt, n_substeps, fake):
             dt /= n_substeps
@@ -107,8 +107,8 @@ class CondensationMethods:
         return step_impl
 
     @staticmethod
-    def make_calculate_ml_old():
-        @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
+    def make_calculate_ml_old(jit_flags):
+        @numba.njit(**jit_flags)
         def calculate_ml_old(v, n, cell_idx):
             result = 0
             for drop in cell_idx:
@@ -118,15 +118,15 @@ class CondensationMethods:
         return calculate_ml_old
 
     @staticmethod
-    def make_calculate_ml_new(dx_dt, volume_of_x, x, phys_r_dr_dt):
-        @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
+    def make_calculate_ml_new(jit_flags, dx_dt, volume_of_x, x, phys_r_dr_dt):
[…]
