+++ PySDM_tests/smoke_tests/Arabas_et_al_2015/test_spin_up.py
@@ -7,6 +7,7 @@ from PySDM_examples.Arabas_et_al_2015.settings import Settings
 from PySDM.physics import si
 import numpy as np
 from matplotlib import pyplot
+import pytest
 
 
 class DummyStorage:
@@ -20,9 +21,10 @@ class DummyStorage:
             self.profiles.append({"qv_env": np.mean(data, axis=0)})
 
 
-def test_spin_up(plot=False):
+@pytest.mark.parametrize("fastmath", (True, False))
+def test_spin_up(fastmath, plot=False):
     # Arrange
-    settings = Settings()
+    settings = Settings(fastmath=fastmath)
     settings.dt = .5 * si.second
     settings.simulation_time = 20 * settings.dt
     settings.output_interval = 1 * settings.dt
