### 2021-04-23 2934be3
jit_flags handling (fastmath) in r_wet_init
