+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -16,7 +16,7 @@ from functools import lru_cache
 
 class CondensationMethods:
     @staticmethod
-    def make_adapt_substeps(jit_flags, dt, step_fake, dt_range, fuse=100, multiplier=2):
+    def make_adapt_substeps(jit_flags, dt, step_fake, dt_range, fuse, multiplier):
         if not isinstance(multiplier, int):
             raise ValueError()
         if dt_range[1] > dt:
@@ -31,24 +31,35 @@ class CondensationMethods:
         @numba.njit(**jit_flags)
         def adapt_substeps(args, n_substeps, thd, rtol_thd):
             n_substeps = np.maximum(n_substeps_min, n_substeps // multiplier)
-            thd_new_long, success = step_fake(args, dt, n_substeps)
-            if not success:
-                return 0, success
+            success = False
             for burnout in range(fuse + 1):
                 if burnout == fuse:
-                    print("burnout")
+                    print("burnout (long)")
                     return 0, False
+                thd_new_long, success = step_fake(args, dt, n_substeps)
+                if success:
+                    break
+                else:
+                    n_substeps *= multiplier
+            for burnout in range(fuse + 1):
+                if burnout == fuse:
+                    print("burnout (short)")
+                    return 0, False
+                if n_substeps > n_substeps_max:
+                    print("n_substeps > n_substeps_max (", n_substeps, ") - reached dt_range[0] limit")
+                    break
                 thd_new_short, success = step_fake(args, dt, n_substeps * multiplier)
+                if not success:
+                    print("short failed")
+                    return 0, False
                 dthd_long = thd_new_long - thd
                 dthd_short = thd_new_short - thd
                 error_estimate = np.abs(dthd_long - multiplier * dthd_short)
                 thd_new_long = thd_new_short
-                if within_tolerance(error_estimate, thd, rtol_thd) and success:
+                if within_tolerance(error_estimate, thd, rtol_thd):
                     break
                 n_substeps *= multiplier
-                if n_substeps > n_substeps_max:
-                    print("n_substeps > n_substeps_max (", n_substeps, ")")
-                    break
+
             return np.minimum(n_substeps_max, n_substeps), success
 
         return adapt_substeps
@@ -119,7 +130,7 @@ class CondensationMethods:
         return calculate_ml_old
 
     @staticmethod
-    def make_calculate_ml_new(jit_flags, dx_dt, volume_of_x, x, phys_r_dr_dt):
+    def make_calculate_ml_new(jit_flags, dx_dt, volume_of_x, x, phys_r_dr_dt, max_iters, RH_rtol):
         @numba.njit(**jit_flags)
         def minfun(x_new, x_old, dt, p, kappa, rd, T, RH, lv, pvs, D, K):
             r_new = radius(volume_of_x(x_new))
@@ -138,11 +149,11 @@ class CondensationMethods:
                 x_old = x(v[drop])
                 r_old = radius(v[drop])
                 rd = radius(vdry[drop])
-                D = phys_D(r_old, T)
-                K = phys_K(r_old, T, p)
                 RH_eq = phys.RH_eq(r_old, T, kappa, rd)
-                args = (x_old, dt, p, kappa, rd, T, RH, lv, pvs, D, K)
-                if RH != RH_eq:
+                if not within_tolerance(RH - RH_eq, RH, RH_rtol):
+                    D = phys_D(r_old, T)
+                    K = phys_K(r_old, T, p)
+                    args = (x_old, dt, p, kappa, rd, T, RH, lv, pvs, D, K)
                     r_dr_dt_old = phys_r_dr_dt(RH_eq, T, RH, lv, pvs, D, K)
[…]
