### 2021-04-24 d1744de
adaptive cond logic fixes
