### 2021-04-24 b67873b
rd**3 -> rd3; back with x_insane; faster test_spin_ip
