+++ PySDM/backends/numba/test_helpers/bdf.py
@@ -78,7 +78,7 @@ def _make_solve(formulae):
             r = phys.radius(v)
             D = phys.D(r, T)
             K = phys.K(r, T, p)
-            RH_eq = phys.RH_eq(r, T, kappa, rd[i])
+            RH_eq = phys.RH_eq(r, T, kappa, rd[i]**3)
             dy_dt[idx_x + i] = dx_dt(x[i], r_dr_dt(RH_eq, T, RH, lv, pvs, D, K))
         dqv_dt = dot_qv - np.sum(n * volume(x) * dy_dt[idx_x:]) * const.rho_w / m_d_mean
         dy_dt[idx_thd] = dot_thd + phys.dthd_dt(rhod_mean, thd, T, dqv_dt, lv)
