### 2021-04-25 a67d1eb
introducing formulae.trivia
