+++ PySDM/physics/trivia.py
@@ -1,0 +1,5 @@
+
+class Trivia:
+    @staticmethod
+    def volume_of_density_mass(rho, m):
+        return m / rho
