+++ PySDM/products/state/cloud_droplet_effective_radius.py
@@ -9,7 +9,7 @@ from PySDM.physics import constants as const
 from PySDM.physics import formulae as phys
 from PySDM.products.product import MomentProduct
 
-GEOM_FACTOR = phys.volume(radius=1)**(-1/3)
+GEOM_FACTOR = const.pi_4_3**(-1/3)
 
 
 class CloudDropletEffectiveRadius(MomentProduct):
@@ -33,12 +33,12 @@ class CloudDropletEffectiveRadius(MomentProduct):
     def get(self):
         tmp = np.empty_like(self.buffer)
         self.download_moment_to_buffer('volume', rank=2/3,
-                                       filter_range=(phys.volume(self.radius_range[0]),
-                                                     phys.volume(self.radius_range[1])))
+                                       filter_range=(self.formulae.trivia.volume(self.radius_range[0]),
+                                                     self.formulae.trivia.volume(self.radius_range[1])))
         tmp[:] = self.buffer[:]
         self.download_moment_to_buffer('volume', rank=1,
-                                       filter_range=(phys.volume(self.radius_range[0]),
-                                                     phys.volume(self.radius_range[1])))
+                                       filter_range=(self.formulae.trivia.volume(self.radius_range[0]),
+                                                     self.formulae.trivia.volume(self.radius_range[1])))
         CloudDropletEffectiveRadius.__get_impl(self.buffer, tmp)
         const.convert_to(self.buffer, const.si.micrometre)
         return self.buffer
