+++ PySDM_tests/unit_tests/dummy_core.py
@@ -3,6 +3,7 @@ Created at 19.11.2019
 """
 
 from PySDM.builder import Builder
+from PySDM.physics.formulae import Formulae
 from PySDM.core import Core
 from .dummy_environment import DummyEnvironment
 from PySDM.attributes.physics.multiplicities import Multiplicities
@@ -12,7 +13,7 @@ from PySDM.attributes.numerics.cell_id import CellID
 class DummyCore(Builder, Core):
 
     def __init__(self, backend, n_sd=0):
-        Core.__init__(self, n_sd, backend)
+        Core.__init__(self, n_sd, backend(Formulae()))
         self.core = self
         self.environment = DummyEnvironment()
         self.environment.register(self)
