### 2021-04-25 ed22293
one more lru_cache?
