### 2021-04-26 4d9cdd2
same _magick?
