+++ .github/workflows/main.yml
@@ -1,4 +1,4 @@
-name: Build Status
+name: PySDM
 
 defaults:
   run:
