### 2021-04-26 a34fd35
explicit_euler -> trivia
