+++ PySDM/physics/trivia.py
@@ -1,4 +1,5 @@
 from PySDM.physics import constants as const
+from numpy import abs
 
 
 class Trivia:
@@ -17,3 +18,7 @@ class Trivia:
     @staticmethod
     def explicit_euler(y, dt, dy_dt):
         y += dt * dy_dt
+
+    @staticmethod
+    def within_tolerance(error_estimate, value, rtol):
+        return error_estimate < rtol * abs(value)
