+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -4,8 +4,7 @@ Created at 11.2019
 
 from PySDM.physics import constants as const
 from PySDM.backends.numba import conf
-from PySDM.physics.formulae import temperature_pressure_pv, dthd_dt, \
-    within_tolerance, D as phys_D, K as phys_K
+from PySDM.physics.formulae import temperature_pressure_pv, dthd_dt, D as phys_D, K as phys_K
 import PySDM.physics.formulae as phys
 from PySDM.backends.numba.toms748 import toms748_solve
 import numba
@@ -16,7 +15,7 @@ from functools import lru_cache
 
 class CondensationMethods:
     @staticmethod
-    def make_adapt_substeps(jit_flags, dt, step_fake, dt_range, fuse, multiplier):
+    def make_adapt_substeps(jit_flags, dt, step_fake, dt_range, fuse, multiplier, within_tolerance):
         if not isinstance(multiplier, int):
             raise ValueError()
         if dt_range[1] > dt:
@@ -129,7 +128,8 @@ class CondensationMethods:
         return calculate_ml_old
 
     @staticmethod
-    def make_calculate_ml_new(jit_flags, dx_dt, volume_of_x, x, phys_r_dr_dt, phys_RH_eq, phys_sigma, radius, max_iters, RH_rtol):
+    def make_calculate_ml_new(jit_flags, dx_dt, volume_of_x, x, phys_r_dr_dt, phys_RH_eq, phys_sigma, radius,
+                              within_tolerance, max_iters, RH_rtol):
         @numba.njit(**jit_flags)
         def minfun(x_new, x_old, dt, p, kappa, rd3, T, RH, lv, pvs, D, K):
             vol = volume_of_x(x_new)
@@ -188,7 +188,7 @@ class CondensationMethods:
                             a, b = b, a
                             fa, fb = fb, fa
 
-                        x_new, iters_taken = toms748_solve(minfun, args, a, b, fa, fb, rtol_x, max_iters)
+                        x_new, iters_taken = toms748_solve(minfun, args, a, b, fa, fb, rtol_x, max_iters, within_tolerance)
                         if iters_taken in (-1, max_iters):
                             if not fake:
                                 print("TOMS failed")
@@ -221,6 +221,7 @@ class CondensationMethods:
             phys_RH_eq=self.formulae.hygroscopicity.RH_eq,
             phys_sigma=self.formulae.surface_tension.sigma,
             radius=self.formulae.trivia.radius,
+            within_tolerance=self.formulae.trivia.within_tolerance,
             dx_dt=self.formulae.condensation_coordinate.dx_dt,
             volume=self.formulae.condensation_coordinate.volume,
             x=self.formulae.condensation_coordinate.x,
@@ -232,16 +233,17 @@ class CondensationMethods:
     @staticmethod
     @lru_cache()
     def make_condensation_solver_impl(fastmath, phys_pvs_C, phys_lv, phys_r_dr_dt, phys_RH_eq, phys_sigma, radius,
-                                      dx_dt, volume, x, dt, dt_range, adaptive,
+                                      within_tolerance, dx_dt, volume, x, dt, dt_range, adaptive,
                                       fuse=32, multiplier=2, RH_rtol=1e-7, max_iters=16):
         jit_flags = {**conf.JIT_FLAGS, **{'parallel': False, 'cache': False, 'fastmath': fastmath}}
 
         calculate_ml_old = CondensationMethods.make_calculate_ml_old(jit_flags)
         calculate_ml_new = CondensationMethods.make_calculate_ml_new(jit_flags, dx_dt, volume, x, phys_r_dr_dt, phys_RH_eq,
-                                                                     phys_sigma, radius, max_iters, RH_rtol)
[…]
