+++ PySDM_tests/unit_tests/backends/test_toms748.py
@@ -1,5 +1,6 @@
 from PySDM.backends.numba.toms748 import toms748_solve
 from scipy.optimize.zeros import toms748
+from PySDM.physics.formulae import Formulae
 import numba
 import numpy as np
 import os
@@ -27,7 +28,8 @@ def test_toms748(fun):
     a = -.5
     b = .5
     rtol = 1e-6
-    actual, iters = sut(f2, (), ax=a, bx=b, fax=f2(a), fbx=f2(b), max_iter=10, rtol=rtol)
+    wt = Formulae().trivia.within_tolerance
+    actual, iters = sut(f2, (), ax=a, bx=b, fax=f2(a), fbx=f2(b), max_iter=10, rtol=rtol, within_tolerance=wt)
     expected = toms748(f2, a, b)
 
     np.testing.assert_almost_equal(actual, expected)
