+++ PySDM/physics/hygroscopicity/__init__.py
@@ -1,1 +1,2 @@
 from .kappa_koehler import KappaKoehler
+from .kappa_koehler_leading_terms import KappaKoehlerLeadingTerms
