### 2021-04-26 #3219 @whedon
[REVIEW]: PySDM v1: particle-based cloud modelling package for warm-rain microphysics and aqueous chemistry
**Submitting author:** @piotrbartman (<a href="http://orcid.org/0000-0003-0265-6428">Piotr Bartman</a>)
**Repository:** https://github.com/atmos-cloud-sim-uj/PySDM.git
**Branch with paper.md** (empty if default branch):
**Version:** v1.27
