+++ PySDM/physics/state_variable_triplet/rhod_thd_qv.py
@@ -3,12 +3,21 @@ from numpy import power
 
 
 class RhodThdQv:
+    # A14 in libcloudph++ 1.0 paper
     @staticmethod
-    def temperature_pressure_pv(rhod, thd, qv):
-        # equivalent to eqs A11 & A12 in libcloudph++ 1.0 paper
-        exponent = const.Rd / const.c_pd
-        pd = power((rhod * const.Rd * thd) / const.p1000 ** exponent, 1 / (1 - exponent))
-        T = thd * (pd / const.p1000) ** exponent
-        R = const.Rv / (1 / qv + 1) + const.Rd / (1 + qv)
-        p = rhod * (1 + qv) * R * T
-        return T, p, p - pd
+    def T(rhod, thd):
+        return thd * (
+                power(
+                    (rhod * const.Rd * thd) / const.p1000 ** const.Rd_over_c_pd,
+                    1 / (1 - const.Rd_over_c_pd)
+                ) / const.p1000
+        ) ** const.Rd_over_c_pd
+
+    # A15 in libcloudph++ 1.0 paper
+    @staticmethod
+    def p(rhod, T, qv):
+        return rhod * (1 + qv) * (const.Rv / (1 / qv + 1) + const.Rd / (1 + qv)) * T
+
+    @staticmethod
+    def pv(p, qv):
+        return p * qv / (qv + const.eps)
