### 2021-04-27 atmos-cloud-sim-uj/PySDM#494 @darothen
conda-forge package for installation of non-Python dependencies?
Given this package's capability of running on the GPU via `ThrustRTC`, would it be possible to develop a `conda-forge` recipe to help users install all the pre-requisites for taking advantage of this hardware?
