### 2021-04-27 14e9c52
Update requirements.txt to include pystrict
When installing via the instructions `pip install git+https://github.com/atmos-cloud-sim-uj/PySDM.git` in the README from a fresh conda environment w/ Python=3.8, `pystrict` was not installed so neither the tests nor the demo code could run.
