+++ requirements.txt
@@ -1,4 +1,5 @@
 # test-time requirements
+exdown
 matplotlib>=3.2.2
 ghapi
 pystrict
