### 2021-04-27 #3219 @darothen
# Review - @darothen

`PySDM` is a new open-source package for particle-resolved modeling of the physics and chemistry of cloud and aerosol microphysics. It provides high-performance reference implementations of state-of-the-science process models in a flexible, highly-configurable package that is relatively easy to manipulate and modify to suit a wide variety of cloud and aerosol modeling tasks. It also ships with a back-end for running certain processes on GPUs, although this functionality does not seem to be fully implemented for all configurations which a user may seek to leverage.

Overall this is a high quality and valuable contribution to the field. Many researchers leverage simple box and parcel models for critical work into fundamental aerosol/cloud microphysics, and `PySDM` provides extremely useful tooling for exploring this domain and for replicating the results of other researchers. However, the library is lacking in documentation, both formally (in clearly defining the complex structure of the objects which comprise the code and how they interact) and informally (through docstrings and other comments/annotations in the code which may allow users to better understand arguments to key functions and how to manipulate and extend them). At present, the authors provide a nice "Hello world" example with some prose on how to configure it, as well as a basic section on package structure and API. But a more comprehensive listing of the arguments, attributes, and parameters that the user should know about in order to configure or extend this example would go a very long way to promote practical usability.

I would encourage the authors to improve or augment this documentation before the final publication of this software paper. However, in all other aspects I find this to be a comprehensive and uniquely valuable software contribution to the cloud and aerosol microphysical modeling community and an ideal candidate for publication in JOSS.

Please note the handful of detailed comments below on specific changes necessary to meet JOSS standards.

## General Checks

- Two authors on the paper (Grzegorz Lazarski and Aleksandra Talar) do not seem to appear as [contributors](https://github.com/atmos-cloud-sim-uj/PySDM/graphs/contributors) to the repo and aren't listed in "Author Contributions." I don't really believe there is any issue here - for long-incubated software projects like this with lots of research interfaces there have to be many contributors over time! But I do want to note this for the editors.

## Functionality

- A simple install recipe using PyPI is provided, but given the capability to run on the GPU a `conda-forge` recipe or something which can help readers understand requirements - or provide a turnkey installation with necessary dependencies - would be very helpful! (Issue [494](https://github.com/atmos-cloud-sim-uj/PySDM/issues/494))

- Installation using the documented instructions misses out on installing `pystrict`, a required dependency (Issue/PR [495](https://github.com/atmos-cloud-sim-uj/PySDM/pull/495))

## Documentation

- The automated tests seem to work out of the box, but the smoke tests seem to take a *very* long time to run. Even one of the unit tests (namely `attributes/chemistry/test_pH.py`) took a disproportionate amount of time compared to the others. Please note that I was not able to run the GPU tests as I did not have immediate access to a GPU-powered machine.

- While there is extensive, excellent demonstration and example code available from the authors in related, clearly-linked repositories, the API documentation leaves much to be desired. There is a high degree of customizability and configuration exposed to the user, but very little in the way of curated documentation for the APIs and data structures in the software, or docstrings and annotations. As a motivating example, consider the [spectra](https://github.com/atmos-cloud-sim-uj/PySDM/blob/master/PySDM/initialisation/spectra.py) module; in this module, a nice base class for building probability distributions is exposed, but there are no docstrings describing the required and optional arguments for either the base class or any of the sub-classes which implement it. An effort to improve the documentation with basic docstrings for all the core methods and functionality would greatly improve the quality of the software and the user experience when adapting it for a specialized research task. Ideally, a simple Sphinx-based documentation for the key API and configurations exposed to users would follow very plainly from such an effort, greatly improving the ease of access of the software.

- Community / contribution guidelines are not provided with the repo; the JOSS reviewer guidelines suggest that such documentattion should clearly indicate how third-parties may contribute to the software, report issues, and seek support should they need to do so.

## Software Paper

- The software paper does not include a section which explicitly explains a "Statement of Need." However, I do believe that the prose in the paper covers this topic with the necessary detail. I will defer to the editors whether a stand-alone "Statement of Need" section is required to meet JOSS guidelines.

- It's much appreciated that the authors include links to open source software which attempts to tackle subsets of the broad functionality that `PySDM` provides. I would strongly encourage the author to include a brief discussion of each package (particularly `pyrcel` and `PyBox`) and compare/contrast the unique or improved functionalities of `PySDM` relative to each. My take (as the author of one of those libraries, `pyrcel`!) is that `PySDM` offers a substantial improvement over each in terms of the breadth of calculations and science/research that can be performed with it as well as the cleanliness of its interface and design for future improvements and extensions. These are all very positive things for `PySDM` and I would strongly encourage the authors to lean into the fantastic work here to promote the capabilities and utility of their library relative to the competition!
