+++ PySDM/physics/formulae.py
@@ -7,10 +7,8 @@ from PySDM.physics import constants as const
 import numpy as np
 from PySDM.physics.constants import R_str
 from numpy import exp, log, pi
-from pystrict import strict
 
 
-@strict
 class Formulae:
     def __init__(self, condensation_coord: str='volume logarithm'):
         if condensation_coord == 'volume':
