### 2021-04-27 baf40e3
removing leftover @strict
