### 2021-04-27 atmos-cloud-sim-uj/PySDM#494 @slayoo
Thanks @darothen!

The only non-Python requirement should be the CUDA toolkit/driver. There seem to be a cuda-forge CUDA toolkit item: https://anaconda.org/conda-forge/cudatoolkit

@tehAgitto @Kyetuur @piotrbartman do you have any experience with Conda-forge?
