### 2021-04-27 bff6df4
bump examples requirement
