### 2021-04-27 #3219 @slayoo
@darothen, on behalf of all authors, let me thank for the feedback!

Let me just address one comment right away:

> Two authors on the paper (Grzegorz Lazarski and Aleksandra Talar) do not seem to appear as contributors to the repo and aren't listed in "Author Contributions." I don't really believe there is any issue here - for long-incubated software projects like this with lots of research interfaces there have to be many contributors over time! But I do want to note this for the editors.

GŁ's (@golui) and AT's (@aleksandratal) contributions were subject to refactors during which we've not preserved commit authorship, yet these can be traced back, e.g., here: https://github.com/atmos-cloud-sim-uj/PySDM/pull/157 & https://github.com/atmos-cloud-sim-uj/PySDM/pull/142 Both are mentioned in the "Author Contributions" section:

![image](https://user-images.githubusercontent.com/89685/116309753-71790400-a7a9-11eb-9012-cde6f6b06311.png)

Thanks!
