### 2021-04-28 f4537b9
check if importing PySDM works right after pip install
