+++ .github/workflows/main.yml
@@ -27,10 +27,13 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - run: |
-          pip install -r requirements.txt
+          pip install -e .
 
       - run: |
-          pip install -e .
+          python -We -c "import PySDM"
+
+      - run: |
+          pip install -r requirements.txt
 
       # TODO #184 (https://github.com/numba/numba/issues/6350#issuecomment-728174860)
       - if: matrix.platform == 'ubuntu-latest'
