+++ README.md
@@ -124,14 +124,18 @@ jupyter-notebook
 
 ![animation](https://github.com/atmos-cloud-sim-uj/PySDM/wiki/files/kinematic_2D_example.gif)
 
-## Hello-world example
+## Hello-world coalescence example in Python, Julia and Matlab
 
 In order to depict the PySDM API with a practical example, the following
-  listings provide a sample code roughly reproducing the 
-  Figure 2 from [Shima et al. 2009 paper](http://doi.org/10.1002/qj.441).
+  listings provide sample code roughly reproducing the 
+  Figure 2 from [Shima et al. 2009 paper](http://doi.org/10.1002/qj.441)
+  using PySDM from Python, Julia and Matlab (click ...).
 It is a coalescence-only set-up in which the initial particle size 
   spectrum is exponential and is deterministically sampled to match
   the condition of each super-droplet having equal initial multiplicity:
+<details open>
+<summary>Python</summary>
+
 ```Python
 from PySDM.physics import si
 from PySDM.initialisation.spectral_sampling import ConstantMultiplicity
@@ -142,11 +146,61 @@ initial_spectrum = Exponential(norm_factor=8.39e12, scale=1.19e5 * si.um**3)
 attributes = {}
 attributes['volume'], attributes['n'] = ConstantMultiplicity(spectrum=initial_spectrum).sample(n_sd)
 ```
+</details>
+<details>
+<summary>Julia (click to expand)</summary>
+
+```Julia
+# using Pkg
+# Pkg.add("Conda")
+# using Conda
+# Conda.pip_interop(true)
+# Conda.pip("install --pre", "git+https://github.com/atmos-cloud-sim-uj/PySDM.git")
+# Pkg.add("PyCall")
+
+using PyCall
+PySDM = pyimport("PySDM")
+PySDM_initialisation_spectra = pyimport("PySDM.initialisation.spectra")
+PySDM_initialisation_spectral_sampling = pyimport("PySDM.initialisation.spectral_sampling")
+si = PySDM.physics.si
+
+n_sd = 2^15
+initial_spectrum = PySDM_initialisation_spectra.Exponential(
+    norm_factor=8.39e12, scale=1.19e5 * si.um^3
+)
+attributes = environment.init_attributes(
+  spectral_discretisation=PySDM_initialisation_spectral_sampling.ConstantMultiplicity(spectrum=initial_spectrum)
+)
+```
+</details>
+<details>
+<summary>Matlab (click to expand)</summary>
+
+```Matlab
+PySDM = py.importlib.import_module('PySDM');
+PySDM_initialisation_spectra = py.importlib.import_module('PySDM.initialisation.spectra');
+PySDM_initialisation_spectral_sampling = py.importlib.import_module('PySDM.initialisation.spectral_sampling');
+si = PySDM_physics.constants.si;
+
+n_sd = 2 ^ 15;
+initial_spectrum = PySDM_initialisation_spectra.Exponential(pyargs(...
+    'norm_factor', 8.39e12, ...
+    'scale', 1.19e5 * si.um ^ 3 ...
+));
+attributes = environment.init_attributes(pyargs( ...
+    'spectral_discretisation', ...
+    PySDM_initialisation_spectral_sampling.ConstantMultiplicity(initial_spectrum)...
+));
+```
+</details>
 
 The key element of the PySDM interface is the [``Core``](https://github.com/atmos-cloud-sim-uj/PySDM/blob/master/PySDM/core.py) 
   class which instances are used to manage the system state and control the simulation.
[…]
