+++ README.md
@@ -359,8 +359,262 @@ The resultant plot looks as follows:
 
 ## Hello-world condensation example in Python, Julia and Matlab
 
+<details open>
+<summary>Python</summary>
 
+```Python
+import numpy as np
+from matplotlib import pyplot
+from PySDM.physics import si
+from PySDM.initialisation import spectral_sampling
+from PySDM.initialisation import multiplicities
+from PySDM.initialisation.spectra import Lognormal
+from PySDM.backends import CPU
+from PySDM import Builder
+from PySDM.dynamics import AmbientThermodynamics
+from PySDM.dynamics import Condensation
+from PySDM.environments import Parcel
+from PySDM.physics import formulae as phys
+from PySDM.initialisation.r_wet_init import r_wet_init
+from PySDM.products import ParticleMeanRadius, RelativeHumidity, CloudDropletConcentration
 
+builder = Builder(backend=CPU, n_sd=1)
+environment = Parcel(
+    dt=1 * si.s,
+    mass_of_dry_air=1 * si.kg,
+    p0=1000 * si.hPa,
+    q0=20 * si.g / si.kg,
+    T0=300 * si.K,
+    w= 1 * si.m / si.s
+)
+builder.set_environment(environment)
+
+kappa = 1 * si.dimensionless
+builder.add_dynamic(AmbientThermodynamics())
+builder.add_dynamic(Condensation(kappa=kappa))
+
+attributes = {}
+r_dry, specific_concentration = spectral_sampling.Logarithmic(
+            spectrum=Lognormal(
+                norm_factor=1000 / si.milligram,
+                m_mode=50 * si.nanometre,
+                s_geom=1.4 * si.dimensionless
+            ),
+            size_range=(10.633 * si.nanometre, 513.06 * si.nanometre)
+        ).sample(n_sd=builder.core.n_sd)
+attributes['dry volume'] = phys.volume(radius=r_dry)
+attributes['n'] = multiplicities.discretise_n(specific_concentration * environment.mass_of_dry_air)
+r_wet = r_wet_init(r_dry, environment, np.zeros_like(attributes['n']), kappa)
+attributes['volume'] = phys.volume(radius=r_wet)
+
+products = [ParticleMeanRadius(), RelativeHumidity(), CloudDropletConcentration(radius_range=(.5 * si.um, 25 * si.um))]
+
+core = builder.build(attributes, products)
+output = {product.name: [product.get()[0]] for product in core.products.values()}
+output['z'] = [environment['z'][0]]
+
+for step in range(100):
+    core.run(steps=10)
+    for product in core.products.values():
+        output[product.name].append(product.get()[0])
+    output['z'].append(environment['z'][0])
+
+fig, axs = pyplot.subplots(1, len(core.products), sharey="all")
+for i, (key, product) in enumerate(core.products.items()):
+    axs[i].step(output[key], output['z'])
+    axs[i].set_title(product.description)
+    axs[i].set_xlabel(product.unit)
+```
+</details>
+<details>
+<summary>Julia (click to expand)</summary>
+using PyCall
+PySDM = pyimport("PySDM")
+PySDM_backends = pyimport("PySDM.backends")
+PySDM_physics_formulae = pyimport("PySDM.physics.formulae")
+PySDM_environments = pyimport("PySDM.environments")
+PySDM_dynamics = pyimport("PySDM.dynamics")
+PySDM_initialisation = pyimport("PySDM.initialisation")
+PySDM_initialisation_spectra = pyimport("PySDM.initialisation.spectra")
[…]
