### 2021-04-28 273e676
fixing formulae references in README; removing tutorials
