+++ PySDM/physics/__init__.py
@@ -3,6 +3,7 @@ Crated at 2019
 """
 
 from .constants import si
+from .formulae import Formulae
 import PySDM.physics.condensation_coordinate
 import PySDM.physics.latent_heat
 import PySDM.physics.saturation_vapour_pressure
