+++ .github/workflows/main.yml
@@ -32,8 +32,13 @@ jobs:
       - run: |
           python -We -c "import PySDM"
 
+      # note: this installs PySDM-examples including its dependencies which inclde PySDM itself...
       - run: |
           pip install -r requirements.txt
+
+      # note: ... hence reverting to the currently tested version
+      - run: |
+          pip install -e .
 
       # TODO #184 (https://github.com/numba/numba/issues/6350#issuecomment-728174860)
       - if: matrix.platform == 'ubuntu-latest'
