### 2021-04-29 b2870fc
fixing explicit Euler call in chemistry_methods; reevant assert added to the test
