### 2021-04-29 3bd4cc7
TpRH using formulae in Thrust backend + fixes
