+++ PySDM_tests/unit_tests/backends/test_physics_methods.py
@@ -1,6 +1,6 @@
 import PySDM.physics.formulae
 from PySDM_tests.backends_fixture import backend
-from PySDM.physics.formulae import Formulae
+from PySDM.physics import si, Formulae
 import numpy as np
 
 
@@ -9,9 +9,9 @@ class TestPhysicsMethods:
     def test_temperature_pressure_RH(backend):
         # Arrange
         sut = backend(Formulae()).temperature_pressure_RH
-        rhod = backend.Storage.from_ndarray(np.asarray((1,1.1)))
-        thd = backend.Storage.from_ndarray(np.asarray((300,301)))
-        qv = backend.Storage.from_ndarray(np.asarray((.01,.02)))
+        rhod = backend.Storage.from_ndarray(np.asarray((1, 1.1)))
+        thd = backend.Storage.from_ndarray(np.asarray((300., 301)))
+        qv = backend.Storage.from_ndarray(np.asarray((.01, .02)))
 
         T = backend.Storage.from_ndarray(np.zeros_like(qv))
         p = backend.Storage.from_ndarray(np.zeros_like(qv))
@@ -21,6 +21,6 @@ class TestPhysicsMethods:
         sut(rhod, thd, qv, T, p, RH)
 
         # Assert
-        assert RH.amin() != 0
-        assert p.amin() != 0
-        assert T.amin() != 0
+        assert 282 * si.K < T.amin() < 283 * si.K
+        assert 820 * si.hPa < p.amin() < 830 * si.hPa
+        assert 1.10 < RH.amin() < 1.11
