+++ PySDM/backends/thrustRTC/impl/c_inline.py
@@ -4,9 +4,10 @@ import re
 from PySDM.backends.thrustRTC.impl.precision_resolver import PrecisionResolver
 from PySDM.physics import constants as const
 
+
 def c_inline(fun, **args):
-    prae = r"([+\-*/( ]|^)"
-    post = r"([ )/*\-+]|$)"
+    prae = r"([,+\-*/( ]|^)"
+    post = r"([ )/*\-+,]|$)"
     real_t = PrecisionResolver.get_C_type()
     real_fmt = ".32g"
     source = ''
@@ -28,4 +29,5 @@ def c_inline(fun, **args):
         source
     )
     source = eval(f'f"""{source}"""')
+    print(source)
     return f'{real_t}({source})'
