+++ PySDM/physics/__init__.py
@@ -14,3 +14,4 @@ import PySDM.physics.diffusion_kinetics
 import PySDM.physics.ventilation
 import PySDM.physics.state_variable_triplet
 import PySDM.physics.trivia
+import PySDM.physics.particle_advection
