### 2021-05-02 97fad42
debug leftover removed; one of the displacement tests should work now on Thrust
