### 2021-05-02 eda185a
formulae: cleanup, dthd_dt -> state vars triplet
