+++ PySDM/physics/formulae.py
@@ -92,11 +92,6 @@ def R(q):
 
 
 @_formula
-def dthd_dt(rhod, thd, T, dqv_dt, lv):
-    return - lv * dqv_dt / const.c_pd / T * thd * rhod
-
-
-@_formula
 def th_dry(th_std, qv):
     return th_std * np.power(1 + qv / const.eps, const.Rd / const.c_pd)
 
