+++ PySDM/physics/diffusion_kinetics/__init__.py
@@ -1,2 +1,2 @@
 from .fuchs_sutugin import FuchsSutugin
-from .fukuta_walter import FukutaWalter
+from .neglect import Neglect
