+++ PySDM/physics/diffusion_kinetics/fukuta_walter.py
@@ -1,2 +1,0 @@
-class FukutaWalter:
-    pass
