### 2021-05-02 e986dd4
diffusion kinetics -> formulae
