+++ PySDM/physics/diffusion_kinetics/neglect.py
@@ -1,2 +1,12 @@
 class Neglect:
-    pass
+    @staticmethod
+    def lambdaD(T):
+        return -1
+
+    @staticmethod
+    def lambdaK(T, p):
+        return -1
+
+    @staticmethod
+    def DK(DK, r, lmbd):
+        return DK
