+++ PySDM/physics/diffusion_kinetics/fuchs_sutugin.py
@@ -1,2 +1,16 @@
+from PySDM.physics import constants as const
+from numpy import sqrt
+
+
 class FuchsSutugin:
-    pass
+    @staticmethod
+    def lambdaD(T):
+        return const.D0 / sqrt(2 * const.Rv * T)
+
+    @staticmethod
+    def lambdaK(T, p):
+        return (4 / 5) * const.K0 * T / p / sqrt(2 * const.Rd * T)
+
+    @staticmethod
+    def DK(DK, r, lmbd):
+        return DK * (1 + lmbd/r) / (1 + 1.71 * lmbd/r + 1.33 * lmbd/r * lmbd/r)
