### 2021-05-02 a924850
chem trivia -> formulae
