+++ PySDM/physics/aqueous_chemistry/support.py
@@ -1,50 +1,56 @@
 from chempy import Substance
-
-from PySDM.physics.formulae import vant_hoff, tdep2enthalpy, arrhenius
 from PySDM.physics import si
 from PySDM.physics.constants import R_str, ROOM_TEMP, H_u, dT_u, M, Md, K_H2O
 import numpy as np
 
 
 class EqConst:
-    def __init__(self, constant_at_T0, dT, T_0):
+    def __init__(self, formulae, constant_at_T0, dT, T_0):
+        self.formulae = formulae
         self.K = constant_at_T0
-        self.dH = tdep2enthalpy(dT)
+        self.dH = formulae.trivia.tdep2enthalpy(dT)
         self.T0 = T_0
 
     def at(self, T):
-        return vant_hoff(self.K, self.dH, T, T_0=self.T0)
+        return self.formulae.trivia.vant_hoff(self.K, self.dH, T, T_0=self.T0)
 
 
 class KinConst:
-    def __init__(self, k, dT, T_0):
-        self.Ea = tdep2enthalpy(dT)
+    def __init__(self, formulae, k, dT, T_0):
+        self.formulae = formulae
+        self.Ea = formulae.trivia.tdep2enthalpy(dT)
         self.A = k * np.exp(self.Ea / (R_str * T_0))
 
     def at(self, T):
-        return arrhenius(self.A, self.Ea, T)
+        return self.formulae.trivia.arrhenius(self.A, self.Ea, T)
 
 
 # TODO #442: unit tests against chempy
-HENRY_CONST = {
-    "HNO3": EqConst(2.1e5 * H_u, 0 * dT_u, T_0=ROOM_TEMP),
-    "H2O2": EqConst(7.45e4 * H_u, 7300 * dT_u, T_0=ROOM_TEMP),
-    "NH3":  EqConst(62 * H_u, 4110 * dT_u, T_0=ROOM_TEMP),
-    "SO2":  EqConst(1.23 * H_u, 3150 * dT_u, T_0=ROOM_TEMP),
-    "CO2":  EqConst(3.4e-2 * H_u, 2440 * dT_u, T_0=ROOM_TEMP),
-    "O3":   EqConst(1.13e-2 * H_u, 2540 * dT_u, T_0=ROOM_TEMP),
-}
+class HenryConsts:
+    def __init__(self, formulae):
+        self.HENRY_CONST = {
+            "HNO3": EqConst(formulae, 2.1e5 * H_u, 0 * dT_u, T_0=ROOM_TEMP),
+            "H2O2": EqConst(formulae, 7.45e4 * H_u, 7300 * dT_u, T_0=ROOM_TEMP),
+            "NH3":  EqConst(formulae, 62 * H_u, 4110 * dT_u, T_0=ROOM_TEMP),
+            "SO2":  EqConst(formulae, 1.23 * H_u, 3150 * dT_u, T_0=ROOM_TEMP),
+            "CO2":  EqConst(formulae, 3.4e-2 * H_u, 2440 * dT_u, T_0=ROOM_TEMP),
+            "O3":   EqConst(formulae, 1.13e-2 * H_u, 2540 * dT_u, T_0=ROOM_TEMP),
+        }
+
 
 # Table 4 in Kreidenweis et al. 2003
-EQUILIBRIUM_CONST = {  # Reaction Specific units, K
-    "K_HNO3": EqConst(15.4 * M, 0 * dT_u, T_0=ROOM_TEMP),
[…]
