+++ PySDM/backends/numba/numba.py
@@ -28,6 +28,7 @@ class Numba(
     def __init__(self, formulae):
         self.formulae = formulae
         PhysicsMethods.__init__(self)
+        ChemistryMethods.__init__(self)
 
     @staticmethod
     def sanity_check():
