### 2021-05-02 4f5154f
more chem support -> trivia
