### 2021-05-03 2b23902
remaining formulae cleanup, removing FakeNumba!
