+++ PySDM_tests/smoke_tests/Arabas_and_Shima_2017/test_initialisation.py
@@ -4,9 +4,7 @@ Created at 2019
 
 import numpy as np
 import pytest
-
-from PySDM.physics import constants as const
-from PySDM.physics import formulae as phys
+from PySDM.physics import Formulae, constants as const
 from PySDM_examples.Arabas_and_Shima_2017.settings import setups
 from PySDM_examples.Arabas_and_Shima_2017.simulation import Simulation
 
@@ -59,5 +57,6 @@ class TestInitialisation:
         setup = setups[settings_idx]
         pv0 = setup.p0 / (1 + const.eps / setup.q0)
         pd0 = setup.p0 - pv0
+        phys = Formulae().trivia
         thd0 = phys.th_std(pd0, setup.T0)
         TestInitialisation.simulation_test('thd', thd0, setup)
