+++ PySDM_tests/smoke_tests/Kreidenweis_et_al_2003/test_Table_3.py
@@ -74,7 +74,7 @@ class TestTable3:
         np.testing.assert_allclose(output['p_env'][-1], 939 * si.mbar, rtol=.005)
         np.testing.assert_allclose(output['T_env'][-1], 284.2 * si.K, rtol=.005)
         np.testing.assert_allclose(
-            phys.rho_of_rhod_qv(rhod=output['rhod_env'][-1], qv=output['qv_env'][-1]*si.g/si.kg),
+            settings.formulae.state_variable_triplet.rho_of_rhod_qv(rhod=output['rhod_env'][-1], qv=output['qv_env'][-1]*si.g/si.kg),
             1.15 * si.kg / si.m**3,
             rtol=.005
         )
