+++ PySDM/physics/__init__.py
@@ -15,3 +15,4 @@ import PySDM.physics.ventilation
 import PySDM.physics.state_variable_triplet
 import PySDM.physics.trivia
 import PySDM.physics.particle_advection
+import PySDM.physics.hydrostatics
