+++ PySDM/physics/state_variable_triplet/rhod_thd_qv.py
@@ -29,3 +29,15 @@ class RhodThdQv:
     @staticmethod
     def th_dry(th_std, qv):
         return th_std * power(1 + qv / const.eps, const.Rd / const.c_pd)
+
+    @staticmethod
+    def rho_d(p, qv, theta_std):
+        return p * (1 - 1 / (1 + const.eps / qv)) / (power(p / const.p1000, const.Rd_over_c_pd) * const.Rd * theta_std)
+
+    @staticmethod
+    def rho_of_rhod_qv(rhod, qv):
+        return rhod * (1 + qv)
+
+    @staticmethod
+    def rhod_of_pd_T(pd, T):
+        return pd / const.Rd / T
