### 2021-05-05 831fcad
fixing Formulae __str__ to properly save POD values
