### 2021-05-05 7ee7226
cleanup in parcel example within README.md (Python code only for now)
