### 2021-05-05 8a76fa7
fixes to match r_wet_init change from prev commit + cleanups
