+++ PySDM_tests/smoke_tests/Yang_et_al_2018/test_just_do_it.py
@@ -13,6 +13,7 @@ import numpy as np
 scheme = ('default', 'BDF')
 adaptive = (True, False)
 
+
 @pytest.mark.parametrize("scheme", scheme)
 @pytest.mark.parametrize("adaptive", adaptive)
 def test_just_do_it(scheme, adaptive):
