+++ PySDM/products/state/water_mixing_ratio.py
@@ -7,6 +7,7 @@ class WaterMixingRatio(MomentProduct):
 
     def __init__(self, radius_range, name='ql', description_prefix='liquid'):
         self.radius_range = radius_range
+        self.volume_range = None
         super().__init__(
             name=name,
             unit='g/kg',
@@ -18,6 +19,7 @@ class WaterMixingRatio(MomentProduct):
     def register(self, builder):
         super().register(builder)
         self.volume_range = self.formulae.trivia.volume(np.asarray(self.radius_range))
+        self.radius_range = None
 
     def get(self):  # TODO #217
         self.download_moment_to_buffer('volume', rank=0, filter_range=self.volume_range, filter_attr='volume')
