### 2021-05-06 a6959fb
checking if prange works now in r_wet_init
