### 2021-05-06 79d0cd2
moving assert out of prange block
