### 2021-05-06 0ba43b3
numba physics_methods and README refactors
