+++ README.md
@@ -125,24 +125,10 @@ jupyter-notebook
 In order to depict the PySDM API with a practical example, the following
   listings provide sample code roughly reproducing the 
   Figure 2 from [Shima et al. 2009 paper](http://doi.org/10.1002/qj.441)
-  using PySDM from Python, Julia and Matlab (click ...).
+  using PySDM from Python, Julia and Matlab.
 It is a coalescence-only set-up in which the initial particle size 
   spectrum is exponential and is deterministically sampled to match
   the condition of each super-droplet having equal initial multiplicity:
-<details open>
-<summary>Python</summary>
-
-```Python
-from PySDM.physics import si
-from PySDM.initialisation.spectral_sampling import ConstantMultiplicity
-from PySDM.initialisation.spectra import Exponential
-
-n_sd = 2**15
-initial_spectrum = Exponential(norm_factor=8.39e12, scale=1.19e5 * si.um**3)
-attributes = {}
-attributes['volume'], attributes['n'] = ConstantMultiplicity(spectrum=initial_spectrum).sample(n_sd)
-```
-</details>
 <details>
 <summary>Julia (click to expand)</summary>
 
@@ -189,30 +175,25 @@ attributes = environment.init_attributes(pyargs( ...
 ));
 ```
 </details>
+<details open>
+<summary>Python</summary>
+
+```Python
+from PySDM.physics import si
+from PySDM.initialisation.spectral_sampling import ConstantMultiplicity
+from PySDM.initialisation.spectra import Exponential
+
+n_sd = 2**15
+initial_spectrum = Exponential(norm_factor=8.39e12, scale=1.19e5 * si.um**3)
+attributes = {}
+attributes['volume'], attributes['n'] = ConstantMultiplicity(spectrum=initial_spectrum).sample(n_sd)
+```
+</details>
 
 The key element of the PySDM interface is the [``Core``](https://github.com/atmos-cloud-sim-uj/PySDM/blob/master/PySDM/core.py) 
   class which instances are used to manage the system state and control the simulation.
 Instantiation of the ``Core`` class is handled by the ``Builder``
   as exemplified below:
-<details open>
-<summary>Python</summary>
-
-<!--exdown-cont-->
-```Python
-from PySDM import Builder
-from PySDM.environments import Box
-from PySDM.dynamics import Coalescence
-from PySDM.physics.coalescence_kernels import Golovin
-from PySDM.backends import CPU
-from PySDM.products.state import ParticlesVolumeSpectrum
-
-builder = Builder(n_sd=n_sd, backend=CPU)
-builder.set_environment(Box(dt=1 * si.s, dv=1e6 * si.m**3))
-builder.add_dynamic(Coalescence(kernel=Golovin(b=1.5e3 / si.s)))
-products = [ParticlesVolumeSpectrum()]
-particles = builder.build(attributes, products)
-```
-</details>
 <details>
 <summary>Julia (click to expand)</summary>
 
@@ -255,7 +236,25 @@ products = py.list({ PySDM_products.ParticlesVolumeSpectrum() });
 core = builder.build(attributes, products);
 ```
 </details>
+<details open>
+<summary>Python</summary>
 
+<!--exdown-cont-->
+```Python
+from PySDM import Builder
+from PySDM.environments import Box
+from PySDM.dynamics import Coalescence
+from PySDM.physics.coalescence_kernels import Golovin
+from PySDM.backends import CPU
+from PySDM.products.state import ParticlesVolumeSpectrum
+
[…]
