### 2021-05-06 cefb775
added TracyWelchPorter diffusion thermics (+ option to neglect)
