+++ PySDM/physics/diffusion_thermics/negclect.py
@@ -1,0 +1,7 @@
+from PySDM.physics import constants as const
+
+
+class Neglect:
+    @staticmethod
+    def D(T, p):
+        return const.D0
