+++ PySDM/physics/__init__.py
@@ -11,6 +11,7 @@ import PySDM.physics.hygroscopicity
 import PySDM.physics.drop_growth
 import PySDM.physics.surface_tension
 import PySDM.physics.diffusion_kinetics
+import PySDM.physics.diffusion_thermics
 import PySDM.physics.ventilation
 import PySDM.physics.state_variable_triplet
 import PySDM.physics.trivia
