### 2021-05-06 1bde872
random seed -> formulae
