+++ PySDM_tests/unit_tests/dynamics/coalescence/test_sdm_single_cell.py
@@ -209,7 +209,7 @@ class TestSDMSingleCell:
                 CountingRandom.calls += 1
                 super(CountingRandom, self).__call__(storage)
 
-        sut.rnd_opt.rnd = CountingRandom(n_sd)
+        sut.rnd_opt.rnd = CountingRandom(n_sd, seed=44)
         sut.stats_n_substep[:] = n_substeps
         sut.adaptive = adaptive
 
