+++ PySDM/backends/__init__.py
@@ -49,9 +49,8 @@ else:
     ThrustRTC.ENABLE = False
 
     class Random:
-        def __init__(self, size, seed=None):
+        def __init__(self, size, seed):
             self.size = size
-            seed = seed or np.random.randint(0, 2 * 16)
             self.generator = np.random.default_rng(seed)
 
         def __call__(self, storage):
