+++ requirements.txt
@@ -5,4 +5,4 @@ pytest
 exdown>=0.8.8
 
 # note: if cloning both PySDM and PySDM examples, consider "pip install -e"
-git+git://github.com/slayoo/PySDM-examples@8bc6fa4#egg=PySDM-examples
+git+git://github.com/slayoo/PySDM-examples@f8e1e77#egg=PySDM-examples
