### 2021-05-06 ed709d4
removing range and scale from products
