+++ PySDM_tests/smoke_tests/Arabas_et_al_2015/test_initialisation.py
@@ -6,7 +6,6 @@ import numpy as np
 from PySDM_examples.Arabas_et_al_2015.settings import Settings
 from PySDM_examples.Arabas_et_al_2015.simulation import Simulation
 from PySDM.physics.constants import si
-from PySDM.physics import formulae as phys
 from matplotlib import pyplot
 from PySDM.backends import CPU
 
