+++ PySDM_tests/smoke_tests/Yang_et_al_2018/test_initialisation.py
@@ -5,7 +5,6 @@ Created at 2019
 from PySDM_examples.Yang_et_al_2018.example import Simulation
 from PySDM_examples.Yang_et_al_2018.settings import Settings
 from PySDM.physics.constants import si
-from PySDM.physics import formulae as phys
 import matplotlib.pyplot as plt
 import numpy as np
 
