+++ PySDM/products/dynamics/aqueous_chemistry/aqueous_mass_spectrum.py
@@ -3,8 +3,6 @@ Created at 28.04.2020
 """
 
 import numpy as np
-
-from PySDM.physics import formulae as phys
 from PySDM.physics.constants import convert_to, si
 from PySDM.products.product import MomentProduct
 from PySDM.physics.aqueous_chemistry.support import AQUEOUS_COMPOUNDS
