+++ PySDM_tests/unit_tests/physics/test_formulae.py
@@ -5,7 +5,6 @@ Created at 2019
 from PySDM.physics import constants
 from PySDM.physics.formulae import Formulae
 from PySDM.physics.dimensional_analysis import DimensionalAnalysis
-from PySDM.physics import formulae as phys
 
 
 class TestFormulae:
