+++ PySDM/products/state/particles_size_spectrum.py
@@ -5,7 +5,6 @@ Created at 23.04.2020
 import numpy as np
 
 from PySDM.physics import constants as const
-from PySDM.physics import formulae as phys
 from PySDM.products.product import MomentProduct
 
 
