+++ PySDM/products/state/particles_concentration.py
@@ -3,9 +3,7 @@ Created at 05.02.2020
 """
 
 import numpy as np
-
 from PySDM.physics import constants as const
-from PySDM.physics import formulae as phys
 from PySDM.products.product import MomentProduct
 
 
