+++ setup.py
@@ -15,10 +15,10 @@ setup(
     install_requires=['ThrustRTC==0.3.10',
                       'CURandRTC==0.1.6',
                       'numba==0.53.1',
-                      'numpy>=1.19.4',
-                      'Pint>=0.11',
-                      'chempy>=0.7.10',
-                      'scipy>=1.4.1'],
+                      'numpy>=1.20.2',
+                      'Pint>=0.17',
+                      'chempy>=0.8.1',
+                      'scipy>=1.6.3'],
     author='https://github.com/atmos-cloud-sim-uj/PySDM/graphs/contributors',
     long_description=get_long_description(),
     long_description_content_type="text/markdown",
