### 2021-05-06 c3b2637
fix import (used within skipif string); one more cleanup
