### 2021-05-06 f4f7120
temporarily disabling travis and appveyor; fixing exdown installation
