+++ .github/workflows/matlab.yml
@@ -19,7 +19,7 @@ jobs:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v1
       - run: pip install -e .
-      - run: pip install exdown
+      - run: pip install exdown pytest
       - run: python -c "import exdown; blocks=exdown.extract('README.md', syntax_filter='Matlab'); f=open('readme.m', 'w'); f.writelines(block[0] for block in bocks); f.close()"
       - uses: matlab-actions/setup-matlab@v0
       - uses: matlab-actions/run-command@v0
