+++ PySDM_tests/devops_tests/test_readme_code.py
@@ -2,7 +2,7 @@ import exdown
 import pathlib
 
 
-test_readme = exdown.pytests(
+test_readme = exdown.pytests_from_file(
     pathlib.Path(__file__).parent.parent.parent.joinpath("README.md").absolute(),
     syntax_filter="Python",
     encoding='utf8'
