+++ README.md
@@ -159,19 +159,18 @@ attributes = environment.init_attributes(
 <summary>Matlab (click to expand)</summary>
 
 ```Matlab
-PySDM = py.importlib.import_module('PySDM');
-PySDM_initialisation_spectra = py.importlib.import_module('PySDM.initialisation.spectra');
-PySDM_initialisation_spectral_sampling = py.importlib.import_module('PySDM.initialisation.spectral_sampling');
-si = PySDM_physics.constants.si;
+si = py.importlib.import_module('PySDM.physics').si;
+ConstantMultiplicity = py.importlib.import_module('PySDM.initialisation.spectral_sampling').ConstantMultiplicity;
+Exponential = py.importlib.import_module('PySDM.initialisation.spectra').Exponential;
 
-n_sd = 2 ^ 15;
-initial_spectrum = PySDM_initialisation_spectra.Exponential(pyargs(...
+n_sd = 2^15;
+initial_spectrum = Exponential(pyargs(...
     'norm_factor', 8.39e12, ...
     'scale', 1.19e5 * si.um ^ 3 ...
 ));
 attributes = environment.init_attributes(pyargs( ...
     'spectral_discretisation', ...
-    PySDM_initialisation_spectral_sampling.ConstantMultiplicity(initial_spectrum)...
+    ConstantMultiplicity(initial_spectrum)...
 ));
 ```
 </details>
