+++ README.md
@@ -168,7 +168,7 @@ initial_spectrum = Exponential(pyargs(...
     'norm_factor', 8.39e12, ...
     'scale', 1.19e5 * si.um ^ 3 ...
 ));
-tmp = ConstantMultiplicity(initial_spectrum).sample(n_sd);
+tmp = ConstantMultiplicity(initial_spectrum).sample(int32(n_sd));
 attributes = py.dict(pyargs('volume', tmp{1}, 'n', tmp{2}));
 ```
 </details>
