+++ README.md
@@ -433,74 +433,57 @@ savefig("plot.svg")
 <summary>Matlab (click to expand)</summary>
 
 ```Matlab
-PySDM = py.importlib.import_module('PySDM');
-PySDM_environments = py.importlib.import_module('PySDM.environments');
-PySDM_products = py.importlib.import_module('PySDM.products');
-PySDM_initialisation = py.importlib.import_module('PySDM.initialisation');
-PySDM_initialisation_spectra = py.importlib.import_module('PySDM.initialisation.spectra');
-PySDM_initialisation_spectral_sampling = py.importlib.import_module('PySDM.initialisation.spectral_sampling');
-PySDM_physics = py.importlib.import_module('PySDM.physics');
-PySDM_dynamics = py.importlib.import_module('PySDM.dynamics');
-PySDM_backends = py.importlib.import_module('PySDM.backends');
+si = py.importlib.import_module('PySDM.physics').si;
+spectral_sampling = py.importlib.import_module('PySDM.initialisation').spectral_sampling;
+multiplicities = py.importlib.import_module('PySDM.initialisation').multiplicities;
+spectra = py.importlib.import_module('PySDM.initialisation').spectra;
+r_wet_init = py.importlib.import_module('PySDM.initialisation').r_wet_init;
+CPU = py.importlib.import_module('PySDM.backends').CPU;
+AmbientThermodynamics = py.importlib.import_module('PySDM.dynamics').AmbientThermodynamics;
+Condensation = py.importlib.import_module('PySDM.dynamics').Condensation;
+Parcel = py.importlib.import_module('PySDM.environments').Parcel;
+Builder = py.importlib.import_module('PySDM').Builder;
+products = py.importlib.import_module('PySDM.products');
 
-si = PySDM_physics.constants.si;
-
-n_sd = 100;
-kappa = 1;
-spectrum = PySDM_initialisation_spectra.Lognormal(pyargs(...
-    'norm_factor', 1000 / si.milligram, ...
-    'm_mode', 50 * si.nanometre, ...
-    's_geom', 1.4 * si.dimensionless ...
-));
-environment = PySDM_environments.Parcel(pyargs( ...
-    'dt', 1 * si.s, ...
-    'mass_of_dry_air', 1 * si.kg, ...
-    'p0', 1000 * si.hPa, ...
+env = Parcel(pyargs( ...
+    'dt', .25 * si.s, ...
+    'mass_of_dry_air', 1e3 * si.kg, ...
+    'p0', 1122 * si.hPa, ...
     'q0', 20 * si.g / si.kg, ...
     'T0', 300 * si.K, ...
-    'w', 1 * si.m / si.s ...
+    'w', 2.5 * si.m / si.s ...
 ));
-size_range = py.list({10.633 * si.nanometre, 513.06 * si.nanometre});
-radius_range = py.list({.5 * si.um, 25 * si.um});
-steps = 100;
-substeps = 10;
+spectrum = spectra.Lognormal(pyargs('norm_factor', 1e4/si.mg, 'm_mode', 50 * si.nm, 's_geom', 1.4));
+kappa = .5;
+cloud_range = py.tuple({.5 * si.um, 25 * si.um});
+output_interval = 1;
+output_points = 40;
+n_sd = 256;
 
-formulae = PySDM_physics.Formulae();
-builder = PySDM.Builder(pyargs( ...
-    'backend', PySDM_backends.CPU, ...
-    'n_sd', int32(n_sd), ...
-    'formulae', formulae ...
-));
-builder.set_environment(environment);
-builder.add_dynamic(PySDM_dynamics.AmbientThermodynamics())
-builder.add_dynamic(PySDM_dynamics.Condensation(pyargs('kappa', kappa)))
[…]
