+++ README.md
@@ -133,26 +133,17 @@ It is a coalescence-only set-up in which the initial particle size
 <summary>Julia (click to expand)</summary>
 
 ```Julia
-# using Pkg
-# Pkg.add("Conda")
-# using Conda
-# Conda.pip_interop(true)
-# Conda.pip("install --pre", "git+https://github.com/atmos-cloud-sim-uj/PySDM.git")
-# Pkg.add("PyCall")
+using Pkg
+Pkg.add("PyCall")
 
 using PyCall
-PySDM = pyimport("PySDM")
-PySDM_initialisation_spectra = pyimport("PySDM.initialisation.spectra")
-PySDM_initialisation_spectral_sampling = pyimport("PySDM.initialisation.spectral_sampling")
-si = PySDM.physics.si
+ConstantMultiplicity = pyimport("PySDM.initialisation.spectral_sampling").ConstantMultiplicity
+Exponential = pyimport("PySDM.initialisation.spectra").Exponential
 
 n_sd = 2^15
-initial_spectrum = PySDM_initialisation_spectra.Exponential(
-    norm_factor=8.39e12, scale=1.19e5 * si.um^3
-)
-attributes = environment.init_attributes(
-  spectral_discretisation=PySDM_initialisation_spectral_sampling.ConstantMultiplicity(spectrum=initial_spectrum)
-)
+initial_spectrum = Exponential(norm_factor=8.39e12, scale=1.19e5 * si.um^3)
+attributes = Dict()
+attributes['volume'], attributes['n'] = ConstantMultiplicity(spectrum=initial_spectrum).sample(n_sd)
 ```
 </details>
 <details>
@@ -196,18 +187,17 @@ Instantiation of the ``Core`` class is handled by the ``Builder``
 
 <!--exdown-cont-->
 ```Julia
-PySDM_backends = pyimport("PySDM.backends")
-PySDM_physics_coalescence_kernels = pyimport("PySDM.physics.coalescence_kernels")
-PySDM_environments = pyimport("PySDM.environments")
-PySDM_dynamics = pyimport("PySDM.dynamics")
-PySDM_products = pyimport("PySDM.products")
+Builder = pyimport("PySDM").Builder
+Box = pyimport("PySDM.environments").Box
+Coalescence = pyimport("PySDM.dynamics").Coalescence
+Golovin = pyimport("PySDM.physics.coalescence_kernels").Golovin
+ParticlesVolumeSpectrum = pyimport("PySDM.products.state").ParticlesVolumeSpectrum
 
-builder = PySDM.Builder(backend=PySDM_backends.CPU, n_sd=n_sd)
-environment = PySDM_environments.Box(dt=1 * si.s, dv=1e6 * si.m^3)
-builder.set_environment(environment)
-builder.add_dynamic(PySDM_dynamics.Coalescence(kernel=PySDM_physics_coalescence_kernels.Golovin(b=1.5e3 / si.s)))
-products = [PySDM_products.ParticlesVolumeSpectrum()] 
-core = builder.build(attributes, products)
+builder = PySDM.Builder(n_sd=n_sd, backend=CPU)
+builder.set_environment(Box(dt=1 * si.s, dv=1e6 * si.m^3))
+builder.add_dynamic(Coalescence(kernel=Golovin(b=1.5e3 / si.s)))
+products = [ParticlesVolumeSpectrum()] 
+particles = builder.build(attributes, products)
 ```
 </details>
 <details>
@@ -272,7 +262,7 @@ In the listing below, its usage is interleaved with plotting logic
 
 <!--exdown-cont-->
 ```Julia
-PySDM_physics_constants = pyimport("PySDM.physics.constants")
+rho_w = pyimport("PySDM.physics.constants").rho_w
 using Plots
 
[…]
