### 2021-05-07 fae269d
fix plot invocation
