### 2021-05-08 fefbb66
fix exdown <!--exdown-cont--> logic for Python pytest tests
