+++ .github/workflows/matlab.yml
@@ -1,4 +1,4 @@
-name: PySDM-matlab
+name: PySDM-readme.m
 
 defaults:
   run:
