+++ README.md
@@ -267,10 +267,10 @@ using Plots
 radius_bins_edges = 10 .^ range(log10(10*si.um), log10(5e3*si.um), length=32) 
 
 for step = 0:1200:3600
-    core.run(step - core.n_steps)
+    particles.run(step - particles.n_steps)
     plot!(
         radius_bins_edges[1:end-1] / si.um,
-        core.products["dv/dlnr"].get(radius_bins_edges) * rho_w / si.g,
+        particles.products["dv/dlnr"].get(radius_bins_edges) * rho_w / si.g,
         linetype=:steppost,
         xaxis=:log,
         xlabel="particle radius [µm]",
@@ -290,9 +290,9 @@ rho_w = py.importlib.import_module('PySDM.physics.constants').rho_w;
 radius_bins_edges = logspace(log10(10 * si.um), log10(5e3 * si.um), 32);
 
 for step = 0:1200:3600
-    core.run(int32(step - core.n_steps))
+    particles.run(int32(step - particles.n_steps))
     x = radius_bins_edges / si.um;
-    y = core.products{"dv/dlnr"}.get(py.numpy.array(radius_bins_edges)) * rho_w / si.g;
+    y = particles.products{"dv/dlnr"}.get(py.numpy.array(radius_bins_edges)) * rho_w / si.g;
     stairs(...
         x(1:end-1), ... 
         double(py.array.array('d',py.numpy.nditer(y))), ...
@@ -383,7 +383,7 @@ attributes["n"] = multiplicities.discretise_n(specific_concentration * env.mass_
 attributes["dry volume"] = builder.formulae.trivia.volume(radius=r_dry)
 attributes["volume"] = builder.formulae.trivia.volume(radius=r_wet) 
 
-core = builder.build(attributes, products=[
+particles = builder.build(attributes, products=[
     products.PeakSupersaturation(),
     products.CloudDropletEffectiveRadius(radius_range=cloud_range),
     products.CloudDropletConcentration(radius_range=cloud_range),
@@ -392,22 +392,22 @@ core = builder.build(attributes, products=[
     
 cell_id=1
 output = Dict("z" => Array{Float32}(undef, output_points+1))
-for (_, product) in core.products
+for (_, product) in particles.products
     output[product.name] = Array{Float32}(undef, output_points+1)
     output[product.name][1] = product.get()[cell_id]
 end 
 output["z"][1] = env.__getitem__("z")[cell_id]
     
 for step = 2:output_points+1
-    core.run(steps=output_interval)
-    for (_, product) in core.products
+    particles.run(steps=output_interval)
+    for (_, product) in particles.products
         output[product.name][step] = product.get()[cell_id]
     end 
     output["z"][step]=env.__getitem__("z")[cell_id]
 end 
 
 plots = []
-for (_, product) in core.products
+for (_, product) in particles.products
     append!(plots, [plot(output[product.name], output["z"], ylabel="z [m]", xlabel=product.unit, title=product.name)])
 end
 plot(plots..., layout=(1,4))
@@ -461,7 +461,7 @@ attributes = py.dict(pyargs( ...
     'volume', builder.formulae.trivia.volume(r_wet) ...
 ));
 
-core = builder.build(attributes, py.list({ ...
+particles = builder.build(attributes, py.list({ ...
     products.PeakSupersaturation(), ...
     products.CloudDropletEffectiveRadius(pyargs('radius_range', cloud_range)), ...
     products.CloudDropletConcentration(pyargs('radius_range', cloud_range)), ...
@@ -469,16 +469,16 @@ core = builder.build(attributes, py.list({ ...
 }));
 
 cell_id = int32(0);
-output_size = [output_points+1, 1 + length(py.list(core.products.keys()))];
[…]
