+++ .github/workflows/python.yml
@@ -13,7 +13,7 @@ on:
     - cron: '0 13 * * 4'
 
 jobs:
-  julia:
+  python:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
