### 2021-05-09 efaa8c2
README cleanups
