+++ .github/workflows/pdoc.yml
@@ -1,0 +1,27 @@
+name: generate and deploy API docs
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+        with:
+          persist-credentials: false
+      - uses: actions/setup-python@v2
+      - name: Build
+        run: |
+          pip3 install pdoc3
+          pdoc --html PySDM
+      - name: Deploy
+        if: github.ref == 'refs/heads/master'
+        uses: JamesIves/github-pages-deploy-action@4.1.1
+        with:
+          BRANCH: pdoc
+          FOLDER: build
+          CLEAN: true
