### 2021-05-09 d510fae
adding pip install to pdoc workflow
