+++ .github/workflows/pdoc.yml
@@ -8,6 +8,7 @@ on:
 
 jobs:
   deploy:
+    environment: pydoc
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
