+++ .github/workflows/pdoc.yml
@@ -24,6 +24,7 @@ jobs:
         #if: github.ref == 'refs/heads/master'
         uses: JamesIves/github-pages-deploy-action@4.1.1
         with:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           BRANCH: pdoc
           FOLDER: html/PySDM
           CLEAN: true
