+++ .github/workflows/pdoc.yml
@@ -8,7 +8,6 @@ on:
 
 jobs:
   deploy:
-    environment: pydoc
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
