+++ .github/workflows/pdoc.yml
@@ -20,10 +20,9 @@ jobs:
           pip install -e .
           pdoc --html PySDM
       - name: Deploy
-        #if: github.ref == 'refs/heads/master'
+        if: github.ref == 'refs/heads/master'
         uses: JamesIves/github-pages-deploy-action@4.1.1
         with:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
           BRANCH: pdoc
           FOLDER: html/PySDM
           CLEAN: true
