### 2021-05-09 2b39072
using SCM version from __init__
