+++ PySDM/backends/numba/impl/_algorithmic_methods.py
@@ -4,7 +4,6 @@ Created at 04.11.2019
 
 import numba
 import numpy as np
-from numba import void, f8, i8, prange, b1
 
 from PySDM.backends.numba import conf
 from PySDM.backends.numba.storage import Storage
@@ -46,7 +45,7 @@ class AlgorithmicMethods:
     def adaptive_sdm_gamma_body(gamma, idx, length, n, cell_id, dt_left, dt, dt_range, is_first_in_pair,
                                 stats_n_substep, stats_dt_min):
         dt_todo = np.empty_like(dt_left)
-        for cid in prange(len(dt_todo)):
+        for cid in numba.prange(len(dt_todo)):
             dt_todo[cid] = min(dt_left[cid], dt_range[1])
         for i in range(length // 2):  # TODO #401
             if gamma[i] == 0:
@@ -58,12 +57,12 @@ class AlgorithmicMethods:
             dt_optimal = max(dt_optimal, dt_range[0])
             dt_todo[cid] = min(dt_todo[cid], dt_optimal)
             stats_dt_min[cid] = min(stats_dt_min[cid], dt_optimal)
-        for i in prange(length // 2):
+        for i in numba.prange(length // 2):
             if gamma[i] == 0:
                 continue
             j, _ = pair_indices(i, idx, is_first_in_pair)
             gamma[i] *= dt_todo[cell_id[j]] / dt
-        for cid in prange(len(dt_todo)):
+        for cid in numba.prange(len(dt_todo)):
             dt_left[cid] -= dt_todo[cid]
             if dt_todo[cid] > 0:
                 stats_n_substep[cid] += 1
@@ -78,7 +77,7 @@ class AlgorithmicMethods:
     @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
     def calculate_displacement_body_1d(dim, scheme, displacement, courant, cell_origin, position_in_cell):
         length = displacement.shape[1]
-        for droplet in prange(length):
+        for droplet in numba.prange(length):
             # Arakawa-C grid
             _l = cell_origin[0, droplet]
             _r = cell_origin[0, droplet] + 1
@@ -88,7 +87,7 @@ class AlgorithmicMethods:
     @numba.njit(**{**conf.JIT_FLAGS, **{'parallel': False, 'cache': False}})
     def calculate_displacement_body_2d(dim, scheme, displacement, courant, cell_origin, position_in_cell):
         length = displacement.shape[1]
-        for droplet in prange(length):
+        for droplet in numba.prange(length):
             # Arakawa-C grid
             _l = (cell_origin[0, droplet], cell_origin[1, droplet])
             _r = (cell_origin[0, droplet] + 1 * (dim == 0), cell_origin[1, droplet] + 1 * (dim == 1))
@@ -119,11 +118,9 @@ class AlgorithmicMethods:
         return AlgorithmicMethods.cell_id_body(cell_id.data, cell_origin.data, strides.data)
 
     @staticmethod
-    @numba.njit(
-        void(i8[:], i8[:], i8, f8[:, :], f8[:], i8[:], b1[:]),
-        **conf.JIT_FLAGS)
+    @numba.njit(**conf.JIT_FLAGS)
     def coalescence_body(n, idx, length, attributes, gamma, healthy, is_first_in_pair):
-        for i in prange(length // 2):
+        for i in numba.prange(length // 2):
             if gamma[i] == 0:
                 continue
             j, k = pair_indices(i, idx, is_first_in_pair)
@@ -158,7 +155,7 @@ class AlgorithmicMethods:
[…]
