### 2021-05-10 ca4d821
one-line fix in the example code snippet to match API change
