### 2021-05-10 #3219 @slayoo
> @whedon generate pdf from branch JOSS

This was just to reflect an API change within one of the code snippets presented in the text:
https://github.com/atmos-cloud-sim-uj/PySDM/commit/ca4d8214ce49abeb6b6b08892428eb04b794bc1f
