### 2021-05-10 5d2d17d
fixing CUDA T()
