### 2021-05-10 5e4e969
fixing CUDA T() (re.sub -> str.replace)
