+++ PySDM/backends/thrustRTC/impl/c_inline.py
@@ -20,7 +20,7 @@ def c_inline(fun, **args):
         if stripped.startswith('def '):
             continue
         source += stripped
-    source = source.reaplce("power(", "pow(")
+    source = source.replace("power(", "pow(")
     source = re.sub("^return ", "", source)
     for arg in inspect.signature(fun).parameters:
         source = re.sub(f"{prae}({arg}){post}", f"\\1{real_t}({args[arg]})\\3", source)
