### 2021-05-10 834ea1f
fix power operators in trivia; introduce some scaffolding for GPU condensation; move explicit_euler and critical volume to backend so that parcel stuff works on ThrustRTC backend; cleanups
