### 2021-05-10 6cbe645
towards 0D GPU condensation
