+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -13,6 +13,65 @@ from functools import lru_cache
 
 class CondensationMethods:
     @staticmethod
+    def condensation(
+            solver,
+            n_cell, cell_start_arg,
+            v, v_cr, n, vdry, idx, rhod, thd, qv, dv, prhod, pthd, pqv, kappa,
+            rtol_x, rtol_thd, dt, counters, cell_order, RH_max, success, _
+    ):
+        n_threads = min(numba.get_num_threads(), n_cell)
+        CondensationMethods._condensation(
+            solver, n_threads, n_cell, cell_start_arg.data,
+            v.data, v_cr.data, n.data, vdry.data, idx.data,
+            rhod.data, thd.data, qv.data, dv, prhod.data, pthd.data, pqv.data, kappa,
+            rtol_x, rtol_thd, dt,
+            counters['n_substeps'].data,
+            counters['n_activating'].data,
+            counters['n_deactivating'].data,
+            counters['n_ripening'].data,
+            cell_order, RH_max.data, success.data
+        )
+
+    @staticmethod
+    @numba.njit(**{**conf.JIT_FLAGS, **{'cache': False}})
+    def _condensation(
+            solver, n_threads, n_cell, cell_start_arg,
+            v, v_cr, n, vdry, idx, rhod, thd, qv, dv_mean, prhod, pthd, pqv, kappa,
+            rtol_x, rtol_thd, dt,
+            counter_n_substeps, counter_n_activating, counter_n_deactivating, counter_n_ripening,
+            cell_order, RH_max, success
+    ):
+        for thread_id in numba.prange(n_threads):
+            for i in range(thread_id, n_cell, n_threads):
+                cell_id = cell_order[i]
+
+                cell_start = cell_start_arg[cell_id]
+                cell_end = cell_start_arg[cell_id + 1]
+                n_sd_in_cell = cell_end - cell_start
+                if n_sd_in_cell == 0:
+                    continue
+
+                dthd_dt = (pthd[cell_id] - thd[cell_id]) / dt
+                dqv_dt = (pqv[cell_id] - qv[cell_id]) / dt
+                rhod_mean = (prhod[cell_id] + rhod[cell_id]) / 2
+                md = rhod_mean * dv_mean
+
+                success_in_cell, qv_new, thd_new, substeps_hint, n_activating, n_deactivating, n_ripening, RH_max_in_cell = solver(
+                    v, v_cr, n, vdry,
+                    idx[cell_start:cell_end],
+                    kappa, thd[cell_id], qv[cell_id], dthd_dt, dqv_dt, md, rhod_mean,
+                    rtol_x, rtol_thd, dt, counter_n_substeps[cell_id]
+                )
+                counter_n_substeps[cell_id] = substeps_hint
+                counter_n_activating[cell_id] = n_activating
+                counter_n_deactivating[cell_id] = n_deactivating
+                counter_n_ripening[cell_id] = n_ripening
+                RH_max[cell_id] = RH_max_in_cell
+                success[cell_id] = success_in_cell
+                pqv[cell_id] = qv_new
+                pthd[cell_id] = thd_new
+
+    @staticmethod
     def make_adapt_substeps(jit_flags, dt, step_fake, dt_range, fuse, multiplier, within_tolerance):
         if not isinstance(multiplier, int):
             raise ValueError()
