+++ PySDM/backends/numba/numba.py
@@ -7,6 +7,7 @@ from PySDM.backends.numba.impl._pair_methods import PairMethods
 from PySDM.backends.numba.impl._physics_methods import PhysicsMethods
 from PySDM.backends.numba.impl._index_methods import IndexMethods
 from PySDM.backends.numba.impl.condensation_methods import CondensationMethods
+from .impl.moments_methods import MomentsMethods
 from PySDM.backends.numba.impl._chemistry_methods import ChemistryMethods
 from PySDM.backends.numba.random import Random as ImportedRandom
 from PySDM.backends.numba.storage import Storage as ImportedStorage
@@ -18,7 +19,8 @@ class Numba(
     IndexMethods,
     PhysicsMethods,
     CondensationMethods,
-    ChemistryMethods
+    ChemistryMethods,
+    MomentsMethods
 ):
     Storage = ImportedStorage
     Random = ImportedRandom
