### 2021-05-12 a2382a6
development of the condensation prototype on the GPU
add Storage.all()
replace python literals with constants
improve Storage.__setitem__ to handle np.dtypes (np.float64 has data, dtype, shape)
add GPU backend to test_water_mass_conservation, test_just_do_it
