+++ PySDM/backends/thrustRTC/impl/condensation_methods.py
@@ -2,19 +2,45 @@ from PySDM.physics import constants as const
 from PySDM.backends.thrustRTC.conf import NICE_THRUST_FLAGS
 from PySDM.backends.thrustRTC.impl import nice_thrust, c_inline
 from .precision_resolver import PrecisionResolver
+from PySDM.backends.thrustRTC.bisection import BISECTION
 from ..conf import trtc
+from PySDM.backends.thrustRTC.storage import Storage
 
 
 class CondensationMethods:
     def __init__(self):
         phys = self.formulae
+        self.RH_rtol = None
 
         # TODO #228: precision for consts
         self.__calculate_m_l = trtc.For(("ml", "v", "n", "cell_id"), "i", f'''
             atomicAdd((real_type*) &ml[cell_id[i]], n[i] * v[i] * {const.rho_w}); 
         '''.replace("real_type", PrecisionResolver.get_C_type()))
 
-        self.__update_volume = trtc.For(("v", "vdry"), "i", f'''
+        self.__update_volume = trtc.For(("v", "vdry", "T", "kappa", "DTp", "lambdaD", "lambdaK", "RH", "pvs", "lv", "dt", "p", "RH_rtol", "rtol_x", "max_iters"), "i",
+            '''
+            struct Args {
+                real_type x_old, dt, p, kappa, rd3, T, RH, lv, pvs, Dr, Kr;
+            };
+            struct Minfun {
+                static __device__ real_type value(const real_type &x_new, void* args_p)
+                {
+                '''
+                f'''
+                    auto args = static_cast<Args*>(args_p);
+                    auto vol = {c_inline(phys.condensation_coordinate.volume, x="x_new")};
+                    auto r_new = {c_inline(phys.trivia.radius, volume="vol")};
+                    auto sgm = {c_inline(phys.surface_tension.sigma, T="args->T", v_wet="vol", v_dry="const.pi_4_3 * args->rd3")};
+                    auto RH_eq = {c_inline(phys.hygroscopicity.RH_eq, r="r_new", T="args->T", kp="args->kappa", rd3="args->rd3", sgm="sgm")};
+                    auto r_dr_dt = {c_inline(phys.drop_growth.r_dr_dt, RH_eq="RH_eq", T="args->T", RH="args->RH", lv="args->lv", pvs="args->pvs", D="args->Dr", K="args->Kr")};
+                    return args->x_old - x_new + args->dt * {c_inline(phys.condensation_coordinate.dx_dt, x="x_new", r_dr_dt="r_dr_dt")};
+                '''
+                '''
+                }
+            };
+            '''
+            f'''
+            {BISECTION}
             auto x_old = {c_inline(phys.condensation_coordinate.x, volume="v[i]")};
             auto r_old = {c_inline(phys.trivia.radius, volume="v[i]")};
             auto x_insane = {c_inline(phys.condensation_coordinate.x, volume="vdry[i]/100")};
@@ -22,7 +48,71 @@ class CondensationMethods:
             auto sgm = {c_inline(phys.surface_tension.sigma, T="T", v_wet="v[i]", v_dry="vdry[i]")};
             auto RH_eq = {c_inline(phys.hygroscopicity.RH_eq, r="r_old", T="T", kp="kappa", rd3="rd3", sgm="sgm")};
 
-            v[i] = 0;
+            real_type Dr, Kr, r_dr_dt_old, dx_old;
+            Args args;
+            real_type x_new;
+            if ( ! {c_inline(phys.trivia.within_tolerance, error_estimate="abs(RH - RH_eq)", value="RH", rtol="RH_rtol")})
+            {{
[…]
