+++ PySDM_tests/smoke_tests/Yang_et_al_2018/test_just_do_it.py
@@ -1,14 +1,16 @@
 """
 Created at 2020
 """
+import pytest
+import numpy as np
 
 from PySDM_examples.Yang_et_al_2018.example import Simulation
 from PySDM_examples.Yang_et_al_2018.settings import Settings
 from PySDM.physics.constants import si
 from PySDM.backends.numba.test_helpers import bdf
-import pytest
-import numpy as np
 
+# noinspection PyUnresolvedReferences
+from PySDM_tests.backends_fixture import backend
 
 scheme = ('default', 'BDF')
 adaptive = (True, False)
@@ -16,7 +18,7 @@ adaptive = (True, False)
 
 @pytest.mark.parametrize("scheme", scheme)
 @pytest.mark.parametrize("adaptive", adaptive)
-def test_just_do_it(scheme, adaptive):
+def test_just_do_it(backend, scheme, adaptive):
     # Arrange
     if scheme == 'BDF' and not adaptive:
         return
@@ -28,7 +30,7 @@ def test_just_do_it(scheme, adaptive):
     elif not adaptive:
         settings.dt_max = 1 * si.second
 
-    simulation = Simulation(settings)
+    simulation = Simulation(settings, backend)
     if scheme == 'BDF':
         bdf.patch_core(simulation.core)
 
