+++ PySDM/physics/constants.py
@@ -19,6 +19,9 @@ def convert_to(value, unit):
 
 pi = sci.pi
 pi_4_3 = pi * 4 / 3
+three = 3
+one_third = 1/3
+two_thirds = 2/3
 
 Md = (
         0.78 * Substance.from_formula('N2').mass * si.gram / si.mole +
@@ -31,7 +34,10 @@ R_str = sci.R * si.joule / si.kelvin / si.mole
 eps = Mv / Md
 Rd = R_str / Md
 Rv = R_str / Mv
+
 D0 = 2.26e-5 * si.metre ** 2 / si.second
+D_exp = 1.81
+
 K0 = 2.4e-2 * si.joules / si.metres / si.seconds / si.kelvins
 
 p1000 = 1000 * si.hectopascals

[…]
