+++ PySDM/physics/breakup_fragmentations/always_n.py
@@ -1,0 +1,18 @@
+"""
+Created at 13.05.2021 by edejong
+"""
+
+import numpy as np
+
+
+class AlwaysN:
+
+    def __init__(self, n):
+        self.core = None
+        self.N = n
+
+    def __call__(self, output, is_first_in_pair):
+        output = self.N
+
+    def register(self, builder):
+        self.core = builder.core
