+++ PySDM/dynamics/breakup.py
@@ -1,0 +1,168 @@
+"""
+Created at 05.13.21 by edejong
+"""
+
+import numpy as np
+from PySDM.physics import si
+from PySDM.dynamics.impl.random_generator_optimizer import RandomGeneratorOptimizer
+import warnings
+
+default_dt_coal_range = (.1 * si.second, 100 * si.second)
+
+
+class Breakup:
+
+    def __init__(self,
+                 kernel,
+                 fragmentation,
+                 croupier=None,
+                 optimized_random=False,
+                 substeps: int = 1,
+                 adaptive: bool = False,
+                 dt_coal_range=default_dt_coal_range
+                 ):
+        assert substeps == 1 or adaptive is False
+
+        self.core = None
+        self.enable = True
+
+        self.kernel = kernel
+        self.fragmentation = fragmentation
+
+        assert dt_coal_range[0] > 0
+        self.croupier = croupier
+        self.optimized_random = optimized_random
+        self.__substeps = substeps
+        self.adaptive = adaptive
+        self.stats_n_substep = None
+        self.stats_dt_min = None
+        self.dt_coal_range = tuple(dt_coal_range)
+
+        self.kernel_temp = None
+        self.norm_factor_temp = None
+        self.prob = None
+        self.is_first_in_pair = None
+        self.dt_left = None
+
+        self.collision_rate = None
+        self.collision_rate_deficit = None
+
+    def register(self, builder):
+        self.core = builder.core
+        self.rnd_opt = RandomGeneratorOptimizer(optimized_random=self.optimized_random,
+                                                dt_min=self.dt_coal_range[0],
+                                                seed=self.core.formulae.seed)
+        self.optimised_random = None
+
+        if self.core.n_sd < 2:
+            raise ValueError("No one to collide with!")
+        if self.dt_coal_range[1] > self.core.dt:
+            self.dt_coal_range = (self.dt_coal_range[0], self.core.dt)
+        assert self.dt_coal_range[0] <= self.dt_coal_range[1]
+
+        self.kernel_temp = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=float)
+        self.norm_factor_temp = self.core.Storage.empty(self.core.mesh.n_cell, dtype=float)  # TODO #372
+        self.prob = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=float)
+        self.is_first_in_pair = self.core.PairIndicator(self.core.n_sd)
+        self.dt_left = self.core.Storage.empty(self.core.mesh.n_cell, dtype=float)
+
+        self.stats_n_substep = self.core.Storage.empty(self.core.mesh.n_cell, dtype=int)
+        self.stats_n_substep[:] = 0 if self.adaptive else self.__substeps
+        self.stats_dt_min = self.core.Storage.empty(self.core.mesh.n_cell, dtype=float)
+        self.stats_dt_min[:] = np.nan
+
+        self.rnd_opt.register(builder)
+        self.kernel.register(builder)
+
+        if self.croupier is None:
+            self.croupier = self.core.backend.default_croupier
+        
+        
+        # TODO Emily: check whether this is still correct to use.
+        self.collision_rate = self.core.Storage.from_ndarray(np.zeros(self.core.mesh.n_cell, dtype=int))
+        self.collision_rate_deficit = self.core.Storage.from_ndarray(np.zeros(self.core.mesh.n_cell, dtype=int))
+
+    def __call__(self):
+        if self.enable:
+            if not self.adaptive:
+                for _ in range(self.__substeps):
+                    self.step()
+            else:
+                self.dt_left[:] = self.core.dt
+
+                while self.core.particles.get_working_length() != 0:
[…]
