+++ PySDM/physics/__init__.py
@@ -17,3 +17,4 @@ import PySDM.physics.state_variable_triplet
 import PySDM.physics.trivia
 import PySDM.physics.particle_advection
 import PySDM.physics.hydrostatics
+import PySDM.physics.breakup_fragmentations
