+++ PySDM/products/dynamics/__init__.py
@@ -7,3 +7,4 @@ from .condensation import *
 from .displacement import *
 from .dynamic_wall_time import DynamicWallTime
 from .aqueous_chemistry import *
+from .breakup import *
