+++ PySDM/products/dynamics/breakup/__init__.py
@@ -1,0 +1,5 @@
+"""
+Created at 14.05.21 by edejong
+"""
+
+from .fragmentation import BreakupFragments
