+++ appveyor.yml
@@ -32,7 +32,7 @@ install:
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
       
       # Add-Content requirements.txt "`npywin32>=300"
-      iex "$env:PYTHON -m pip install -r requirements.txt"
+      iex "$env:PYTHON -m pip install -r test-time-requirements.txt"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
 
 test_script:
