+++ .travis.yml
@@ -30,7 +30,7 @@ jobs:
       env: PATH=/c/Python37:/c/Python37/Scripts:$PATH
 
 install:
-  - pip$PY_SFX install -U pytest-cov codecov setuptools>=50.3.2  # TODO #462
+  - pip$PY_SFX install -U pytest-cov codecov setuptools>=50.3.2
   - |
     # https://stackoverflow.com/questions/55486248/deprecationwarning-in-python-3-6-and-3-7-with-pillow-distutils-imp
     if [[ $TRAVIS_OS_NAME == linux ]]; then
