### 2021-05-16 b653e8e
Adding debug print statements
