+++ PySDM/dynamics/breakup.py
@@ -48,6 +48,7 @@ class Breakup:
         self.collision_rate = None
         self.collision_rate_deficit = None
         self.n_fragment = None
+        print("initialized")
 
     def register(self, builder):
         self.core = builder.core
@@ -84,6 +85,7 @@ class Breakup:
         
         self.collision_rate = self.core.Storage.from_ndarray(np.zeros(self.core.mesh.n_cell, dtype=int))
         self.collision_rate_deficit = self.core.Storage.from_ndarray(np.zeros(self.core.mesh.n_cell, dtype=int))
+        print("registered")
 
     def __call__(self):
         if self.enable:
@@ -102,6 +104,7 @@ class Breakup:
             self.rnd_opt.reset()
 
     def step(self):
+        print("called step")
         # (1) Make the superdroplet list 
         pairs_rand, rand = self.rnd_opt.get_random_arrays()
         print("list made")
@@ -120,8 +123,11 @@ class Breakup:
         
         # (4) Compute gamma...
         self.compute_gamma(self.prob, rand, self.is_first_in_pair)
+        print('computed gamma')
+        
         # (5) Perform the collisional-breakup step: 
         self.core.particles.breakup(gamma=self.prob, n_fragment=self.n_fragment, is_first_in_pair=self.is_first_in_pair)
+        print('breakup done')
         
         if self.adaptive:
             self.core.particles.cut_working_length(self.core.particles.adaptive_sdm_end(self.dt_left))
@@ -147,12 +153,13 @@ class Breakup:
         
     # (4a) Compute n_fragment
     def compute_n_fragment(self, n_fragment, is_first_in_pair):
+        #self.fragmentation(self.n_fragment, is_first_in_pair)
         self.fragmentation(self.fragmentation_temp, is_first_in_pair)
         n_fragment.max(self.core.particles['n'], is_first_in_pair)
-        n_fragment *= self.fragmentation_temp
+        n_fragment = self.fragmentation_temp
         
         # TODO: normalize?
-        self.core.normalize(n_fragment, self.norm_factor_temp)
+        #self.core.normalize(n_fragment, self.norm_factor_temp)
 
     # (4) Compute gamma, i.e. whether the collision leads to breakup
     def compute_gamma(self, prob, rand, is_first_in_pair):
