+++ setup.py
@@ -1,4 +1,5 @@
 from setuptools import setup, find_packages
+import os
 
 
 def get_long_description():
