### 2021-05-17 5b242a6
Fixed issue: do not use ConstK
