### 2021-05-18 b205daf
disable ripening test in S&H smoke test
