### 2021-05-18 db71d3c
disable deactivation assert as well
