### 2021-05-18 7f5e16e
add minimal requirements for non-CI env for ThrustRTC, CURandRTC and numba
