+++ PySDM/physics/breakup_fragmentations/__init__.py
@@ -2,3 +2,4 @@
 Created 13.05.2021 by edejong
 """
 from .always_n import AlwaysN
+from .slams import SLAMS
