### 2021-05-19 c34b1fa
test_initialisation.py for GPU
