### 2021-05-19 38054e0
multi-cell condensation on GPU
added backend fixture to test_spin_up
added backend fixture to test_export
