+++ PySDM_tests/smoke_tests/Arabas_et_al_2015/test_export.py
@@ -8,6 +8,9 @@ from PySDM_examples.Arabas_et_al_2015.simulation import Simulation
 from PySDM_examples.Arabas_et_al_2015.storage import Storage
 import tempfile
 
+# noinspection PyUnresolvedReferences
+from PySDM_tests.backends_fixture import backend
+
 
 def test_export(tmp_path):
     # Arrange
@@ -16,7 +19,7 @@ def test_export(tmp_path):
     settings.output_interval = settings.dt
 
     storage = Storage()
-    simulator = Simulation(settings, storage)
+    simulator = Simulation(settings, storage, backend)
     _, temp_file = tempfile.mkstemp(dir=tmp_path, suffix='.nc')
     sut = NetCDFExporter(storage, settings, simulator, temp_file)
 
