+++ PySDM/backends/thrustRTC/impl/condensation_methods.py
@@ -6,6 +6,8 @@ from PySDM.backends.thrustRTC.bisection import BISECTION
 from ..conf import trtc
 from PySDM.backends.thrustRTC.storage import Storage
 
+keys = ['T', 'p', 'pv', 'lv', 'pvs', 'RH', 'DTp', 'lambdaK', 'lambdaD']
+
 
 class CondensationMethods:
     def __init__(self):
@@ -17,17 +19,24 @@ class CondensationMethods:
             atomicAdd((real_type*) &ml[cell_id[i]], n[i] * v[i] * {const.rho_w}); 
         '''.replace("real_type", PrecisionResolver.get_C_type()))
 
-        self.__update_volume = trtc.For(("v", "vdry", "T", "kappa", "DTp", "lambdaD", "lambdaK", "RH",
-                                         "pvs", "lv", "dt", "p", "RH_rtol", "rtol_x", "max_iters"), "i",
-            '''
-            struct Args {
+        self.__update_volume = trtc.For(("v", "vdry", *keys, "kappa", "dt", "RH_rtol", "rtol_x", "max_iters", "cell_id"), "i",
+            f'''
+            auto _T = T[cell_id[i]];
+            auto _p = p[cell_id[i]];
+            auto _pv = pv[cell_id[i]];
+            auto _lv = lv[cell_id[i]];
+            auto _pvs = pvs[cell_id[i]];
+            auto _RH = RH[cell_id[i]];
+            auto _DTp = DTp[cell_id[i]];
+            auto _lambdaK = lambdaK[cell_id[i]];
+            auto _lambdaD = lambdaD[cell_id[i]];
+            
+            struct Args {{
                 real_type x_old, dt, p, kappa, rd3, T, RH, lv, pvs, Dr, Kr;
-            };
-            struct Minfun {
+            }};
+            struct Minfun {{
                 static __device__ real_type value(const real_type &x_new, void* args_p)
-                {
-                '''
-                f'''
+                {{
                     auto args = static_cast<Args*>(args_p);
                     auto vol = {c_inline(phys.condensation_coordinate.volume, x="x_new")};
                     auto r_new = {c_inline(phys.trivia.radius, volume="vol")};
@@ -35,29 +44,26 @@ class CondensationMethods:
                     auto RH_eq = {c_inline(phys.hygroscopicity.RH_eq, r="r_new", T="args->T", kp="args->kappa", rd3="args->rd3", sgm="sgm")};
                     auto r_dr_dt = {c_inline(phys.drop_growth.r_dr_dt, RH_eq="RH_eq", T="args->T", RH="args->RH", lv="args->lv", pvs="args->pvs", D="args->Dr", K="args->Kr")};
                     return args->x_old - x_new + args->dt * {c_inline(phys.condensation_coordinate.dx_dt, x="x_new", r_dr_dt="r_dr_dt")};
-                '''
-                '''
-                }
-            };
-            '''
-            f'''
+                }}
+            }};
             {BISECTION}
+            
             auto x_old = {c_inline(phys.condensation_coordinate.x, volume="v[i]")};
             auto r_old = {c_inline(phys.trivia.radius, volume="v[i]")};
             auto x_insane = {c_inline(phys.condensation_coordinate.x, volume="vdry[i]/100")};
             auto rd3 = vdry[i] / {const.pi_4_3};
[…]
