### 2021-05-19 d0f536a
refactored bisection (for FakeThrust) added return_type to c_inline refactored condensation_methods.py (for FakeThrust) extended cpp2python fixed test_export
