+++ PySDM/backends/thrustRTC/impl/condensation_methods.py
@@ -19,8 +19,13 @@ class CondensationMethods:
             atomicAdd((real_type*) &ml[cell_id[i]], n[i] * v[i] * {const.rho_w}); 
         '''.replace("real_type", PrecisionResolver.get_C_type()))
 
+        args_vars = ('x_old', 'dt', '_p', 'kappa', 'rd3', '_T', '_RH', '_lv', '_pvs', 'Dr', 'Kr')
+
+        def args(arg):
+            return f"args[{args_vars.index(arg)}]"
+
         self.__update_volume = trtc.For(("v", "vdry", *keys, "kappa", "dt", "RH_rtol", "rtol_x", "max_iters", "cell_id"), "i",
-            f'''
+            f'''            
             auto _T = T[cell_id[i]];
             auto _p = p[cell_id[i]];
             auto _pv = pv[cell_id[i]];
@@ -31,19 +36,15 @@ class CondensationMethods:
             auto _lambdaK = lambdaK[cell_id[i]];
             auto _lambdaD = lambdaD[cell_id[i]];
             
-            struct Args {{
-                real_type x_old, dt, p, kappa, rd3, T, RH, lv, pvs, Dr, Kr;
-            }};
             struct Minfun {{
-                static __device__ real_type value(const real_type &x_new, void* args_p)
-                {{
-                    auto args = static_cast<Args*>(args_p);
+                static __device__ real_type value(real_type x_new, void* args_p) {{
+                    auto args = static_cast<real_type*>(args_p);
                     auto vol = {c_inline(phys.condensation_coordinate.volume, x="x_new")};
                     auto r_new = {c_inline(phys.trivia.radius, volume="vol")};
-                    auto sgm = {c_inline(phys.surface_tension.sigma, T="args->T", v_wet="vol", v_dry="const.pi_4_3 * args->rd3")};
-                    auto RH_eq = {c_inline(phys.hygroscopicity.RH_eq, r="r_new", T="args->T", kp="args->kappa", rd3="args->rd3", sgm="sgm")};
-                    auto r_dr_dt = {c_inline(phys.drop_growth.r_dr_dt, RH_eq="RH_eq", T="args->T", RH="args->RH", lv="args->lv", pvs="args->pvs", D="args->Dr", K="args->Kr")};
-                    return args->x_old - x_new + args->dt * {c_inline(phys.condensation_coordinate.dx_dt, x="x_new", r_dr_dt="r_dr_dt")};
+                    auto sgm = {c_inline(phys.surface_tension.sigma, T=args('_T'), v_wet="vol", v_dry=f"const.pi_4_3 * {args('rd3')}")};
+                    auto RH_eq = {c_inline(phys.hygroscopicity.RH_eq, r="r_new", T=args('_T'), kp=args("kappa"), rd3=args("rd3"), sgm="sgm")};
+                    auto r_dr_dt = {c_inline(phys.drop_growth.r_dr_dt, RH_eq="RH_eq", T=args("_T"), RH=args("_RH"), lv=args("_lv"), pvs=args("_pvs"), D=args("Dr"), K=args("Kr"))};
+                    return {args("x_old")} - x_new + {args("dt")} * {c_inline(phys.condensation_coordinate.dx_dt, x="x_new", r_dr_dt="r_dr_dt")};
                 }}
             }};
             {BISECTION}
@@ -55,21 +56,23 @@ class CondensationMethods:
             auto sgm = {c_inline(phys.surface_tension.sigma, T="_T", v_wet="v[i]", v_dry="vdry[i]")};
[…]
