+++ PySDM/physics/coalescence_kernels/__init__.py
@@ -2,4 +2,4 @@ from .golovin import Golovin
 from .geometric import Geometric
 from .hydrodynamic import Hydrodynamic
 from .electric import Electric
-from .constantK import ConstantK
+from .paramWithEff import ParamWithEff
