### 2021-05-20 9dae700
fixing volume coordinate code to work with c_inline, refactoring test_conservation
