### 2021-05-21 14710ce
Revert "Clean up, work on docs"
This reverts commit e4413cfd0bd121707376676c484904e938c2fc0f.
