+++ PySDM/physics/coalescence_efficiencies/__init__.py
@@ -1,1 +1,0 @@
-from .berry1967 import Berry1967
