### 2021-05-21 f8fa720
Revert "Fixed issue: do not use ConstK"
This reverts commit 5b242a6be36ffa3c6794270ad4c50b4eea89d3e2.

Conflicts:
	PySDM/dynamics/breakup.py
	PySDM_tests/breakup_test.ipynb
