+++ PySDM_tests/breakup_test.py
@@ -1,62 +1,0 @@
-import sys
-import numpy as np
-
-from PySDM.backends import CPU
-from PySDM.builder import Builder
-from PySDM.environments import Box
-from PySDM.dynamics import Coalescence
-from PySDM.dynamics import Breakup
-from PySDM.initialisation.spectral_sampling import ConstantMultiplicity
-
-from PySDM.products.state import ParticlesVolumeSpectrum, ParticlesConcentration, ParticleMeanRadius
-
-from PySDM.initialisation.spectra import Exponential
-from PySDM.physics.coalescence_kernels import ConstantK, Golovin
-from PySDM.physics.breakup_fragmentations import AlwaysN
-from PySDM.physics.constants import si
-from PySDM.physics.formulae import Formulae
-
-class Settings:
-
-    def __init__(self):
-        self.formulae = Formulae()
-        self.n_sd = 2
-        self.n_part = 2 / si.cm**3
-        self.X0 = self.formulae.trivia.volume(radius=30.531 * si.micrometres)
-        self.dv = 1 * si.cm**3
-        self.norm_factor = self.n_part * self.dv
-        self.rho = 1000 * si.kilogram / si.metre**3
-        self.dt = 1 * si.seconds
-        self.adaptive = False
-        self.seed = 44
-        self._steps = [0, 1]
-        self.kernel = Golovin(b=1.5e3 / si.second)
-        self.breakup_kernel = ConstantK(a=1 * si.cm**3 / si.second)
-        self.fragmentation = AlwaysN(n=2)
-        self.spectrum = Exponential(norm_factor=self.norm_factor, scale=self.X0)
-        self.radius_bins_edges = np.logspace(np.log10(10 * si.um), np.log10(100 * si.um), num=128, endpoint=True)
-        self.radius_range = [0 * si.um, 1e6 * si.um]
-
-    @property
-    def output_steps(self):
-        return [int(step/self.dt) for step in self._steps]
-
-
-settings = Settings()
-backend = CPU
-
-builder = Builder(n_sd=settings.n_sd, backend=backend, formulae=settings.formulae)
-builder.set_environment(Box(dv=settings.dv, dt=settings.dt))
-attributes = {}
-attributes['volume'], attributes['n'] = ConstantMultiplicity(settings.spectrum).sample(settings.n_sd)
-breakup = Breakup(settings.breakup_kernel, settings.fragmentation, adaptive=settings.adaptive)
-builder.add_dynamic(breakup)
-products = [ParticlesVolumeSpectrum(), ParticleMeanRadius(), ParticlesConcentration(radius_range = settings.radius_range)]
-core = builder.build(attributes, products)
-
-for step in settings.output_steps:
-    print(step)
-    print(core.n_steps)
-    print(step - core.n_steps)
-    core.run(step - core.n_steps)
-    print(core.products['radius_m1'].get(), core.products['n_a_cm3'].get())
