### 2021-05-21 56a0a5d
At last, fixed the issue of constant values not registering
