### 2021-05-21 9684fe1
Clean up print statements
