### 2021-05-21 432adca
more readable code in chemistry methods
