### 2021-05-23 6e12f6d
making c_inline compatible with precision resolver
