+++ PySDM/backends/thrustRTC/impl/c_inline.py
@@ -23,11 +23,11 @@ def c_inline(fun, return_type=PrecisionResolver.get_C_type(), **args):
     source = source.replace("power(", "pow(")
     source = re.sub("^return ", "", source)
     for arg in inspect.signature(fun).parameters:
-        source = re.sub(f"{prae}({arg}){post}", f"\\1{real_t}({args[arg]})\\3", source)
+        source = re.sub(f"{prae}({arg}){post}", f"\\1({real_t})({args[arg]})\\3", source)
     source = re.sub(
         f"{prae}const\\.([^\\d\\W]\\w*]*){post}",
-        "\\1" + real_t + "({const.\\2:" + real_fmt + "})\\3",
+        "\\1(" + real_t + ")({const.\\2:" + real_fmt + "})\\3",
         source
     )
     source = eval(f'f"""{source}"""')
-    return f'{return_type}({source})'
+    return f'({return_type})({source})'
