### 2021-05-24 3a040f7
bisection tol fix; GPU int division problem in Fuchs-Sutugin
