### 2021-05-24 d36e4b3
fixing qv/thd updates on GPU
