### 2021-05-24 889a739
bump examples req; adapt test_vs_BDF to check also GPU impl.
