+++ PySDM/physics/breakup_fragmentations/slams.py
@@ -13,12 +13,14 @@ class SLAMS:
     def __call__(self, output, is_first_in_pair):
         p = 0.0
         r = np.random.rand()       # TODO: optimize random generator outside the fragmentation function
-        output = 1
+        nf = 1
         for i in range(22):
             p += 0.91 * (i + 2)**(-1.56)
             if (r < p):
-                output = i + 2
+                nf = i + 2
                 break
+        output *= 0
+        output += nf
 
     def register(self, builder):
         self.core = builder.core
