### 2021-05-24 d90c04b
README and docstring updates
