+++ PySDM/dynamics/__init__.py
@@ -1,7 +1,3 @@
-"""
-Created at 24.10.2019
-"""
-
 from .coalescence import Coalescence
 from .condensation import Condensation
 from .displacement import Displacement
