+++ PySDM/dynamics/impl/__init__.py
@@ -1,3 +1,0 @@
-"""
-Created at 30.04.2020
-"""
