+++ PySDM/attributes/physics/temperature.py
@@ -1,5 +1,4 @@
 """
-Created at 14.05.2020
 """
 
 from PySDM.attributes.impl.intensive_attribute import IntensiveAttribute
