+++ PySDM/products/__init__.py
@@ -1,7 +1,3 @@
-"""
-Created at 28.08.2020
-"""
-
 from .environments import *
 from .dynamics import *
 from .state import *
