+++ PySDM/core.py
@@ -1,7 +1,3 @@
-"""
-Created at 09.11.2019
-"""
-
 import numpy as np
 
 from PySDM.state.particles import Particles
