+++ PySDM/environments/__init__.py
@@ -1,7 +1,3 @@
-"""
-Created at 30.04.2020
-"""
-
 from .box import Box
 from .kinematic_2d import Kinematic2D
 from .parcel import Parcel
