+++ PySDM/attributes/__init__.py
@@ -1,3 +1,6 @@
 """
-Created at 11.05.2020
+Classes representing particle properties such as
+`PySDM.attributes.physics.multiplicities.Multiplicities`,
+`PySDM.attributes.chemistry.pH.pH`, `PySDM.attributes.physics.volume.Volume`,
+...
 """
