+++ PySDM/attributes/physics/dry_volume.py
@@ -1,5 +1,4 @@
 """
-Created at 11.05.2020
 """
 
 from PySDM.attributes.impl.extensive_attribute import ExtensiveAttribute
