+++ PySDM/initialisation/__init__.py
@@ -1,5 +1,1 @@
-"""
-Crated at 2019
-"""
-
 from .r_wet_init import r_wet_init
