### 2021-05-25 3eecb2c
Add coalescence efficiency and testing
