+++ PySDM/attributes/impl/mapper.py
@@ -1,6 +1,3 @@
-"""
-"""
-
 from PySDM.attributes.physics.multiplicities import Multiplicities
 from PySDM.attributes.physics.volume import Volume
 from PySDM.attributes.physics.dry_volume import DryVolumeDynamic, DryVolumeStatic
