+++ PySDM/attributes/numerics/position_in_cell.py
@@ -1,6 +1,3 @@
-"""
-"""
-
 from PySDM.attributes.impl.cell_attribute import CellAttribute
 
 
