+++ PySDM/physics/coalescence_kernels/_parameterized.py
@@ -1,7 +1,3 @@
-"""
-Created at 05.07.2020
-"""
-
 from ._gravitational import Gravitational
 from PySDM.physics import constants as const
 
