+++ PySDM/backends/numba/impl/_algorithmic_methods.py
@@ -1,7 +1,3 @@
-"""
-Created at 04.11.2019
-"""
-
 import numba
 import numpy as np
 
