+++ PySDM_tests/smoke_tests/Yang_et_al_2018/test_just_do_it.py
@@ -1,6 +1,3 @@
-"""
-Created at 2020
-"""
 import pytest
 import numpy as np
 
