### 2021-05-25 5107437
fixing test_oxidation to work with Numba, other fixes
