+++ PySDM_tests/unit_tests/backends/test_oxidation.py
@@ -1,8 +1,9 @@
 import numpy as np
 from PySDM.backends.numba.impl._chemistry_methods import ChemistryMethods
+from PySDM.backends.numba.storage import Storage
 from PySDM.physics import si, Formulae
 from PySDM.physics.aqueous_chemistry.support import KineticConsts, EquilibriumConsts, \
-    DISSOCIATION_FACTORS
+    DISSOCIATION_FACTORS, k4
 from PySDM.physics.constants import T_STP, pi_4_3
 import pytest
 
@@ -21,22 +22,21 @@ equilibrium_consts = EquilibriumConsts(formulae)
 
 T = T_STP
 
-k0 = np.full(1, kinetic_consts.KINETIC_CONST['k0'].at(T))
-k1 = np.full(1, kinetic_consts.KINETIC_CONST['k1'].at(T))
-k2 = np.full(1, kinetic_consts.KINETIC_CONST['k2'].at(T))
-k3 = np.full(1, kinetic_consts.KINETIC_CONST['k3'].at(T))
-k4 = kinetic_consts.KINETIC_CONST['k4']
-K_SO2 = np.full(1, equilibrium_consts.EQUILIBRIUM_CONST['K_SO2'].at(T))
-K_HSO3 = np.full(1, equilibrium_consts.EQUILIBRIUM_CONST['K_HSO3'].at(T))
+k0 = Storage.from_ndarray(np.full(1, kinetic_consts.KINETIC_CONST['k0'].at(T)))
+k1 = Storage.from_ndarray(np.full(1, kinetic_consts.KINETIC_CONST['k1'].at(T)))
+k2 = Storage.from_ndarray(np.full(1, kinetic_consts.KINETIC_CONST['k2'].at(T)))
+k3 = Storage.from_ndarray(np.full(1, kinetic_consts.KINETIC_CONST['k3'].at(T)))
+K_SO2 = Storage.from_ndarray(np.full(1, equilibrium_consts.EQUILIBRIUM_CONST['K_SO2'].at(T)))
+K_HSO3 = Storage.from_ndarray(np.full(1, equilibrium_consts.EQUILIBRIUM_CONST['K_HSO3'].at(T)))
 
 volume = pi_4_3 * (1 * si.um)**3
 pH = 5.
 n_sd = 1
 eqc = {
-    k: np.full(n_sd, equilibrium_consts.EQUILIBRIUM_CONST[k].at(T))
+    k: Storage.from_ndarray(np.full(n_sd, equilibrium_consts.EQUILIBRIUM_CONST[k].at(T)))
     for k in ('K_HSO3', 'K_SO2')
 }
-cell_ids = np.zeros(n_sd, dtype=int)
+cell_ids = Storage.from_ndarray(np.zeros(n_sd, dtype=int))
 H = formulae.trivia.pH2H(pH)
 DF = DISSOCIATION_FACTORS['SO2'](H, eqc, 0)
 
@@ -57,7 +57,7 @@ def test_oxidation(conc):
     dt = 1
 
     moles = {
-        k: np.full(n_sd, 0. if k not in conc['input'] else conc['input'][k] * volume)
+        k: Storage.from_ndarray(np.full(n_sd, 0. if k not in conc['input'] else conc['input'][k] * volume))
         for k in ('S_IV', 'S_VI', 'H2O2', 'O3')
     }
 
@@ -65,7 +65,7 @@ def test_oxidation(conc):
     sut.oxidation(
         n_sd=n_sd,
         cell_ids=cell_ids,
-        do_chemistry_flag=np.full(n_sd, True),
+        do_chemistry_flag=Storage.from_ndarray(np.full(n_sd, True)),
         k0=k0,
         k1=k1,
         k2=k2,
@@ -74,9 +74,9 @@ def test_oxidation(conc):
[…]
