### 2021-05-26 5d7c43b
parcel example narrative in README [skip ci]
