+++ PySDM/backends/thrustRTC/impl/_pair_methods.py
@@ -24,7 +24,7 @@ class PairMethods:
 
     __find_pairs_body = trtc.For(['cell_start', 'perm_cell_id', 'is_first_in_pair', 'length'], "i", '''
         is_first_in_pair[i] = (
-            i < length - 1 &&
+            i < length - 1 && // note: just to set the last element within the same loop
             perm_cell_id[i] == perm_cell_id[i+1] &&
             (i - cell_start[perm_cell_id[i]]) % 2 == 0
         );
@@ -34,7 +34,7 @@ class PairMethods:
     @nice_thrust(**NICE_THRUST_FLAGS)
     def find_pairs(cell_start, is_first_in_pair, cell_id, cell_idx, idx):  # TODO #330 handle cell_idx
         perm_cell_id = trtc.DVPermutation(cell_id.data, idx.data)
-        d_length = trtc.DVInt64(len(idx))  # TODO #350 length-1 as we use i+1 in __find_pairs_body
+        d_length = trtc.DVInt64(len(idx))
         PairMethods.__find_pairs_body.launch_n(
             len(idx), [cell_start.data, perm_cell_id, is_first_in_pair.indicator.data, d_length])
 
