### 2021-05-26 b746ecb
using backends fixture in Sh&H test_few_steps
