+++ PySDM_tests/smoke_tests/Shipway_and_Hill_2012/test_few_steps.py
@@ -1,12 +1,16 @@
-from PySDM_examples.Shipway_and_Hill_2012 import Simulation, Settings
-from PySDM.physics import si
 import numpy as np
 
+from PySDM_examples.Shipway_and_Hill_2012 import Simulation, Settings
+from PySDM.physics import si
 
-def test_few_steps(plot=False):
+# noinspection PyUnresolvedReferences
+from PySDM_tests.backends_fixture import backend
+
+
+def test_few_steps(backend, plot=False):
     # Arrange
     settings = Settings(n_sd_per_gridbox=50, dt=30*si.s, dz=50*si.m)
-    simulation = Simulation(settings)
+    simulation = Simulation(settings, backend=backend)
 
     # Act
     output = simulation.run(nt=100)
@@ -29,4 +33,4 @@ def test_few_steps(plot=False):
     assert .1 < max(profile('ql')) < 1
     # assert max(profile('ripening_rate')) > 0 # TODO #521
     assert max(profile('activating_rate')) == 0
-    #assert max(profile('deactivating_rate')) > 0 TODO #521
+    # assert max(profile('deactivating_rate')) > 0 TODO #521
