### 2021-05-26 e3f4c29
refactored GPU condensation
rollback test_few_steps.py
