### 2021-05-26 ad6c2a3
RH_max product works with GPU
