### 2021-05-26 648137c
fixed pair_methods
