+++ PySDM/backends/thrustRTC/impl/_pair_methods.py
@@ -7,10 +7,7 @@ class PairMethods:
 
     __distance_pair_body = trtc.For(['data_out', 'data_in', 'is_first_in_pair'], "i", '''
         if (is_first_in_pair[i]) {
-            data_out[i] = abs(data_in[i] - data_in[i + 1]);
-        } 
-        else {
-            data_out[i] = 0;
+            data_out[(int64_t)(i/2)] = abs(data_in[i] - data_in[i + 1]);
         }
         ''')
 
@@ -62,9 +59,6 @@ class PairMethods:
                 data_out[i] = data_in[i];
                 data_out[i + 1] = data_in[i + 1];
             }
-        } 
-        else {
-            data_out[i] = 0;
         }
         ''')
 
