### 2021-05-26 9cc8e6c
bump examples req; fix test_few_steps
