### 2021-05-26 fd98c80
rework parcel example in README; cleanups
