### 2021-05-26 7ec3fcf
Matlab code cleanup
