+++ PySDM/initialisation/__init__.py
@@ -5,3 +5,7 @@ Initialisation logic incl.
 `PySDM.initialisation.spectra`
 """
 from .r_wet_init import r_wet_init
+from . import spectral_sampling
+from . import spectra
+from . import spatial_sampling
+from . import multiplicities
