### 2021-05-26 e1c9f47
run README code on GPU as well
