### 2021-05-26 befbb85
sed usage fixes; Julia global var syntax fix
