### 2021-05-26 056174e
Fix GPU issue: initialize neg_ones vec
