+++ PySDM/dynamics/breakup.py
@@ -69,7 +69,8 @@ class Breakup:
         self.kernel_temp = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=float)
         self.fragmentation_temp = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=int)
         self.coal_eff_temp = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=float)
-        self.neg_ones = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=int)
+        neg_ones_tmp = np.tile([-1], self.core.n_sd // 2)
+        self.neg_ones = self.core.PairwiseStorage.from_ndarray(neg_ones_tmp)
         self.norm_factor_temp = self.core.Storage.empty(self.core.mesh.n_cell, dtype=float)
         self.prob = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=float)
         self.n_fragment = self.core.PairwiseStorage.empty(self.core.n_sd // 2, dtype=int)
