+++ PySDM/backends/thrustRTC/impl/_algorithmic_methods.py
@@ -190,9 +190,9 @@ class AlgorithmicMethods:
     def breakup(n, idx, length, attributes, gamma, n_fragment, healthy, is_first_in_pair):
         n_sd = trtc.DVInt64(attributes.shape[1])
         n_attr = trtc.DVInt64(attributes.shape[0])
-        print(n.data.to_host(), flush=True)
-        print(gamma.data.to_host(), flush=True)
-        print(n_fragment.data.to_host(), flush=True)
+        #print(n.data.to_host(), flush=True)
+        #print(gamma.data.to_host(), flush=True)
+        #print(n_fragment.data.to_host(), flush=True)
         AlgorithmicMethods.__breakup_body.launch_n(length // 2,
                                                    [n.data, idx.data, n_sd,
                                                     attributes.data, n_attr,
@@ -223,12 +223,12 @@ class AlgorithmicMethods:
     @nice_thrust(**NICE_THRUST_FLAGS)
     def compute_gamma(gamma, rand, n, cell_id,
                       collision_rate_deficit, collision_rate, is_first_in_pair):
-        print("compute_gamma ", gamma.data.to_host())
+        #print("compute_gamma ", gamma.data.to_host())
         AlgorithmicMethods.__compute_gamma_body.launch_n(
             len(n) // 2,
             [gamma.data, rand.data, n.idx.data, n.data, cell_id.data,
              collision_rate_deficit.data, collision_rate.data])
-        print("computed_gamma ", gamma.data.to_host())
+        #print("computed_gamma ", gamma.data.to_host())
 
     @staticmethod
     @nice_thrust(**NICE_THRUST_FLAGS)
