### 2021-05-27 a3faf3b
c_inline refactors
