+++ PySDM/backends/thrustRTC/impl/condensation_methods.py
@@ -2,7 +2,7 @@ from typing import Dict, Optional
 
 from PySDM.physics import constants as const
 from PySDM.backends.thrustRTC.conf import NICE_THRUST_FLAGS
-from PySDM.backends.thrustRTC.impl import nice_thrust, c_inline
+from PySDM.backends.thrustRTC.impl import nice_thrust
 from .precision_resolver import PrecisionResolver
 from PySDM.backends.thrustRTC.bisection import BISECTION
 from ..conf import trtc
@@ -40,12 +40,12 @@ class CondensationMethods():
             struct Minfun {{
                 static __device__ real_type value(real_type x_new, void* args_p) {{
                     auto args = static_cast<real_type*>(args_p);
-                    auto vol = {c_inline(phys.condensation_coordinate.volume, x="x_new")};
-                    auto r_new = {c_inline(phys.trivia.radius, volume="vol")};
-                    auto sgm = {c_inline(phys.surface_tension.sigma, T=args('_T'), v_wet="vol", v_dry=f"const.pi_4_3 * {args('rd3')}")};
-                    auto RH_eq = {c_inline(phys.hygroscopicity.RH_eq, r="r_new", T=args('_T'), kp=args("kappa"), rd3=args("rd3"), sgm="sgm")};
-                    auto r_dr_dt = {c_inline(phys.drop_growth.r_dr_dt, RH_eq="RH_eq", T=args("_T"), RH=args("_RH"), lv=args("_lv"), pvs=args("_pvs"), D=args("Dr"), K=args("Kr"))};
-                    return {args("x_old")} - x_new + {args("dt")} * {c_inline(phys.condensation_coordinate.dx_dt, x="x_new", r_dr_dt="r_dr_dt")};
+                    auto vol = {phys.condensation_coordinate.volume.c_inline(x="x_new")};
+                    auto r_new = {phys.trivia.radius.c_inline(volume="vol")};
+                    auto sgm = {phys.surface_tension.sigma.c_inline(T=args('_T'), v_wet="vol", v_dry=f"const.pi_4_3 * {args('rd3')}")};
+                    auto RH_eq = {phys.hygroscopicity.RH_eq.c_inline(r="r_new", T=args('_T'), kp=args("kappa"), rd3=args("rd3"), sgm="sgm")};
+                    auto r_dr_dt = {phys.drop_growth.r_dr_dt.c_inline(RH_eq="RH_eq", T=args("_T"), RH=args("_RH"), lv=args("_lv"), pvs=args("_pvs"), D=args("Dr"), K=args("Kr"))};
+                    return {args("x_old")} - x_new + {args("dt")} * {phys.condensation_coordinate.dx_dt.c_inline(x="x_new", r_dr_dt="r_dr_dt")};
                 }}
             }};
             {BISECTION}
@@ -59,12 +59,12 @@ class CondensationMethods():
             auto _lambdaK = lambdaK[cell_id[i]];
             auto _lambdaD = lambdaD[cell_id[i]];
             
-            auto x_old = {c_inline(phys.condensation_coordinate.x, volume="v[i]")};
-            auto r_old = {c_inline(phys.trivia.radius, volume="v[i]")};
-            auto x_insane = {c_inline(phys.condensation_coordinate.x, volume="vdry[i]/100")};
+            auto x_old = {phys.condensation_coordinate.x.c_inline(volume="v[i]")};
[…]
