+++ PySDM/backends/thrustRTC/impl/__init__.py
@@ -1,2 +1,1 @@
 from .nice_thrust import nice_thrust
-from .c_inline import c_inline
