### 2021-05-27 3ca8471
skipping one unneeded run in Y test_just_do_it; uncommenting one skipped test in test_advection
