### 2021-05-27 ae1ec28
Fix SLAMS kernel for GPU
