+++ PySDM/products/state/__init__.py
@@ -13,3 +13,4 @@ from .particles_volume_spectrum import ParticlesVolumeSpectrum
 from .water_mixing_ratio import WaterMixingRatio
 from .cloud_droplet_effective_radius import CloudDropletEffectiveRadius
 from .total_dry_mass_mixing_ratio import TotalDryMassMixingRatio
+from .k_moments import KMoments
