+++ PySDM/physics/breakup_fragmentations/slams.py
@@ -9,6 +9,8 @@ class SLAMS:
 
     def __init__(self):
         self.core = None
+        self.N_vec = None
+        self.zeros = None
 
     def __call__(self, output, is_first_in_pair):
         p = 0.0
@@ -19,8 +21,14 @@ class SLAMS:
             if (r < p):
                 nf = i + 2
                 break
-        output *= 0
-        output += nf
+        output *= self.zeros
+        output += self.N_vec
+        output *= nf
+        print(nf, flush=True)
 
     def register(self, builder):
         self.core = builder.core
+        N_vec_tmp = np.tile([1], self.core.n_sd // 2)
+        zeros_tmp = np.tile([0], self.core.n_sd // 2)
+        self.N_vec = self.core.PairwiseStorage.from_ndarray(N_vec_tmp)
+        self.zeros = self.core.PairwiseStorage.from_ndarray(zeros_tmp)
