+++ PySDM/backends/numba/impl/_algorithmic_methods.py
@@ -160,13 +160,13 @@ class AlgorithmicMethods:
                 n[k] = n[k] * int(n_fragment[i])
                 for a in range(0, len(attributes)):
                     attributes[a, k] += gamma[i] * attributes[a, j]
-                    attributes[a, k] /= int(n_fragment[i])
+                    attributes[a, k] /= n_fragment[i]
                     
             else:  # new_n == 0
-                n[j] = (int(n_fragment[i]) * n[k]) // 2
-                n[k] = int(n_fragment[i]) * n[k] - n[j]
+                n[j] = (n_fragment[i] * n[k]) // 2
+                n[k] = n_fragment[i] * n[k] - n[j]
                 for a in range(0, len(attributes)):
-                    attributes[a, j] = (gamma[i] * attributes[a, j] + attributes[a, k])/int(n_fragment[i])
+                    attributes[a, j] = (gamma[i] * attributes[a, j] + attributes[a, k])/n_fragment[i]
                     attributes[a, k] = attributes[a, j]
             if n[k] == 0 or n[j] == 0:
                 healthy[0] = 0
