### 2021-05-29 39246ce
better chem spec test. closes #481
