### 2021-05-30 07db830
more Thrust storage asserts, fixing array slicing in Thrust backend tests
