+++ PySDM/backends/thrustRTC/storage.py
@@ -29,8 +29,14 @@ class Storage:
             else:
                 raise NotImplementedError("Only 2 or less dimensions array is supported.")
             result = Storage(result_data, result_shape, self.dtype)
-        elif isinstance(item, tuple) and dim == 2 and isinstance(item[1], slice):
-            result_data = self.data.range(self.shape[1] * item[0], self.shape[1] * (item[0] + 1))
+        elif isinstance(item, tuple) and dim == 2 and isinstance(item[0], int):
+            if isinstance(item[1], slice):
+                assert item[1].start is None or item[1].start == 0
+                assert item[1].stop is None or item[1].stop == self.shape[1]
+                assert item[1].step is None or item[1].step == 1
+                result_data = self.data.range(self.shape[1] * item[0], self.shape[1] * (item[0] + 1))
+            else:
+                raise NotImplementedError()
             result = Storage(result_data, (*self.shape[1:],), self.dtype)
         else:
             result = self.to_ndarray()[item]
