+++ PySDM_tests/smoke_tests/Kreidenweis_et_al_2003/test_Fig_1.py
@@ -68,4 +68,5 @@ class TestFig1:
             pyplot.xlabel('pH')
             pyplot.show()
 
-        assert 4.9 < example_output['pH'][-1] < 5
+        assert 4.9 < example_output['pH_number_weighted'][-1] < 5
+        assert 4.9 < example_output['pH_volume_weighted'][-1] < 5
