### 2021-05-31 a0afff2
bringing back unintentionally lost code fragment! (ThrustRTC __linear_collection_efficiency_body)
