+++ PySDM_tests/smoke_tests/Kreidenweis_et_al_2003/test_Fig_1.py
@@ -23,9 +23,7 @@ class TestFig1:
         # Plot
         if plot:
             name = 'ql'
-            #prod = simulation.core.products['ql']
             pyplot.plot(example_output[name], np.asarray(example_output['t']) - Z_CB * si.s)
-            #pyplot.xlabel(f"{prod.name} [{prod.unit}]")  # TODO #442
             pyplot.ylabel(f"time above cloud base [s]")
             pyplot.grid()
             pyplot.show()
