### 2021-06-01 33bdb64
cleanup, different defaults for pH ctor args
