### 2021-06-02 450bfdc
kind of fixing ionic_strength test (pH was infinite for one or two super droplets, OK for more)
