+++ PySDM_tests/unit_tests/attributes/chemistry/test_pH.py
@@ -15,6 +15,7 @@ from collections import defaultdict
 formulae = Formulae()
 EQUILIBRIUM_CONST = EquilibriumConsts(formulae).EQUILIBRIUM_CONST
 
+
 class Test_pH:
     @staticmethod
     def test_equilibrate_pH_pure_water():
@@ -117,73 +118,3 @@ class Test_pH:
         )
 
         np.testing.assert_allclose(actual_pH[0], expected_pH, rtol=1e-5)
-
-    @staticmethod
-    @pytest.mark.parametrize("nt", (0, 1, 2, 3))
-    @pytest.mark.parametrize("n_sd", (1, 2, 10))
-    def test_calc_ionic_strength(nt, n_sd):
-        from chempy.electrolytes import ionic_strength
-        from PySDM_examples.Kreidenweis_et_al_2003 import Settings, Simulation
-        from PySDM.backends.numba.impl._chemistry_methods import calc_ionic_strength
-        from PySDM.physics.constants import rho_w, ROOM_TEMP
-
-        K_NH3 = EQUILIBRIUM_CONST["K_NH3"].at(ROOM_TEMP)
-        K_SO2 = EQUILIBRIUM_CONST["K_SO2"].at(ROOM_TEMP)
-        K_HSO3 = EQUILIBRIUM_CONST["K_HSO3"].at(ROOM_TEMP)
-        K_HSO4 = EQUILIBRIUM_CONST["K_HSO4"].at(ROOM_TEMP)
-        K_HCO3 = EQUILIBRIUM_CONST["K_HCO3"].at(ROOM_TEMP)
-        K_CO2 = EQUILIBRIUM_CONST["K_CO2"].at(ROOM_TEMP)
-        K_HNO3 = EQUILIBRIUM_CONST["K_HNO3"].at(ROOM_TEMP)
-
-        settings = Settings(dt=1, n_sd=n_sd, n_substep=5)
-        settings.t_max = nt * settings.dt
-        simulation = Simulation(settings)
-        simulation.run()
-
-        H = simulation.core.particles['conc_N_mIII'].data
-        conc = {
-            'H+': H,
-            'N-3': simulation.core.particles['conc_N_mIII'].data,
-            'N+5': simulation.core.particles['conc_N_V'].data,
-            'S+4': simulation.core.particles['conc_S_IV'].data,
-            'S+6': simulation.core.particles['conc_S_VI'].data,
-            'C+4': simulation.core.particles['conc_C_IV'].data,
-        }
-
-        alpha_C = (1 + K_CO2 / conc['H+'] + K_CO2 * K_HCO3 / conc['H+'] ** 2)
-        alpha_S = (1 + K_SO2 / conc['H+'] + K_SO2 * K_HSO3 / conc['H+'] ** 2)
-        alpha_N3 = (1 + conc['H+'] * K_NH3 / K_H2O)
-        alpha_N5 = (1 + K_HNO3 / conc['H+'])
-
-        actual = calc_ionic_strength(
-            H=conc['H+'],
-            N_mIII=conc['N-3'],
-            N_V=conc['N+5'],
-            C_IV=conc['C+4'],
-            S_IV=conc['S+4'],
-            S_VI=conc['S+6'],
-            K_NH3=K_NH3,
-            K_SO2=K_SO2,
-            K_HSO3=K_HSO3,
-            K_HSO4=K_HSO4,
-            K_HCO3=K_HCO3,
-            K_CO2=K_CO2,
-            K_HNO3=K_HNO3
-        )
-
-        expected = ionic_strength({
[…]
