### 2021-06-02 dc0a257
fixed getitem and test_advection.py
