### 2021-06-02 78f9bd8
"fixed" "An internal error happened" i.e. size==0 & added rainfall to flag_precipitated (GPU)
