### 2021-06-02 374bf09
Fix GPU backend for SLAMS
