### 2021-06-02 f4de34b
Fix power and typecast for ThrustRTC
