### 2021-06-02 62645b2
Fix final typecast errors with GPU backend
