### 2021-06-02 88ef8e3
removing two more unneeded test skips
