+++ PySDM_tests/unit_tests/dynamics/displacement/test_advection.py
@@ -23,10 +23,6 @@ class TestExplicitEulerWithInterpolation:
 
     @staticmethod
     def test_advection(backend):
-        from PySDM.backends import ThrustRTC
-        if backend is ThrustRTC:
-            return  # TODO #332
-
         # Arrange
         settings = DisplacementSettings()
         settings.grid = (3, 3)
@@ -111,10 +107,6 @@ class TestExplicitEulerWithInterpolation:
 
     @staticmethod
     def test_update_cell_origin(backend):
-        from PySDM.backends import ThrustRTC
-        if backend is ThrustRTC:
-            return  # TODO #332
-
         # Arrange
         settings = DisplacementSettings()
         sut, core = settings.get_displacement(backend, scheme='ImplicitInSpace')
