### 2021-06-03 4d4d302
Numba backend cell_id_body simplification (strides are always n-dim now)
