+++ PySDM/backends/numba/impl/_algorithmic_methods.py
@@ -104,10 +104,7 @@ class AlgorithmicMethods:
     @staticmethod
     # @numba.njit(**conf.JIT_FLAGS)  # Note: in Numba 0.51 "np.dot() only supported on float and complex arrays"
     def cell_id_body(cell_id, cell_origin, strides):
-        if strides.size == 1:
-            cell_id[:] = cell_origin
-        else:
-            cell_id[:] = np.dot(strides, cell_origin)
+        cell_id[:] = np.dot(strides, cell_origin)
 
     @staticmethod
     def cell_id(cell_id, cell_origin, strides):
