### 2021-06-04 4fbc30f
moving spectra to physics. Closes #550
