+++ PySDM/initialisation/__init__.py
@@ -6,6 +6,5 @@ Initialisation logic incl.
 """
 from .r_wet_init import r_wet_init
 from . import spectral_sampling
-from . import spectra
 from . import spatial_sampling
 from . import multiplicities
