### 2021-06-08 e08d1b7
added smoke test for Berry_1967 example
closing issue #549
