### 2021-06-08 87e2d4a
added test_coalescence_2_sd
handling len(idx) < 1 in GPU backend
removing length from arguments (coalescence, sort_within_pair_by_attr)
