### 2021-06-08 61c106d
fixed MomentsMethods.__moments_body_0 and test_few_steps for GPU
