### 2021-06-08 1713fda
C cast syntax compatible with FakeThrust...
