+++ PySDM/backends/thrustRTC/impl/moments_methods.py
@@ -16,7 +16,7 @@ class MomentsMethods:
             '''
             auto i = idx[fake_i];
             if (min_x < x_attr[i] && x_attr[i] < max_x) {
-                atomicAdd((real_type*)&moment_0[cell_id[i]], (real_type)n[i]);
+                atomicAdd((real_type*)&moment_0[cell_id[i]], (real_type)(n[i]));
                 for (auto k = 0; k < n_ranks; k+=1) {
                     atomicAdd((real_type*) &moments[n_cell * k + cell_id[i]], n[i] * pow((real_type)(attr_data[i]), (real_type)(ranks[k])));
                }
