### 2021-06-08 0a911a4
mark test_coalescence_2_sd() as xfail for 32 bit Python
