+++ PySDM_tests/smoke_tests/Berry_1967/test_coalescence.py
@@ -46,7 +46,7 @@ def test_coalescence(backend, kernel, croupier, adaptive):
         x_max = np.amax(volume)
 
 
-@pytest.mark.xfail(struct.calcsize("P") * 8 == 32, strict=True, reason="32 bit")
+@pytest.mark.xfail(struct.calcsize("P") * 8 == 32, reason="32 bit")  # TODO #565
 def test_coalescence_2_sd(backend):
     # Arrange
     s = Settings()
