### 2021-06-15 ba04bed
fixed test_moments.py for GPU backend
