### 2021-06-15 2b3aa6f
change sgm in constants file to sgm_w for water and then add content to placeholder for calculating surface tension when aerosol has some organic fraction via the compressed film framework. still TODO is adding f_org as a superdroplet attribute and creating some structure for carrying around the chemical parameters relevant for the organic species
