+++ PySDM/physics/surface_tension/constant.py
@@ -4,4 +4,4 @@ import PySDM.physics.constants as const
 class Constant:
     @staticmethod
     def sigma(T, v_wet, v_dry):
-        return const.sgm
+        return const.sgm_w
