+++ PySDM/physics/constants.py
@@ -67,7 +67,7 @@ FWC_C7 = 0.379534310e-13 * si.hPa / si.K**7
 FWC_C8 = -.321582393e-15 * si.hPa / si.K**8
 
 rho_w = 1 * si.kilograms / si.litres
-sgm = 0.072 * si.joule / si.metre ** 2  # TODO #223 temperature dependence
+sgm_w = 0.072 * si.joule / si.metre ** 2  # TODO #223 temperature dependence
 
 p_tri = 611.73 * si.pascal
 T_tri = 273.16 * si.kelvin
