### 2021-06-16 1e8a74d
change sgm to sgm_w in tests
