### 2021-06-17 537e00a
first attempt to add organic and inorganic dry volumes as separate attributes
