+++ PySDM/attributes/physics/dry_volume.py
@@ -2,11 +2,27 @@ from PySDM.attributes.impl.extensive_attribute import ExtensiveAttribute
 from PySDM.attributes.impl.derived_attribute import DerivedAttribute
 from PySDM.physics import constants as const
 
-
-class DryVolumeStatic(ExtensiveAttribute):
+#TODO 223
+# first attempt to add inorganic and organic dry volumes as extensive attributes
+# and make DryVolumeStatic a derived attribute
+class DryVolumeInorganic(ExtensiveAttribute):
     def __init__(self, builder):
-        super().__init__(builder, name='dry volume')
+        super().__init__(builder, name='dry volume inorganic')
 
+class DryVolumeOrganic(ExtensiveAttribute):
+    def __init__(self, builder):
+        super().__init__(builder, name='dry volume organic')
+
+class DryVolumeStatic(DerivedAttribute):
+    def __init__(self, builder):
+        self.volume_dry_org = builder.get_attribute('dry volume organic')
+        self.volume_dry_inorg = builder.get_attribute('dry volume inorganic')
+        dependencies = [self.volume_dry_org, self.volume_dry_inorg]
+        super().__init__(builder, name='dry volume', dependencies=dependencies)
+
+    def recalculate(self):
+        self.data.idx = self.volume_dry_org.data.idx
+        self.data.sum(self.volume_dry_org.get(), self.volume_dry_inorg.get())
 
 class DryVolumeDynamic(DerivedAttribute):
     def __init__(self, builder):
