+++ PySDM/attributes/impl/mapper.py
@@ -1,6 +1,7 @@
 from PySDM.attributes.physics.multiplicities import Multiplicities
 from PySDM.attributes.physics.volume import Volume
-from PySDM.attributes.physics.dry_volume import DryVolumeOrganic, DryVolumeInorganic, DryVolumeDynamic, DryVolumeStatic
+from PySDM.attributes.physics.dry_volume import DryVolumeOrganic, DryVolumeInorganic, 
+from PySDM.attributes.physics.dry_volume import DryVolumeOrgInorg, DryVolumeDynamic, DryVolumeStatic
 from PySDM.attributes.physics.radius import Radius
 from PySDM.attributes.physics.dry_radius import DryRadius
 from PySDM.attributes.physics.terminal_velocity import TerminalVelocity
@@ -20,9 +21,9 @@ from functools import partial
 attributes = {
     'n': lambda _: Multiplicities,
     'volume': lambda _: Volume,
-    'dry volume organic': lambda _: DryVolumeOrganic, #TODO 223
-    'dry volume inorganic': lambda _: DryVolumeInorganic, #TODO 223
-    'dry volume': lambda dynamics: DryVolumeDynamic if 'AqueousChemistry' in dynamics else DryVolumeStatic,
+    'dry volume organic': lambda _: DryVolumeOrganic,
+    'dry volume inorganic': lambda _: DryVolumeInorganic,
+    'dry volume': lambda dynamics: DryVolumeDynamic if 'AqueousChemistry' in dynamics (DryVolumeOrgInorg if 'SurfaceOrganics' in dynamics else DryVolumeStatic),
     'radius': lambda _: Radius,
     'dry radius': lambda _: DryRadius,
     'terminal velocity': lambda _: TerminalVelocity,
