+++ PySDM/dynamics/__init__.py
@@ -9,3 +9,4 @@ from .displacement import Displacement
 from .eulerian_advection import EulerianAdvection
 from .ambient_thermodynamics import AmbientThermodynamics
 from .aqueous_chemistry import AqueousChemistry
+from .surface_organics import SurfaceOrganics
