+++ PySDM/attributes/physics/dry_volume.py
@@ -17,7 +17,7 @@ class DryVolumeDynamic(DerivedAttribute):
         self.data.data[:] = self.moles_S_VI.data.data[:]
         self.data.data[:] *= dynamic.dry_molar_mass / dynamic.dry_rho
 
-#TODO 223
+# TODO #233
 # first attempt to add inorganic and organic dry volumes as extensive attributes
 # and make DryVolumeStatic a derived attribute
 class DryVolumeInorganic(ExtensiveAttribute):
