### 2021-06-20 atmos-cloud-sim-uj/PySDM#494 @slayoo
@Golui has started exploring what needs to be done here.
First, it would be great to have ThrustRTC and CURandRTC available for conda, relevant requests reported:
- https://github.com/fynv/ThrustRTC/issues/13
- https://github.com/fynv/CURandRTC/issues/3
