### 2021-06-21 8549d3b
draft of atomic_min in adaptive_sdm_gamma()
atomic_min does not work with floating points, so something like that is necessary.
It's hacky, but on the other hand it should work since high precision is not crucial here.

addressed in #571
