### 2021-06-22 62aae8f
hardcoded parameters in CompressedFilm surface tensions; temporarily disabling organic/inorganic attributes in the mapper; draft of smoke test for Lowe_et_al_2019
