+++ PySDM/attributes/impl/mapper.py
@@ -23,7 +23,8 @@ attributes = {
     'volume': lambda _: Volume,
     'dry volume organic': lambda _: DryVolumeOrganic,
     'dry volume inorganic': lambda _: DryVolumeInorganic,
-    'dry volume': lambda dynamics: DryVolumeDynamic if 'AqueousChemistry' in dynamics else (DryVolumeOrgInorg if 'SurfaceOrganics' in dynamics else DryVolumeStatic),
+    # 'dry volume': lambda dynamics: DryVolumeDynamic if 'AqueousChemistry' in dynamics else (DryVolumeOrgInorg if 'SurfaceOrganics' in dynamics else DryVolumeStatic),
+    'dry volume': lambda dynamics: DryVolumeDynamic if 'AqueousChemistry' in dynamics else DryVolumeStatic,
     'radius': lambda _: Radius,
     'dry radius': lambda _: DryRadius,
     'terminal velocity': lambda _: TerminalVelocity,
