+++ PySDM/physics/surface_tension/compressed_film.py
@@ -1,17 +1,14 @@
-import PySDM.physics.constants as const
+from PySDM.physics import constants as const#, si
 
 class CompressedFilm:
     @staticmethod
     def sigma(T, v_wet, v_dry):
-        # TODO #223
-        # placeholder for chem structure and f_org
-        f_org = 0.5
-        chem = {"delta_min":1, "sgm_org":const.sgm_w*2}
-
+        # TODO #223 - these parameters should be passed from Settings
         # chemical parameters
-        delta_min = chem.delta_min  # minimum organic film thickness [nm]
-        sgm_org = chem.sgm_org      # organic surface tension [N m-1]
-    
+        delta_min = 0.2e-9  # minimum organic film thickness [nm]
+        sgm_org = const.sgm_w*0.5      # organic surface tension [N m-1]
+        f_org = .5
+
         # convert volumes to diameters
         d_coat = 2 * ((3*v_dry) / (4*const.pi))**(1/3)
         d_seed = (d_coat**3 - f_org * d_coat**3)**(1/3)
@@ -28,5 +25,4 @@ class CompressedFilm:
         
         # calculate sigma
         sgm = (1-c_beta)*const.sgm_w + c_beta*sgm_org
-        
-        return sgm 
+        return sgm
