+++ PySDM/physics/surface_tension/__init__.py
@@ -1,1 +1,2 @@
 from .constant import Constant
+from .compressed_film import CompressedFilm
