### 2021-06-22 10e9ff4
wip add organic fraction as derived attribute
