### 2021-06-23 atmos-cloud-sim-uj/PySDM#494 @slayoo
Both ThrustRTC and CURandRTC are now available as conda packages! Thanks @fynv!
