+++ PySDM/products/state/particles_size_spectrum.py
@@ -1,14 +1,16 @@
 import numpy as np
 
 from PySDM.physics import constants as const
-from PySDM.products.product import MomentProduct
 
 
-class ParticlesSizeSpectrum(MomentProduct):
+from PySDM.products.product import SpectrumMomentProduct
 
-    def __init__(self, v_bins, name, dry=False, normalise_by_dv=False):
+
+class ParticlesSizeSpectrum(SpectrumMomentProduct):
+
+    def __init__(self, radius_bins_edges, name, dry=False, normalise_by_dv=False):
         self.volume_attr = 'dry volume' if dry else 'volume'
-        self.v_bins = v_bins
+        self.radius_bins_edges = radius_bins_edges
         self.normalise_by_dv = normalise_by_dv
         super().__init__(
             name=name,
@@ -18,15 +20,20 @@ class ParticlesSizeSpectrum(MomentProduct):
 
     def register(self, builder):
         builder.request_attribute(self.volume_attr)
+
+        volume_bins_edges = builder.core.formulae.trivia.volume(self.radius_bins_edges)
+        self.attr_bins_edges = builder.core.bck.Storage.from_ndarray(volume_bins_edges)
+
         super().register(builder)
-        self.shape = (*builder.core.mesh.grid, len(self.v_bins) - 1)
+
+        self.shape = (*builder.core.mesh.grid, len(self.attr_bins_edges) - 1)
 
     def get(self):
-        vals = np.empty([self.core.mesh.n_cell, len(self.v_bins) - 1])
-        for i in range(len(self.v_bins) - 1):
-            self.download_moment_to_buffer(
-                attr=self.volume_attr, rank=0, filter_attr=self.volume_attr, filter_range=(self.v_bins[i], self.v_bins[i + 1])
-            )
+        vals = np.empty([self.core.mesh.n_cell, len(self.attr_bins_edges) - 1])
+        self.recalculate_spectrum_moment(attr=self.volume_attr, rank=1, filter_attr=self.volume_attr)
+
+        for i in range(vals.shape[1]):
+            self.download_spectrum_moment_to_buffer(rank=0, bin_number=i)
             vals[:, i] = self.buffer.ravel()
 
         if self.normalise_by_dv:
@@ -34,9 +41,9 @@ class ParticlesSizeSpectrum(MomentProduct):
 
         self.download_to_buffer(self.core.environment['rhod'])
         rhod = self.buffer.ravel()
-        for i in range(len(self.v_bins) - 1):
-            dr = self.formulae.trivia.radius(volume=self.v_bins[i + 1]) - \
-                 self.formulae.trivia.radius(volume=self.v_bins[i])
+        for i in range(len(self.attr_bins_edges) - 1):
+            dr = self.formulae.trivia.radius(volume=self.attr_bins_edges[i + 1]) - \
+                 self.formulae.trivia.radius(volume=self.attr_bins_edges[i])
             vals[:, i] /= rhod * dr
 
         const.convert_to(vals, const.si.micrometre**-1 * const.si.milligram**-1)
@@ -45,10 +52,10 @@ class ParticlesSizeSpectrum(MomentProduct):
 
 
 class ParticlesWetSizeSpectrum(ParticlesSizeSpectrum):
-    def __init__(self, v_bins, normalise_by_dv=False):
-        super().__init__(v_bins, dry=False, normalise_by_dv=normalise_by_dv, name='Particles Wet Size Spectrum')
+    def __init__(self, radius_bins_edges, normalise_by_dv=False):
+        super().__init__(radius_bins_edges, dry=False, normalise_by_dv=normalise_by_dv, name='Particles Wet Size Spectrum')
 
 
[…]
