+++ README.md
@@ -240,6 +240,7 @@ particles = builder.build(attributes, products);
 <summary>Python</summary>
 
 ```Python
+import numpy as np
 from PySDM import Builder
 from PySDM.environments import Box
 from PySDM.dynamics import Coalescence
@@ -250,7 +251,8 @@ from PySDM.products.state import ParticlesVolumeSpectrum
 builder = Builder(n_sd=n_sd, backend=CPU)
 builder.set_environment(Box(dt=1 * si.s, dv=1e6 * si.m**3))
 builder.add_dynamic(Coalescence(kernel=Golovin(b=1.5e3 / si.s)))
-products = [ParticlesVolumeSpectrum()]
+radius_bins_edges = np.logspace(np.log10(10 * si.um), np.log10(5e3 * si.um), num=32)
+products = [ParticlesVolumeSpectrum(radius_bins_edges)]
 particles = builder.build(attributes, products)
 ```
 </details>
@@ -329,14 +331,11 @@ legend()
 ```Python
 from PySDM.physics.constants import rho_w
 from matplotlib import pyplot
-import numpy as np
-
-radius_bins_edges = np.logspace(np.log10(10 * si.um), np.log10(5e3 * si.um), num=32)
 
 for step in [0, 1200, 2400, 3600]:
     particles.run(step - particles.n_steps)
     pyplot.step(x=radius_bins_edges[:-1] / si.um,
-                y=particles.products['dv/dlnr'].get(radius_bins_edges) * rho_w / si.g,
+                y=particles.products['dv/dlnr'].get()[0] * rho_w / si.g,
                 where='post', label=f"t = {step}s")
 
 pyplot.xscale('log')
