### 2021-06-25 9f126ed
Update python code (spectrum moments)
