+++ paper/paper.md
@@ -148,6 +148,7 @@ Subsequently, a `Builder` object is created to orchestrate dependency injection 
   the `Core` class of `PySDM`:
 
 ```python
+import numpy as np
 from PySDM.builder import Builder
 from PySDM.environments import Box
 from PySDM.dynamics import Coalescence
@@ -155,10 +156,13 @@ from PySDM.physics.coalescence_kernels import Golovin
 from PySDM.backends import CPU
 from PySDM.products import ParticlesVolumeSpectrum
 
+radius_bins_edges = np.logspace(
+    np.log10(10 * si.um), np.log10(5e3 * si.um), num=32)
+
 builder = Builder(n_sd=n_sd, backend=CPU)
 builder.set_environment(Box(dt=1 * si.s, dv=1e6 * si.m ** 3))
 builder.add_dynamic(Coalescence(kernel=Golovin(b=1.5e3 / si.s)))
-products = [ParticlesVolumeSpectrum()]
+products = [ParticlesVolumeSpectrum(radius_bins_edges)]
 core = builder.build(attributes, products)
 ```
 
@@ -172,16 +176,12 @@ A minimal simulation example is depicted below with a code snippet and a resulta
 ```python
 from PySDM.physics.constants import rho_w
 from matplotlib import pyplot
-import numpy as np
-
-radius_bins_edges = np.logspace(
-    np.log10(10 * si.um), np.log10(5e3 * si.um), num=32)
 
 for step in [0, 1200, 2400, 3600]:
     core.run(step - core.n_steps)
     pyplot.step(
         x=radius_bins_edges[:-1] / si.um,
-        y=core.products['dv/dlnr'].get(radius_bins_edges) * rho_w/si.g,
+        y=core.products['dv/dlnr'].get()[0] * rho_w/si.g,
         where='post', label=f"t = {step}s")
 
 pyplot.xscale('log')
