### 2021-07-05 0412a4d
uncommenting thrustrtc and curandrtc - they are available now
