### 2021-07-07 c792b67
more physics docstrings
