+++ PySDM/physics/terminal_velocity/__init__.py
@@ -1,0 +1,3 @@
+"""
+Particle terminal velocity formulae
+"""
