### 2021-07-07 atmos-cloud-sim-uj/PySDM#578 @slayoo
with no comments here, I've just posted the reply in the JOSS review issue
