### 2021-07-08 #3219 @slayoo
For the record, PySDM v1.4 is now available on PyPI as well: https://pypi.org/project/PySDM/
