### 2021-07-16 5aace18
adding pypi badge for examples [skip ci]
