### 2021-07-22 #3219 @josephhardinee
I could probably help generically on the python and modeling side, but let me check if a colleague with more domain knowledge, especially around chemistry, would be interested in reviewing for JOSS first.
