### 2021-07-30 #3219 @dhhagan
Hey @darothen thanks for checking in! As of now, we're just waiting on the review from @josephhardinee. I believe the authors made changes in response to your initial comments, so you can either review and update your checklist now or wait until the final revision comes in. Everything is a bit slow right now during the pandemic, but thanks for being on top of it!
