### 2021-08-03 c4476e9
some more README cleanups [skip ci]
