### 2021-08-03 dbe4f2e
trying plotly instead of plotlyjs (to avoid dependency installation)
