### 2021-08-06 79bc577
cleanups, introducing draft of warning handling within condensation (Numba for now) incl. test
