+++ PySDM/backends/numba/impl/condensation_methods.py
@@ -5,7 +5,18 @@ import numba
 import numpy as np
 import math
 from functools import lru_cache
+import sys
 
+
+@numba.njit(**conf.JIT_FLAGS)
+def _warn(msg, context=None, return_value=None):
+    with numba.objmode():
+        print("condensation error:", msg, file=sys.stderr)
+        if context is not None:
+            print("context:", file=sys.stderr)
+            for var in context:
+                print("\t", var, file=sys.stderr)
+    return return_value
 
 class CondensationMethods:
     @staticmethod
@@ -86,8 +97,7 @@ class CondensationMethods:
             success = False
             for burnout in range(fuse + 1):
                 if burnout == fuse:
-                    print("burnout (long)")
-                    return 0, False
+                    return _warn("burnout (long)", context=("thd", thd,), return_value=(0, False))
                 thd_new_long, success = step_fake(args, dt, n_substeps)
                 if success:
                     break
@@ -95,12 +105,10 @@ class CondensationMethods:
                     n_substeps *= multiplier
             for burnout in range(fuse + 1):
                 if burnout == fuse:
-                    print("burnout (short)")
-                    return 0, False
+                    return _warn("burnout (short)", return_value=(0, False))
                 thd_new_short, success = step_fake(args, dt, n_substeps * multiplier)
                 if not success:
-                    print("short failed")
-                    return 0, False
+                    return _warn("short failed", return_value=(0, False))
                 dthd_long = thd_new_long - thd
                 dthd_short = thd_new_short - thd
                 error_estimate = np.abs(dthd_long - multiplier * dthd_short)
@@ -109,7 +117,6 @@ class CondensationMethods:
                     break
                 n_substeps *= multiplier
                 if n_substeps > n_substeps_max:
-                    # print("n_substeps > n_substeps_max (", n_substeps, ") - reached dt_range[0] limit")
                     break
             return np.minimum(n_substeps_max, n_substeps), success
 
@@ -234,7 +241,7 @@ class CondensationMethods:
                         counter += 1
                         if counter > max_iters:
                             if not fake:
-                                print("failed to find interval")
+                                _warn("failed to find interval", context=("T", T, "p", p, "RH", RH, "a", a, "b", b, "fa", fa, "fb", fb))
                             success = False
                             break
                         b = max(x_insane, a + math.ldexp(dx_old, counter))
@@ -250,7 +257,7 @@ class CondensationMethods:
                         x_new, iters_taken = toms748_solve(minfun, args, a, b, fa, fb, rtol_x, max_iters, within_tolerance)
                         if iters_taken in (-1, max_iters):
                             if not fake:
-                                print("TOMS failed")
+                                _warn("TOMS failed")
                             success = False
                             break
                     else:
