### 2021-08-06 2f536b2
cleanup in Numba warn method, condensation/test_diagnostics beef up
