### 2021-08-10 023c103
moving one constant from formulae into constants
