+++ PySDM/physics/surface_tension/compressed_film.py
@@ -1,4 +1,6 @@
-from PySDM.physics import constants as const#, si
+from PySDM.physics import constants as const, si
+import numpy as np
+
 
 class CompressedFilm:
     @staticmethod
@@ -6,7 +8,6 @@ class CompressedFilm:
         # TODO #223 - these parameters should be passed from Settings
         # chemical parameters
         delta_min = 0.2e-9  # minimum organic film thickness [nm]
-        sgm_org = const.sgm_w*0.5      # organic surface tension [N m-1]
         f_org = .5
 
         # convert volumes to diameters
@@ -21,8 +22,8 @@ class CompressedFilm:
         v_delta = const.pi/6 * (d_wet**3 - (d_wet**3 - 2*delta_min)**3)
         
         # calculate the coverage parameter using Eq. (S3.4.2)
-        c_beta = min(v_beta/v_delta, 1)
+        c_beta = np.minimum(v_beta/v_delta, 1)
         
         # calculate sigma
-        sgm = (1-c_beta)*const.sgm_w + c_beta*sgm_org
+        sgm = (1-c_beta)*const.sgm_w + c_beta*const.sgm_org
         return sgm
