### 2021-08-11 a6f929a
cleanups in kindematic_2d examples, moving example list to PySDM-examples README
