+++ PySDM/exporters/__init__.py
@@ -2,3 +2,4 @@
 Exporters module
 """
 from .vtk_exporter import VTKExporter
+from .netcdf_exporter import NetCDFExporter
