+++ PySDM_tests/smoke_tests/Arabas_et_al_2015/test_export.py
@@ -1,8 +1,7 @@
-from PySDM_examples.Arabas_et_al_2015.netcdf_exporter import NetCDFExporter
-from PySDM_examples.Arabas_et_al_2015.settings import Settings
-from PySDM_examples.Arabas_et_al_2015.simulation import Simulation
-from PySDM_examples.Arabas_et_al_2015.storage import Storage
-from PySDM.exporters import VTKExporter
+from PySDM.exporters import NetCDFExporter, VTKExporter
+from PySDM_examples.Arabas_et_al_2015 import Settings, SpinUp
+from PySDM_examples.Szumowski_et_al_1998 import Simulation, Storage
+from PySDM_examples.utils import DummyController
 import tempfile
 
 # noinspection PyUnresolvedReferences
@@ -16,7 +15,7 @@ def test_export(backend, tmp_path):
     settings.output_interval = settings.dt
 
     storage = Storage()
-    simulator = Simulation(settings, storage, backend)
+    simulator = Simulation(settings, storage, SpinUp=SpinUp, backend=backend)
     _, temp_file = tempfile.mkstemp(dir=tmp_path, suffix='.nc')
     sut = NetCDFExporter(storage, settings, simulator, temp_file)
     
@@ -26,6 +25,6 @@ def test_export(backend, tmp_path):
     simulator.run(vtk_exporter=vtk_exporter)
 
     # Act
-    sut.run()
+    sut.run(controller=DummyController())
 
     # Assert
