+++ PySDM/products/dynamics/freezing/ice_water_content.py
@@ -1,0 +1,10 @@
+from ...product import MomentProduct
+
+class IceWaterContent(MomentProduct):
+
+    def __init__(self):
+        super().__init__(
+            name='qi',
+            unit='g/kg',
+            description=f'Ice water mixing ratio'
+        )
