+++ PySDM/products/dynamics/PartMC/__init__.py
@@ -1,0 +1,1 @@
+from .VolumeFractalDimension import VolumeFractalDimension
