### 2021-08-16 ae778d7
fixing lower/upper case issue in product name
