### 2021-08-16 b1dd6ab
sort out uppercase/lowercase issue
