+++ PySDM/products/state/particles_size_spectrum.py
@@ -1,8 +1,5 @@
 import numpy as np
-
 from PySDM.physics import constants as const
-
-
 from PySDM.products.product import SpectrumMomentProduct
 
 
