### 2021-08-20 0c371b9
fixing Numba backend moments range handling (<= instead of < for minimum) + test
