+++ PySDM/state/particles_factory.py
@@ -1,9 +1,7 @@
 import numpy as np
 
-from PySDM.attributes.impl.derived_attribute import DerivedAttribute
-from PySDM.attributes.impl.extensive_attribute import ExtensiveAttribute
+from PySDM.attributes.impl import DerivedAttribute, ExtensiveAttribute, CellAttribute, MaximumAttribute
 from PySDM.attributes.physics.multiplicities import Multiplicities
-from PySDM.attributes.impl.cell_attribute import CellAttribute
 from PySDM.state.particles import Particles
 
 
@@ -14,22 +12,36 @@ class ParticlesFactory:
         idx = core.Index.identity_index(core.n_sd)
 
         extensive_attr = []
+        maximum_attr = []
         for attr_name in req_attr:
             if isinstance(req_attr[attr_name], ExtensiveAttribute):
                 extensive_attr.append(attr_name)
+            elif isinstance(req_attr[attr_name], MaximumAttribute):
+                maximum_attr.append(attr_name)
             elif not isinstance(req_attr[attr_name], (DerivedAttribute, Multiplicities, CellAttribute)):
                 raise AssertionError()
 
         extensive_attributes = core.IndexedStorage.empty(idx, (len(extensive_attr), core.n_sd), float)
+        maximum_attributes = core.IndexedStorage.empty(idx, (len(maximum_attr), core.n_sd), float)
+
         for attr in req_attr.values():
             if isinstance(attr, DerivedAttribute):
                 attr.allocate(idx)
 
         extensive_keys = {}
-        for i, attr in enumerate(extensive_attr):
-            extensive_keys[attr] = i
-            req_attr[attr].set_data(extensive_attributes[i, :])
-            req_attr[attr].init(attributes[attr])
+        maximum_keys = {}
+
+        def helper(req_attr, all_attr, names, data, keys):
+            for i, attr in enumerate(names):
+                keys[attr] = i
+                req_attr[attr].set_data(data[i, :])
+                try:
+                    req_attr[attr].init(all_attr[attr])
+                except KeyError:
+                    raise ValueError(f"attribute '{attr}' required by one of the processes but initial values not provided")
+
+        helper(req_attr, attributes, extensive_attr, extensive_attributes, extensive_keys)
+        helper(req_attr, attributes, maximum_attr, maximum_attributes, maximum_keys)
 
         n = req_attr['n']
         n.allocate(idx)
@@ -60,6 +72,7 @@ class ParticlesFactory:
             core,
             idx,
             extensive_attributes, extensive_keys,
+            # maximum_attributes, maximum_keys, # TODO #594
             cell_start,
             req_attr
         )
