### 2021-08-22 c8c65c3
work in progress on kappa and f_org as particle attributes
