+++ PySDM/attributes/physics/dry_volume.py
@@ -1,12 +1,8 @@
 from PySDM.attributes.impl.extensive_attribute import ExtensiveAttribute
 from PySDM.attributes.impl.derived_attribute import DerivedAttribute
-from PySDM.physics import constants as const
 
-class DryVolumeStatic(ExtensiveAttribute):
-    def __init__(self, builder):
-        super().__init__(builder, name='dry volume')
 
-class DryVolumeDynamic(DerivedAttribute):
+class DryVolumeInorganicDynamic(DerivedAttribute):
     def __init__(self, builder):
         self.core = builder.core
         self.moles_S_VI = builder.get_attribute('moles_S_VI')
@@ -17,18 +13,18 @@ class DryVolumeDynamic(DerivedAttribute):
         self.data.data[:] = self.moles_S_VI.data.data[:]
         self.data.data[:] *= dynamic.dry_molar_mass / dynamic.dry_rho
 
-# TODO #223
-# first attempt to add inorganic and organic dry volumes as extensive attributes
-# and make DryVolumeStatic a derived attribute
+
 class DryVolumeInorganic(ExtensiveAttribute):
     def __init__(self, builder):
         super().__init__(builder, name='dry volume inorganic')
+
 
 class DryVolumeOrganic(ExtensiveAttribute):
     def __init__(self, builder):
         super().__init__(builder, name='dry volume organic')
 
-class DryVolumeOrgInorg(DerivedAttribute):
+
+class DryVolume(DerivedAttribute):
     def __init__(self, builder):
         self.volume_dry_org = builder.get_attribute('dry volume organic')
         self.volume_dry_inorg = builder.get_attribute('dry volume inorganic')
@@ -36,5 +32,15 @@ class DryVolumeOrgInorg(DerivedAttribute):
         super().__init__(builder, name='dry volume', dependencies=dependencies)
 
     def recalculate(self):
-        self.data.idx = self.volume_dry_org.data.idx
         self.data.sum(self.volume_dry_org.get(), self.volume_dry_inorg.get())
+
+
+class OrganicFraction(DerivedAttribute):
+    def __init__(self, builder):
+        self.volume_dry_org = builder.get_attribute('dry volume organic')
+        self.volume_dry = builder.get_attribute('dry volume')
+        dependencies = [self.volume_dry_org, self.volume_dry]
+        super().__init__(builder, name='dry volume organic fraction', dependencies=dependencies)
+
+    def recalculate(self):
+        self.data.ratio(self.volume_dry_org.get(), self.volume_dry.get())
