+++ PySDM/attributes/impl/mapper.py
@@ -1,7 +1,9 @@
 from PySDM.attributes.physics.multiplicities import Multiplicities
 from PySDM.attributes.physics.volume import Volume
-from PySDM.attributes.physics.dry_volume import DryVolumeOrganic, DryVolumeInorganic
-from PySDM.attributes.physics.dry_volume import DryVolumeOrgInorg, DryVolumeDynamic, DryVolumeStatic
+from PySDM.attributes.physics.dry_volume import (DryVolumeOrganic, DryVolumeInorganic,
+                                                 DryVolume, DryVolumeInorganicDynamic,
+                                                 OrganicFraction)
+from PySDM.attributes.physics.hygroscopicity import Kappa, KappaTimesDryVolume
 from PySDM.attributes.physics.radius import Radius
 from PySDM.attributes.physics.dry_radius import DryRadius
 from PySDM.attributes.physics.terminal_velocity import TerminalVelocity
@@ -22,9 +24,12 @@ attributes = {
     'n': lambda _: Multiplicities,
     'volume': lambda _: Volume,
     'dry volume organic': lambda _: DryVolumeOrganic,
-    'dry volume inorganic': lambda _: DryVolumeInorganic,
-    # 'dry volume': lambda dynamics: DryVolumeDynamic if 'AqueousChemistry' in dynamics else (DryVolumeOrgInorg if 'SurfaceOrganics' in dynamics else DryVolumeStatic),
-    'dry volume': lambda dynamics: DryVolumeDynamic if 'AqueousChemistry' in dynamics else DryVolumeStatic,
+    'dry volume inorganic': lambda dynamics:
+        DryVolumeInorganicDynamic if 'AqueousChemistry' in dynamics else DryVolumeInorganic,
+    'dry volume': lambda _: DryVolume,
+    'dry volume organic fraction': lambda _: OrganicFraction,
+    'kappa times dry volume': lambda _: KappaTimesDryVolume,
+    'kappa': lambda _: Kappa,
     'radius': lambda _: Radius,
     'dry radius': lambda _: DryRadius,
     'terminal velocity': lambda _: TerminalVelocity,
