+++ PySDM/environments/kinematic_1d.py
@@ -44,7 +44,11 @@ class Kinematic1D(_Moist):
                 self.mesh.cellular_attributes(positions)
 
             r_dry, n_per_kg = spectral_discretisation.sample(self.core.n_sd)
-            r_wet = r_wet_init(r_dry, self, cell_id=attributes['cell id'], kappa=kappa)
+            attributes['dry volume inorganic'] = self.formulae.trivia.volume(radius=r_dry)
+            attributes['dry volume organic'] = np.zeros_like(r_dry)
+            attributes['kappa times dry volume'] = attributes['dry volume inorganic'] * kappa
+            r_wet = r_wet_init(r_dry, self, cell_id=attributes['cell id'],
+                               kappa_times_dry_volume=attributes['kappa times dry volume'])
 
             rhod = self['rhod'].to_ndarray()
             cell_id = attributes['cell id']
@@ -52,7 +56,6 @@ class Kinematic1D(_Moist):
 
         attributes['n'] = discretise_n(n_per_kg * rhod[cell_id] * domain_volume)
         attributes['volume'] = self.formulae.trivia.volume(radius=r_wet)
-        attributes['dry volume'] = self.formulae.trivia.volume(radius=r_dry)
 
         return attributes
 

[…]
