### 2021-08-22 aa36fe8
fixing Julia and Matlab syntax in README
