+++ README.md
@@ -374,7 +374,7 @@ r_wet = r_wet_init(r_dry, env, kappa * v_dry)
 attributes = Dict()
 attributes["n"] = multiplicities.discretise_n(specific_concentration * env.mass_of_dry_air)
 attributes["dry volume inorganic"] = v_dry
-attributes["dry volume organic"] = Array{Float32}(0, n_sd)
+attributes["dry volume organic"] = zeros(n_sd)
 attributes["kappa times dry volume"] = kappa * v_dry
 attributes["volume"] = builder.formulae.trivia.volume(radius=r_wet) 
 
@@ -457,8 +457,8 @@ r_wet = r_wet_init(r_dry, env, kappa * v_dry);
 attributes = py.dict(pyargs( ...
     'n', multiplicities.discretise_n(specific_concentration * env.mass_of_dry_air), ...
     'dry volume inorganic', v_dry, ...
-    'dry volume organic', zeros(n_sd) ...
-    'kappa times dry volume', kappa * v_dry ... 
+    'dry volume organic', zeros(n_sd), ...
+    'kappa times dry volume', kappa * v_dry, ... 
     'volume', builder.formulae.trivia.volume(r_wet) ...
 ));
 
