+++ README.md
@@ -457,7 +457,7 @@ r_wet = r_wet_init(r_dry, env, kappa * v_dry);
 attributes = py.dict(pyargs( ...
     'n', multiplicities.discretise_n(specific_concentration * env.mass_of_dry_air), ...
     'dry volume inorganic', v_dry, ...
-    'dry volume organic', zeros(n_sd), ...
+    'dry volume organic', py.numpy.zeros(int32(n_sd)), ...
     'kappa times dry volume', kappa * v_dry, ... 
     'volume', builder.formulae.trivia.volume(r_wet) ...
 ));
