### 2021-08-23 5d3878e
fixing bug in sum_out_of_place
