### 2021-08-23 3cb6609
sample randomly from spectrum for initialization sizes. Closes #156
