### 2021-08-23 2fc8685
syntax fix for older Pythons
