+++ PySDM/dynamics/freezing.py
@@ -1,6 +1,7 @@
 class Freezing:
     def __init__(self, *, singular=True):
         self.singular = singular
+        self.enable = True
 
     def register(self, builder):
         if self.singular:
@@ -13,3 +14,16 @@ class Freezing:
     def __call__(self):
         if 'Coalescence' in self.core.dynamics:
             raise NotImplementedError("handling T_fz during collisions not implemented yet")  # TODO #594
+
+        if not self.enable:
+            return
+
+        self.core.bck.freeze(
+            T_fz=self.core.particles['freezing temperature'],
+            v_wet=self.core.particles['volume'],
+            T=self.core.environment['T'],
+            RH=self.core.environment['RH'],
+            cell=self.core.particles['cell id']
+        )
+
+        self.core.particles.attributes['volume'].mark_updated()
