+++ PySDM_tests/unit_tests/physics/test_freezing_temperature_spectra.py
@@ -6,7 +6,7 @@ import numpy as np
 def test_freezing_temperature_spectra(plot=False):
     # Arrange
     formulae = Formulae()
-    T = np.linspace(const.T0 - 40, const.T0, num=100)
+    T = np.linspace(const.T0, const.T0 - 40, num=100)
 
     # Act
     pdf = formulae.freezing_temperature_spectrum.pdf(T)
@@ -19,12 +19,14 @@ def test_freezing_temperature_spectra(plot=False):
         cdfax = pdfax.twinx()
         cdfax.plot(T, cdf, linestyle='--', marker='x', label='cdf')
         pylab.xlabel('T [K]')
+        pylab.xlim(np.amax(T), np.amin(T))
         pdfax.set_ylabel('pdf [K$^{-1}$]')
         cdfax.set_ylabel('cdf [1]')
+        pylab.grid()
         pylab.show()
 
     # Assert
-    dT = T[1] - T[0]
+    dT = abs(T[1] - T[0])
     np.testing.assert_approx_equal(np.sum(pdf * dT), 1)
     np.testing.assert_approx_equal(cdf[0]+1, 1)
     np.testing.assert_approx_equal(cdf[-1], 1)
