### 2021-08-30 9e14b39
improved error handling in r_wet_init
