### 2021-08-31 66f7513
better diagnostics in r_wet_init (using the warn utility)
