### 2021-08-31 93be081
changing delta_min value to the one apparently used in Lowe paper - thanks Clare!
