+++ PySDM/physics/constants.py
@@ -69,9 +69,12 @@ FWC_C8 = -.321582393e-15 * si.hPa / si.K**8
 rho_w = 1 * si.kilograms / si.litres
 rho_i = 916.8 * si.kg / si.metres**3
 pH_w = 7
-sgm_w = 0.072 * si.joule / si.metre ** 2  # TODO #223 temperature dependence
+sgm_w = 0.072 * si.joule / si.metre ** 2
 sgm_org = 40 * si.mN / si.m
-delta_min = 0.2 * si.nm  # minimum organic film thickness [nm]
+
+# minimum organic film thickness
+# TODO #223 0.2 in the paper, but 0.1 matches the paper plots
+delta_min = 0.1 * si.nm
 
 p_tri = 611.73 * si.pascal
 T_tri = 273.16 * si.kelvin
