### 2021-09-02 39a9bb4
fixing (or rather workarounding) test_just_do_it for for the cold cumulus example; handling kappa=0 in r_wet_init
