### 2021-09-03 bd88162
sorting out trouble with r_wet_init left interval bound for RH < RH_eq(r_d) at t=0 + test
