+++ PySDM/environments/parcel.py
@@ -65,9 +65,9 @@ class Parcel(_Moist):
             n_in_dv = np.array([n_in_dv])
 
         attributes = {}
-        attributes['dry volume inorganic'] = self.formulae.trivia.volume(radius=r_dry)
+        attributes['dry volume'] = self.formulae.trivia.volume(radius=r_dry)
         attributes['dry volume organic'] = np.zeros_like(r_dry)
-        attributes['kappa times dry volume'] = attributes['dry volume inorganic'] * kappa
+        attributes['kappa times dry volume'] = attributes['dry volume'] * kappa
         attributes['n'] = discretise_n(n_in_dv)
         r_wet = r_wet_init(r_dry, self, kappa_times_dry_volume=attributes['kappa times dry volume'], rtol=rtol)
         attributes['volume'] = self.formulae.trivia.volume(radius=r_wet)
