### 2021-09-04 72e87a9
setting attribute data to nan for dummy attributes; corrected check for attribute presence in input data for dummy attr; marking base organic attr as dummy (not only f_org); fixing condensation test_diagnostic (removing organic vol attr)
