### 2021-09-04 5aa1212
organic attr cleanup, bump examples req
