+++ PySDM/environments/parcel.py
@@ -66,7 +66,6 @@ class Parcel(_Moist):
 
         attributes = {}
         attributes['dry volume'] = self.formulae.trivia.volume(radius=r_dry)
-        attributes['dry volume organic'] = np.zeros_like(r_dry)
         attributes['kappa times dry volume'] = attributes['dry volume'] * kappa
         attributes['n'] = discretise_n(n_in_dv)
         r_wet = r_wet_init(r_dry, self, kappa_times_dry_volume=attributes['kappa times dry volume'], rtol=rtol)
