+++ PySDM/environments/kinematic_1d.py
@@ -45,7 +45,6 @@ class Kinematic1D(_Moist):
 
             r_dry, n_per_kg = spectral_discretisation.sample(self.core.n_sd)
             attributes['dry volume'] = self.formulae.trivia.volume(radius=r_dry)
-            attributes['dry volume organic'] = np.zeros_like(r_dry)
             attributes['kappa times dry volume'] = attributes['dry volume'] * kappa
             r_wet = r_wet_init(r_dry, self, cell_id=attributes['cell id'],
                                kappa_times_dry_volume=attributes['kappa times dry volume'])
