+++ PySDM/environments/kinematic_2d.py
@@ -54,7 +54,6 @@ class Kinematic2D(_Moist):
                 raise NotImplementedError()
 
             attributes['dry volume'] = self.formulae.trivia.volume(radius=r_dry)
-            attributes['dry volume organic'] = np.zeros_like(r_dry)
             attributes['kappa times dry volume'] = kappa * attributes['dry volume']
             if kappa == 0:
                 r_wet = r_dry
