### 2021-09-04 c85c86b
fix dummy attr name and particles factory error handling
