+++ PySDM/attributes/impl/mapper.py
@@ -1,4 +1,4 @@
-from PySDM.attributes.impl.dummy_attribute import DummyAttribute
+from PySDM.attributes.impl.dummy_attribute import DummyAttributeImpl
 from PySDM.attributes.physics.dry_volume import DryVolumeOrganic, DryVolume, DryVolumeDynamic, OrganicFraction
 from PySDM.attributes.physics.hygroscopicity import Kappa, KappaTimesDryVolume
 from PySDM.attributes.physics import (Multiplicities, Volume, Radius, DryRadius,
@@ -14,13 +14,13 @@ attributes = {
     'n': lambda _: Multiplicities,
     'volume': lambda _: Volume,
     'dry volume organic': lambda dynamics: (
-        DummyAttribute if isinstance(dynamics['Condensation'].core.formulae.surface_tension, Constant)
+        DummyAttributeImpl('dry volume organic') if isinstance(dynamics['Condensation'].core.formulae.surface_tension, Constant)
         else DryVolumeOrganic
     ),
     'dry volume': lambda dynamics:
     DryVolumeDynamic if 'AqueousChemistry' in dynamics else DryVolume,
     'dry volume organic fraction': lambda dynamics: (
-        DummyAttribute if isinstance(dynamics['Condensation'].core.formulae.surface_tension, Constant)
+        DummyAttributeImpl('dry volume organic fraction') if isinstance(dynamics['Condensation'].core.formulae.surface_tension, Constant)
         else OrganicFraction
     ),
     'kappa times dry volume': lambda _: KappaTimesDryVolume,
