### 2021-09-05 12e4348
using --use-deprecated=legacy-resolver for all pip CI invocations
