+++ .github/workflows/julia.yml
@@ -18,8 +18,8 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v1
-      - run: pip install -e .
-      - run: pip install exdown pytest
+      - run: pip install --use-deprecated=legacy-resolver -e .
+      - run: pip install --use-deprecated=legacy-resolver exdown pytest
       - run: python -c "import exdown; code=exdown.extract('README.md'); f=open('readme.jl', 'w'); f.writelines(block.code for block in code if block.syntax=='Julia'); f.close()"
       - uses: julia-actions/setup-julia@v1
       - run: julia readme.jl
