+++ .travis.yml
@@ -26,18 +26,18 @@ jobs:
       os: windows
       before_install:
         - choco install python --version 3.7
-        - python -m pip install --upgrade pip
+        - python -m pip install --use-deprecated=legacy-resolver --upgrade pip
       env: PATH=/c/Python37:/c/Python37/Scripts:$PATH
 
 install:
-  - pip$PY_SFX install -U pytest-cov codecov setuptools>=50.3.2
+  - pip$PY_SFX install --use-deprecated=legacy-resolver -U pytest-cov codecov setuptools>=50.3.2
   - |
     # https://stackoverflow.com/questions/55486248/deprecationwarning-in-python-3-6-and-3-7-with-pillow-distutils-imp
     if [[ $TRAVIS_OS_NAME == linux ]]; then
       python$PY_SFX $PY_FLAGS -c 'import os, virtualenv; virtualenv.install_distutils(os.environ["VIRTUAL_ENV"])' || exit 1
     fi;
-  - pip$PY_SFX install -U -r test-time-requirements.txt
-  - pip$PY_SFX install -e .
+  - pip$PY_SFX install --use-deprecated=legacy-resolver -U -r test-time-requirements.txt
+  - pip$PY_SFX install --use-deprecated=legacy-resolver -e .
   - export PY_FLAGS="-We"
 
 script:
