+++ appveyor.yml
@@ -25,17 +25,17 @@ install:
       }
 
   - ps: |
-      iex "$env:PYTHON -m pip install --upgrade pip"
+      iex "$env:PYTHON -m pip install --use-deprecated=legacy-resolver --upgrade pip"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
       
-      iex "$env:PYTHON -m pip install pytest"
+      iex "$env:PYTHON -m pip install --use-deprecated=legacy-resolver pytest"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
       
       # Add-Content requirements.txt "`npywin32>=300"
-      iex "$env:PYTHON -m pip install -r test-time-requirements.txt"
+      iex "$env:PYTHON -m pip install --use-deprecated=legacy-resolver -r test-time-requirements.txt"
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
       
-      iex "$env:PYTHON -m pip install -e ."
+      iex "$env:PYTHON -m pip install --use-deprecated=legacy-resolver -e ."
       if ($LastExitCode -ne 0) { $host.SetShouldExit($LastExitCode)  }
 
 test_script:
