+++ .github/workflows/main.yml
@@ -27,13 +27,13 @@ jobs:
           python-version: ${{ matrix.python-version }}
 
       - run: |
-          pip install --use-deprecated=legacy-resolver -e .
+          pip --use-deprecated=legacy-resolver install -e .
 
       - run: |
           python -We -c "import PySDM"
 
       - run: |
-          pip install --use-deprecated=legacy-resolver -r test-time-requirements.txt
+          pip --use-deprecated=legacy-resolver install -r test-time-requirements.txt
 
       # https://github.com/numba/numba/issues/6350#issuecomment-728174860
       - if: matrix.platform == 'ubuntu-latest'
