+++ .github/workflows/python.yml
@@ -18,8 +18,8 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v1
-      - run: pip install --use-deprecated=legacy-resolver -e .
-      - run: pip install --use-deprecated=legacy-resolver exdown pytest
+      - run: pip --use-deprecated=legacy-resolver install -e .
+      - run: pip --use-deprecated=legacy-resolver install exdown pytest
       - run: python -c "import exdown; code=exdown.extract('README.md'); f=open('readme.py', 'w'); f.writelines(block.code for block in code if block.syntax=='Python'); f.close()"
       - run: python -We readme.py
       - run: sed -i 's/CPU/GPU/g' readme.py
