### 2021-09-05 3f12406
removing --use-deprecated=legacy-resolver from appveyor file
