### 2021-09-09 b254320
removing .travis.yml; adding codecov GA file
