### 2021-09-09 c7f30a4
fixing codecov (installing PySDM itself and test-time-requirements)
