### 2021-09-10 d4195af
introducing immersed surface area attribute + handling in freezing method
