+++ PySDM/attributes/ice/__init__.py
@@ -1,2 +1,2 @@
 from .freezing_temperature import FreezingTemperature
-from .nucleation_sites import NucleationSites
+from .immersed_surface_area import ImmersedSurfaceArea
