+++ PySDM/physics/spectra.py
@@ -49,6 +49,9 @@ class Lognormal(Spectrum):
     def __init__(self, norm_factor: float, m_mode: float, s_geom: float):
         super().__init__(lognorm, (math.log(s_geom), 0, m_mode), norm_factor)
 
+    @property
+    def s_geom(self):
+        return math.exp(self.distribution_params[0])
 
 class Sum:
 
