### 2021-09-12 f649326
adding section on conda to the README [skip ci]
