### 2021-09-13 af87e71
adding smoke test for Alpert & Knopf 2016 Fig 1
