### 2021-09-13 70596aa
plot=False in smoke test for Alpert & Knopf 2016 Fig 1
