+++ PySDM_tests/smoke_tests/Alpert_and_Knopf_2016/test_AK16_fig_1.py
@@ -14,7 +14,7 @@ import pytest
 n_runs_per_case = 3
 
 @pytest.mark.parametrize("multiplicity", (1, 2, 10))
-def test_fig_1(multiplicity, plot=False):
+def test_AK16_fig_1(multiplicity, plot=False):
     # Arrange
     J_het = 1e3 / si.cm ** 2 / si.s
     A_g = 1e-5 * si.cm ** 2
