+++ PySDM_tests/smoke_tests/Alpert_and_Knopf_2016/test_AK16_fig_1.py
@@ -1,12 +1,6 @@
-from PySDM.backends import CPU
-from PySDM import Builder
-from PySDM.environments import Box
-from PySDM.dynamics import Freezing
+from PySDM_examples.Alpert_and_Knopf_2016 import simulation
 from PySDM.physics import si, constants as const, Formulae
-from PySDM.initialisation.multiplicities import discretise_n
-from PySDM.initialisation import spectral_sampling
 from PySDM.physics.spectra import Lognormal
-from PySDM.products import IceWaterContent
 import numpy as np
 from matplotlib import pylab
 import pytest
@@ -45,65 +39,39 @@ def test_AK16_fig_1(multiplicity, plot=False):
             assert int(n_sd) == n_sd
             n_sd = int(n_sd)
 
-            formulae = Formulae(seed=seed)
-            builder = Builder(n_sd=n_sd, backend=CPU, formulae=formulae)
-            builder.set_environment(Box(dt=dt, dv=dv))
-            builder.add_dynamic(Freezing(singular=False, J_het=J_het))
-
-            if case['ISA'].s_geom != 1:
-                _isa, _conc = spectral_sampling.ConstantMultiplicity(case['ISA']).sample(n_sd)
-            else:
-                _isa, _conc = np.full(n_sd, A_g), np.full(n_sd, multiplicity / dv)
-            attributes = {
-                'n': discretise_n(_conc * dv),
-                'immersed surface area': _isa,
-                'volume': np.full(n_sd, droplet_volume)
-            }
-            np.testing.assert_almost_equal(attributes['n'], multiplicity)
-            products = (IceWaterContent(specific=False),)
-            particulator = builder.build(attributes=attributes, products=products)
-
-            cell_id = 0
-            data = []
-            for i in range(int(total_time / dt) + 1):
-                particulator.run(0 if i == 0 else 1)
-
-                ice_mass_per_volume = particulator.products['qi'].get()[cell_id]
-                ice_mass = ice_mass_per_volume * dv
-                ice_number = ice_mass / (const.rho_w * droplet_volume)
-                unfrozen_fraction = 1 - ice_number / number_of_real_droplets
-                data.append(unfrozen_fraction)
-
+            data = simulation(seed=i, n_sd=n_sd, dt=dt, dv=dv, spectrum=case['ISA'],
+                          droplet_volume=droplet_volume, multiplicity=multiplicity, J_het=J_het,
+                          total_time=total_time, number_of_real_droplets=number_of_real_droplets)
             output[key].append(data)
 
     # Plot
-    pylab.rc('font', size=10)
-    for key in output.keys():
-        for run in range(n_runs_per_case):
-            label = f"{key}: σ=ln({int(cases[key]['ISA'].s_geom)}),N={int(cases[key]['ISA'].norm_factor * dv)}"
+    if plot:
+        for key in output.keys():
+            for run in range(n_runs_per_case):
+                label = f"{key}: σ=ln({int(cases[key]['ISA'].s_geom)}),N={int(cases[key]['ISA'].norm_factor * dv)}"
+                pylab.step(
+                    dt / si.min * np.arange(len(output[key][run])),
+                    output[key][run],
+                    label=label if run == 0 else None,
+                    color=cases[key]['color'],
+                    linewidth=.666
+                )
+            output[key].append(np.mean(np.asarray(output[key]), axis=0))
             pylab.step(
-                dt / si.min * np.arange(len(output[key][run])),
-                output[key][run],
-                label=label if run == 0 else None,
[…]
