### 2021-09-13 f5ee8a8
Temporary fix for non-stochastic behavior
