+++ PySDM/dynamics/coalescence.py
@@ -14,6 +14,7 @@ class Coalescence:
     def __init__(self,
                  kernel,
                  coal_eff,
+                 seed=None,
                  croupier=None,
                  optimized_random=False,
                  substeps: int = 1,
@@ -27,6 +28,7 @@ class Coalescence:
 
         self.kernel = kernel
         self.coal_eff = coal_eff
+        self.seed = seed
 
         assert dt_coal_range[0] > 0
         self.croupier = croupier
@@ -51,7 +53,7 @@ class Coalescence:
         self.core = builder.core
         self.rnd_opt = RandomGeneratorOptimizer(optimized_random=self.optimized_random,
                                                 dt_min=self.dt_coal_range[0],
-                                                seed=self.core.formulae.seed)
+                                                seed=self.seed) #self.core.formulae.seed)
         self.optimised_random = None
 
         if self.core.n_sd < 2:
