### 2021-09-14 848dff9
adding TopHat spectrum (+ tests); bump examples req
