### 2021-09-14 8a361ca
bump examples req; fix plot=True in test
