### 2021-09-16 b08738b
fix in tests for default adaptivity in coalescence
