+++ PySDM_tests/unit_tests/dynamics/coalescence/test_sdm_single_cell.py
@@ -1,221 +1,0 @@
-import numpy as np
-import pytest
-
-from PySDM.storages.pair_indicator import make_PairIndicator
-from PySDM.storages.indexed_storage import make_IndexedStorage
-from PySDM.storages.index import make_Index
-# noinspection PyUnresolvedReferences
-from PySDM_tests.backends_fixture import backend
-from PySDM_tests.unit_tests.dynamics.coalescence.__parametrisation__ import backend_fill
-from PySDM_tests.unit_tests.dynamics.coalescence.__parametrisation__ import get_dummy_core_and_sdm
-# noinspection PyUnresolvedReferences
-from PySDM_tests.unit_tests.dynamics.coalescence.__parametrisation__ import v_2, T_2, n_2
-
-
-class TestSDMSingleCell:
-
-    @staticmethod
-    def test_single_collision(backend, v_2, T_2, n_2):
-        # Arrange
-        const = 1.
-        core, sut = get_dummy_core_and_sdm(backend, len(n_2))
-        sut.compute_gamma = lambda prob, rand, is_first_in_pair: backend_fill(prob, 1)
-        attributes = {'n': n_2, 'volume': v_2, 'heat': const*T_2*v_2, 'temperature': T_2}
-        core.build(attributes)
-
-        # Act
-        sut()
-
-        # Assert
-        particles = core.particles
-        a = particles['n'].to_ndarray()
-        b = particles['volume'].to_ndarray()
-        c = particles['temperature'].to_ndarray()
-        np.testing.assert_approx_equal(
-            const * np.sum(particles['n'].to_ndarray() * particles['volume'].to_ndarray() * particles['temperature'].to_ndarray()),
-            const * np.sum(n_2 * T_2 * v_2),
-            significant=7
-        )
-        new_T = np.sum(T_2 * v_2) / np.sum(v_2)
-        assert np.isin(round(new_T, 7), np.round(particles['temperature'].to_ndarray().astype(float), 7))
-
-        assert np.sum(particles['n'].to_ndarray() * particles['volume'].to_ndarray()) == np.sum(n_2 * v_2)
-        assert np.sum(core.particles['n'].to_ndarray()) == np.sum(n_2) - np.amin(n_2)
-        if np.amin(n_2) > 0:
-            assert np.amax(core.particles['volume'].to_ndarray()) == np.sum(v_2)
-        assert np.amax(core.particles['n'].to_ndarray()) == max(np.amax(n_2) - np.amin(n_2), np.amin(n_2))
-
-    @staticmethod
-    @pytest.mark.parametrize("n_in, n_out", [
-        pytest.param(1, np.array([1, 0])),
-        pytest.param(2, np.array([1, 1])),
-        pytest.param(3, np.array([2, 1])),
-    ])
-    def test_single_collision_same_n(backend, n_in, n_out):
-        # Arrange
-        core, sut = get_dummy_core_and_sdm(backend, 2)
-        sut.compute_gamma = lambda prob, rand, is_first_in_pair: backend_fill(prob, 1)
-        attributes = {'n': np.full(2, n_in), 'volume': np.full(2, 1.)}
-        core.build(attributes)
-
-        # Act
-        sut()
-
-        # Assert
-        np.testing.assert_array_equal(sorted(core.particles['n'].to_ndarray(raw=True)), sorted(n_out))
-
-    @staticmethod
-    @pytest.mark.parametrize("p", [
-        pytest.param(2),
-        pytest.param(4),
-        pytest.param(5),
-        pytest.param(7),
-    ])
-    def test_multi_collision(backend, v_2, n_2, p):
-        # Arrange
-        core, sut = get_dummy_core_and_sdm(backend, len(n_2))
-
-        def _compute_gamma(prob, rand, is_first_in_pair):
[…]

[…]
