### 2021-09-17 atmos-cloud-sim-uj/PySDM#494 @slayoo
Let's close this and the PR for now. If there is any reason why conda users should not just open anaconda prompt and do ``pip install PySDM``, please reopen
