### 2021-09-18 #3219 @slayoo
(API and affiliation updates)
