### 2021-09-27 #3219 @slayoo
(added mention of one additional example based on Lowe et al. 2019 and added Clare Singer, who contributed the example, to the author list)
