### 2021-10-12 #3219 @josephhardinee
I've completed the review (Sorry it took so long, the package was quite comprehensive). Everything looks good with possibly one exception. The API level documentation is lacking detail once you get below the top level with most methods having no documentation, especially on inputs/outputs (See image below for instance). However, the authors have a fairly robust set of example notebooks that I would argue serves as decent documentation(https://github.com/atmos-cloud-sim-uj/PySDM-examples). I do think we should leave a note encouraging authors to expand API level docs. With that note I would be okay approving this if that meets JOSS standards as the authors appear to be very active in development. @dhhagan would this be acceptable from an editorial viewpoint?

![image](https://user-images.githubusercontent.com/761704/136965772-661a7e2f-60a3-40a8-9423-cb276db3d7fa.png)
