### 2021-11-01 #3219 @slayoo
@darothen @dhhagan Let me just confirm, we are actively cleaning up the code towards 100% pylint-clean (incl. docstrings):
- https://github.com/atmos-cloud-sim-uj/PySDM/pull/663
- https://github.com/atmos-cloud-sim-uj/PySDM/pull/664

Still, we need some days to finish it up.

BTW, we have recently rolled out a proof-of-concept coupler for ClimateMachine.jl and PySDM (with 2D prescribed-flow test case so far featuring PySDM particle-based microphyics and ClimateMachine's fluid flow):
https://github.com/CliMA/ClimateMachine.jl/pull/2244
