### 2021-12-14 #3219 @slayoo
> We have already set up a pylint workflow for the project which gives us indications where docstrings (and other things) are missing or need improvement. Will now focus on improving it and report back here within a few weeks.

Dear @dhhagan, @josephhardinee & @darothen,

As of PySDM v1.24 (released just now) 100% of PySDM Python modules have docstrings included. These are top-file module docstrings, mostly 1-2 lines, not yet a full class/method coverage, but given that most of PySDM source files are single-class modules, this should give the users/contributors a meaningful overview of the project when browsing the online docs at https://atmos-cloud-sim-uj.github.io/PySDM.

Moreover, the presence of the module-level docstrings is now ensured at CI level for all PRs using `pylint`, so new code will follow suit. The documentation is generated also during CI builds (and deployed when merging to `master`), with `python -We -m pdoc`, causing any code reference mismatches within the docstrings to be detected and flagged by triggering CI failure. Thus, refactors missing docstring updates will not pass unnoticed.

The documentation sprint has been an occasion for significant cleanups and some refactors aimed at simplifying directory structure and API (all reflected in tests, examples and README Python/Julia/Matlab snippets - as all are checked as a part of CI builds). The code snippets in the paper have also been updated (public API changes were minor, mainly requiring updates of import lines).

We are planning further steady buildup of the docstrings coverage as well as rollouts of other user-oriented resources.

We hope that the reached interim milestone of 100% module docstring coverage and the introduced CI documentation-related automation can serve as a solid indication of our commitment to maintain and enhance PySDM documentation. Looking forward to your comments and acceptance of the paper for publication in JOSS.

Thank you for your consideration,
Sylwester
