### 2021-12-17 #3219 @josephhardinee
I'd like to commend the authors for seeing through their commitment to improve documentation as well as the accompanying code cleanups. This looks good to me and I think this greatly deserves publication in JOSS.
