### 2022-01-06 #3219 @slayoo
For the record, from now on (https://github.com/atmos-cloud-sim-uj/PySDM/pull/717) the code snippets included in the JOSS paper will be executed as a part of PySDM CI build to ensure the example code works well with future updates to PySDM.
