### 2022-01-31 #3219 @darothen
Sorry for the delay folks; between job change + holidays + AMS it took far longer than I had hoped to check in here.

I reviewed all of the changes and the discourse from the past few months. I think all of my comments/recommendations have either been directly addressed or addressed in the spirit of the substantial improvements and additions made to the library since the first version I reviewed last year. In my opinion, the work here is excellent and is ready for publication in JOSS without any additional modifications.
