### 2022-03-01 #3219 @slayoo
We have moved ahead of the "v1" branch some time ago for this project,
so here are DOIs for Zenodo-archived last release from the v1 branch described in this paper:
- PySDM v1.27: https://doi.org/10.5281/zenodo.6321270
- PySDM-examples v1.27: https://doi.org/10.5281/zenodo.6321279

Thank you
