### 2022-03-04 #3219 @slayoo
just added one more funding acknowledgement to the paper.md
