### 2022-03-04 #3219 @slayoo
@dhhagan the markdown-to-latex-to-pdf workflow makes it tricky for us to control line breaks in the list of authors and affiliations:
![image](https://user-images.githubusercontent.com/89685/156810607-9a73177e-846f-4fe6-aba2-8ed67eaafc7e.png)

Some kludge workarounds seem to be possible by inserting non-braking spaces here and there, but it is cumbersome, and I'm not sure if it will actually even be taken into account when producing the final pdf - please confirm if we should "fight" these problems ourselves or if there is any final hand-tuning done for such aspects at JOSS. Thanks!
