+++ paper/paper.md
@@ -1,12 +1,12 @@
 ---
-title: 'PySDM v1: particle-based cloud modelling package for&nbsp;warm-rain microphysics and aqueous chemistry'
+title: 'PySDM v1: particle-based cloud modeling package for&nbsp;warm-rain microphysics and aqueous chemistry'
 date: 31 March 2021
 tags:
   - Python
   - physics-simulation 
   - monte-carlo-simulation 
   - gpu-computing 
-  - atmospheric-modelling 
+  - atmospheric-modeling 
   - particle-system 
   - numba 
   - thrust 
@@ -63,14 +63,14 @@ bibliography: paper.bib
 It is intended to serve as a building block for process-level as well as computational-fluid dynamics simulation systems involving representation
   of a continuous phase (air) and a dispersed phase (aerosol), with `PySDM` being responsible for representation of the dispersed phase. 
 For major version 1 (v1), the development has been focused on atmospheric cloud physics applications, in particular on 
-  modelling the dynamics of particles immersed in moist air using the particle-based 
+  modeling the dynamics of particles immersed in moist air using the particle-based 
   approach to represent 
   the evolution of the size spectrum of aerosol/cloud/rain particles. 
 The particle-based approach contrasts the more commonly used bulk and bin methods
   in which atmospheric particles are segregated into multiple categories (aerosol, cloud and rain) 
   and their evolution is governed by deterministic dynamics solved on the same Eulerian grid as 
   the dynamics of the continuous phase. 
-Particle-based methods employ discrete computational (super) particles for modelling the dispersed phase.
+Particle-based methods employ discrete computational (super) particles for modeling the dispersed phase.
 Each super particle is associated with a set of continuously-valued attributes evolving in Lagrangian manner. 
 Such approach is particularly well suited for using probabilistic representation of 
   particle collisional growth (coagulation) and for representing processes dependent 
@@ -92,7 +92,7 @@ Furthermore, representation of dissolution and, if applicable, dissociation
 Representation of the chemical processes follows the particle-based formulation of @Jaruga_and_Pawlowska_2018.
 
 The usage examples are built on top of four different `environment` classes included in `PySDM` v1
-  which implement common simple atmospheric cloud modelling frameworks: box, adiabatic
+  which implement common simple atmospheric cloud modeling frameworks: box, adiabatic
   parcel, single-column, and 2D prescribed flow kinematic models.
 
 In addition, the package ships with tutorial code depicting how `PySDM` can be used from `Julia` and `Matlab` using
@@ -149,7 +149,7 @@ attributes['volume'], attributes['n'] = sampling.sample(N_SD)
 
 In the above snippet, the `si` is an instance of the `FakeUnitRegistry` class.
 The exponential distribution of particle volumes is sampled at $2^{17}$ points 
-  in order to initialise two key attributes of the super-droplets, namely their volume and multiplicity. 
+  in order to initialize two key attributes of the super-droplets, namely their volume and multiplicity. 
 Subsequently, a `Builder` object is created to orchestrate dependency injection while instantiating
   the `Particulator` class of `PySDM`:
 
@@ -245,7 +245,7 @@ The `Parcel` environment shares the zero-dimensionality of `Box` (i.e., no parti
   hydrostatically stratified atmosphere.
 Adiabatic cooling during the ascent results in supersaturation which triggers activation of
   aerosol particles (condensation nuclei) into cloud droplets through condensation.
-All examples based on the `Parcel` environment utilise the `Condensation` and `AmbientThermodynamics` dynamics.
+All examples based on the `Parcel` environment utilize the `Condensation` and `AmbientThermodynamics` dynamics.
 
 The simplest example uses a monodisperse particle spectrum represented with a single super-droplet
[…]
