### 2022-03-16 #3219 @slayoo
US spelling introduced here: https://github.com/atmos-cloud-sim-uj/PySDM/commit/dfdfc3a22a051ee7e761bbe4207cc27f2b120301

Zenodo titles updated:
- https://zenodo.org/record/6321270
- https://zenodo.org/record/6321280

Thanks!
