### 2022-04-18 #3219 @arfon
@dhhagan – this is failing because the DOI in the BibTeX here is invalid (the preceding `doi:` needs removing): https://github.com/atmos-cloud-sim-uj/PySDM/blob/master/paper/paper.bib#L170
