### 2022-04-18 #3219 @claresinger
@slayoo it seems the DOI being accessed is not the same as what is in the .bib file https://github.com/atmos-cloud-sim-uj/PySDM/blob/master/paper/paper.bib#L48
`10.1175/1520-0469(1967)024<0688:CDGBC>2.0.CO;2`
