### 2022-04-18 #3219 @slayoo
@claresinger, thanks, yes, it seems that recently something has changed in the way the bot parses the bib file, I've also just noted it in the PyMPDATA paper: https://github.com/openjournals/joss-reviews/issues/3896#issuecomment-1101700414
