### 2022-04-19 #3219 @arfon
> ⚠️ [Error prepararing paper acceptance](https://github.com/openjournals/joss-papers/actions/runs/2189150869?check_suite_focus=true). The generated XML metadata file is invalid.

@tarleb – any ideas what this kind of error is about?
