### 2022-04-19 #3219 @tarleb
The figures are defined using LaTeX syntax. We try hard to parse and convert LaTeX as much as sensible to accommodate cases like this, but that appears one of the cases that we miss.

The solution would be to use pandoc's "implicit figure" feature: a paragraph that contains just an image will be treated as a figure. E.g.

``` markdown
![Sample plot generated with the code snippets included in the paper.](readme.pdf){#fig:readme_fig_1 width="60%"}
```
