### 2022-04-19 188609b
fixing LaTeX figure problem in JOSS paper (thanks @tarleb for hint!)
