+++ paper/paper.md
@@ -204,12 +204,7 @@ pyplot.legend()
 pyplot.show()
 ```
 
-\begin{figure}[h]
-    \centering
-    \includegraphics[width=0.6\textwidth]{readme}
-    \caption{Sample plot generated with the code snippets included in the paper.}
-    \label{fig:readme_fig_1}
-\end{figure}
+![Sample plot generated with the code snippets included in the paper.](readme.pdf){#fig:readme_fig_1 width="60%"}
 
 # Usage examples
 
