### 2022-04-19 #3219 @slayoo
> This happens as the referenced figure is not included in the JATS output.
>
> The figures are defined using LaTeX syntax. We try hard to parse and convert LaTeX as much as sensible to accommodate cases like this, but that appears one of the cases that we miss.
>
> The solution would be to use pandoc's "implicit figure" feature: a paragraph that contains just an image will be treated as a figure. E.g.
>
> ```gfm
> ![Sample plot generated with the code snippets included in the paper.](readme.pdf){#fig:readme_fig_1 width="60%"}
> ```

changed here: https://github.com/atmos-cloud-sim-uj/PySDM/commit/188609b26f86956602b0c569057a1bbfdd4f6d63
thank you @tarleb & @arfon !
