### 2022-04-22 #3219 @slayoo
@dhhagan, @arfon, let me ask if you could retry the needed bot command to check if the "Error prepararing paper acceptance" still shows up (BTW, reported the error message typo here: https://github.com/openjournals/joss-papers/pull/3165), thanks.
