### 2022-04-24 #3219 @slayoo
@arfon Thank you!
In the final paper page that one is redirected to by using the DOI the author list is full of unescaped HTML non-braking space:
![image](https://user-images.githubusercontent.com/89685/164995668-772e2028-dd78-4b12-915f-58f3737dfe90.png)
Is it possible to do correct it?
Thanks
