### 2022-04-25 #3219 @slayoo
> @slayoo - please remove all of the &nsbp; characters and I'll re-deposit the paper.

Removed: https://github.com/atmos-cloud-sim-uj/PySDM/commit/cf59297dbe6728d479e0b8d2224903a71b37db70

(note, we have changed the branch name from `master` to `main`)

Thank you, @arfon!
